### 2016-10-14 #79 @adamhsparks
GSODR: Global Surface Summary Daily Weather Data in R
The GSODR package offers an R interface for downloading Global Surface
  Summary of the Day weather data files from the United States National Climatic
  Data Center, applies rules for cleaning and parses them into files useful for
  researchers.

```
Package: GSODR
Title: Global Summary Daily Weather Data in R
Version: 0.3
Authors@R: c(person("Adam", "Sparks", role = c("aut", "cre"),
    email = "adamhsparks@gmail.com"),
    person("Tomislav", "Hengl", role = "aut",
    email = "tom.hengl@isric.org"),
    person("Andrew", "Nelson", role = "aut",
    email = "dr.andy.nelson@gmail.com"))
Maintainer: Adam Sparks <adamhsparks@gmail.com>
Description:  The GSODR package offers automated downloading, parsing, cleaning
    and converting of Global Surface Summary of the Day (GSOD) weather data from
    the from the USA National Climatic Data Center (NCDC) into Comma Separated
    Values (CSV) or Geopackage (GPKG) files. The get_GSOD() function retrieves
    data from the NCDC's public FTP site and reformats it from United States
    Customary System (USCS) units to International System of Units (SI).
    Stations are individually checked for number of missing days as defined by
    the user to help ensure data quality. Stations files with too many missing
    observations are omitted from the final file. Only stations with valid
    reported latitude and longitude values are permitted in the final derived
    data set. A list of values for 200 metre buffered elevation data, derived
    from the CGIAR-CSI SRTM hole-filled 90 metre data set (Jarvis et al. 2008)
    are included for stations between -60 and 60 degrees latitude with valid
    latitude and longitude values. File output is returned as a comma-separated
    values (CSV) or GeoPackage (GPKG) file written to a local disk in a location
    selected by the user, which summarises each year by station. Additional
    useful variables, saturation vapour pressure (es), actual vapour pressure (ea)
    and relative humidity are calculated from the original data and included in
    the final data set. The resulting files can be as larger than 500mb
    depending on the user's stringency for missing data and geographic area of
    interest. Be sure to have sufficient RAM and disk space as well as a
    reasonably fast Internet connection to use this package to perform this
    operation for a global data set over many years. However, for much smaller
    and more manageable data sets, an individual country of interest may be
    selected as well as only stations falling between -60 and 60 degrees
    latitude for agroclimatology work or individual stations if the station ID
    is known. The resulting files include station data (e.g., station name,
    country, latitude, longitude, elevation) for use in a geographic
    information system (GIS). The function was largely based on T. Hengl's
    'getGSOD.R' script, available from
    <http://spatial-analyst.net/book/system/files/getGSOD.R> with enhancements
    to be cross-platform, more efficient with computing resources and more
    flexible. For information on the original data from NCDC, please see the
    GSOD readme.txt file available from,
    <http://www1.ncdc.noaa.gov/pub/data/gsod/readme.txt>.
    doParallel,
    foreach,
    iterators,
    parallel,
    RCurl,
    rgdal,
    sp,
    utils
    chillR,
    plotKML,
    spacetime,
    ggalt,
    rgeos,
    tidyr
NeedsCompilation: no
Repository: CRAN
```
  https://github.com/adamhsparks/GSODR
  Researchers who use daily weather data. Most individuals I've encountered so far
  are crop or plant disease modellers, hydrological modellers, etc. However,
  I'm there that there are others that I'm just unaware of due to my work-sphere.
  **[gsod: obtain daily weather data" from Jacob Van Etten, 2013](https://r-forge.r-project.org/projects/cropcc/)**
  This package was written for a instructional course and has not been submitted
  to CRAN or updated in some time. It does offer the same corrected elevation that
  GSODR offers, however the station list is not updated from the NCDC server,
  meaning that the station list provided with this package is over two years old.
  It is part of a crop modelling course, not designed to only download and clean
  GSOD data. GSODR also offers different types of file outputs, Geopackage or
  comma-separated values (CSV).
  **[GSODTools: R functions to select, download, and process GSOD data](https://github.com/environmentalinformatics-marburg/GSODTools)**
  This package is well written and maintained on GitHub only. The package is not
  available on CRAN at this date, either. It offers many of the same functions as
  GSODR. However, it does not automatically convert the values to International
  System of Units (SI), which GSODR does. It also does not offer the corrected
  elevation values that GSODR offers. Florian indicated that this project was
  not top priority during the completion of his PhD but was happy to consider
  working together in the future.
  **[rnoaa: R interface to many NOAA data APIs](https://github.com/ropensci/rnoaa)**
  This package offers an interface to several of NOAA's weather data sets. However,
  GSOD is not among the data sets that are available through this package.

  - [x] The package is deposited in a long-term repository with the DOI: [![DOI](https://zenodo.org/badge/32764550.svg)](https://zenodo.org/badge/latestdoi/32764550)
  - (_Do not submit your package separately to JOSS_)
  Scott Chamberlain, https://github.com/sckott
  Brooke Anderson, https://github.com/geanders
  Florian Detsch, https://github.com/fdetsch
  Jeff Hansen, https://github.com/jeffreyhanson
