### 2016-10-14 #79 @adamhsparks
@sckott, I've found a bug that I need to fix.

Not sure how it's crept in. I will report back when it's corrected.
