### 2016-10-15 f073972
Fix bug with station data frames
This was embarrassing.

There was a bug that caused the dataframe for the station information
to have NA values present in the LAT/LON columns (shouldn’t happen) and
also to have too few columns (again, shouldn’t happen).
