### 2016-10-15 #79 @adamhsparks
Well that was embarrassing. It's all fixed now.

Glad I checked the package build again this morning.
