### 2016-10-15 #79 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @adamhsparks   - no worries about eth bug

Will be seeking reviewers, but in the meantime: I ran `goodpractice::gp()` on your package. Here are the results, which it may be helpful to get started with before review (you can run it yourself and dig into the results more):

```
It is good practice to

    in general. At least some lines are not covered in this package.

    R/get_GSOD.R:230:NA
    R/get_GSOD.R:231:NA
    R/get_GSOD.R:234:NA
    R/get_GSOD.R:235:NA
    R/get_GSOD.R:238:NA
    ... and 359 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.

    R/get_GSOD.R:283:1
    R/get_GSOD.R:284:1
    R/get_GSOD.R:287:1
    R/get_GSOD.R:306:1
    R/get_GSOD.R:456:1
    ... and 9 more lines
```

* The vignette takes a loooooooooong time to run. Please do figure out a way to make it run much more quickly.
* The`Description` field in the `DESCRIPTION` file is too detailed. Please make it more concise.

---------

Reviewers: xxx
Due date: xxx
