### 2016-10-15 #79 @adamhsparks
Thanks for the valuable feedback. This is exactly the type of suggestions I was looking for here.

I'll get as much of this done as I can shortly.
