### 2016-10-15 19f49f5
Omit date in DESCRIPTION file
