### 2016-10-16 791d2a0
Shorten Description field in DESCRIPTION file
