### 2016-10-16 d70de3b
More refinement of Description field in DESCRIPTION file
