### 2016-10-16 85d3ec2
Revert "More refinement of Description field in DESCRIPTION file"
This reverts commit d70de3b0e7a6920fce0c582984a825ea8dc07bb9.
