### 2016-10-16 a6f2754
More refinement in Description field of DESCRIPTION file
