### 2016-10-16 3484cfe
Lint files
Corrected lines over 80 chars

Fixed incomplete final line found on
'/Users/asparks/Development/GSODR/tests/testthat.R'
