+++ R/get_GSOD.R
@@ -16,7 +16,8 @@
 #'group's (CGIAR-CSI) Shuttle Radar Topography Mission 90 metre (SRTM 90m)
 #'digital elevation data based on NASA's original SRTM 90m data. Further
 #'information on these data and methods can be found on GSODR's GitHub
-#'repository here: \url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
+#'repository here:
+#'\url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
 #'
 #' @param years Year(s) of weather data to download.
 #' @param station Specify a station or multiple stations for which to retrieve,
@@ -30,8 +31,8 @@
 #' @param country Specify a country of interest for which to retrieve weather
 #' data; full name. For stations located in locales
 #' having an ISO code 2 or 3 letter ISO code can also be used if known. See
-#' \code{\link{GSOD_country_list}} for a full list of country names and ISO codes
-#' available.
+#' \code{\link{GSOD_country_list}} for a full list of country names and ISO
+#' codes available.
 #' @param dsn Path to file write to.
 #' @param filename The filename for resulting file(s) to be written with no
 #' file extension. Year and file extension will be automatically appended to
@@ -280,11 +281,13 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL, dsn = "",
       tryCatch(utils::download.file(url = paste0(ftp_site, yr, "/gsod_", yr,
                                                  ".tar"),
                                     destfile = tf, mode = "wb"),
-               error = function(x) message(paste0("\nThe download stoped at year: ", yr,
-                                                  ".\nPlease restart the 'get_GSOD()' function starting at this point.\n")))
+               error = function(x) message(paste0("\nThe download stoped at
+      year: ", yr, ".\nPlease restart the 'get_GSOD()' function starting at this
+                                                  point.\n")))
       utils::untar(tarfile = tf, exdir  = paste0(td, "/", yr, "/"))
 
-      message("\nFinished downloading file. Parsing the indivdual station files now.\n")
+      message("\nFinished downloading file. Parsing the indivdual station files
+              now.\n")
       GSOD_list <- list.files(paste0(td, "/", yr, "/"),
                               pattern = utils::glob2rx("*.gz"),
                               full.names = FALSE)
@@ -303,7 +306,8 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL, dsn = "",
       # If country is set, subset list of stations to clean --------------------
       if (!is.null(country)) {
         country_FIPS <- unlist(as.character(stats::na.omit(
-          GSODR::GSOD_country_list[GSODR::GSOD_country_list$FIPS == country, ][1]),
+          GSODR::GSOD_country_list[GSODR::GSOD_country_list$FIPS == country, ]
+          [1]),
           use.names = FALSE))
         station_list <- stations[stations$CTRY == country_FIPS, ]$STNID
         station_list <- vapply(station_list,
[…]
