+++ vignettes/Using_GSODR.Rmd
@@ -22,7 +22,8 @@ implemented by most major GIS software, summarising each year by station, which
 also includes vapour pressure and relative humidity variables calculated from
 existing data in GSOD.
 
-This package was largely based on Tomislav Hengl's work, [getGSOD.R](http://spatial-analyst.net/book/getGSOD.R)
+This package was largely based on Tomislav Hengl's work,
+[getGSOD.R](http://spatial-analyst.net/book/getGSOD.R)
 in "[A Practical Guide to Geostatistical Mapping](http://spatial-analyst.net)",
 with updates for speed, cross-platform functionality, and more options for data
 retrieval and error correction.
@@ -40,10 +41,10 @@ library(GSODR)
 
 ## Plot Global Station Locations
 
-The `get_GSOD()` function automatically fetches the most recent station data from
-the NCDC website and removes stations that are missing data (as shown below).
-Using this data we can plot the station locations that are included in GSOD that
-provide valid geo-locations after cleaning.
+The `get_GSOD()` function automatically fetches the most recent station data
+from the NCDC website and removes stations that are missing data (as shown
+below). Using this data we can plot the station locations that are included in
+GSOD that provide valid geo-locations after cleaning.
 
 ```{r fetch_stations}
   GSOD_stations <- readr::read_csv(
@@ -55,11 +56,16 @@ provide valid geo-locations after cleaning.
   GSOD_stations[GSOD_stations == -999.9] <- NA
   GSOD_stations[GSOD_stations == -999] <- NA
 
-  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LAT) & !is.na(GSOD_stations$LON), ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LAT != 0 & GSOD_stations$LON != 0, ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LAT > -90 & GSOD_stations$LAT < 90, ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LON > -180 & GSOD_stations$LON < 180, ]
-  GSOD_stations$STNID <- as.character(paste(GSOD_stations$USAF, GSOD_stations$WBAN, sep = "-"))
+  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LAT) &
+                                   !is.na(GSOD_stations$LON), ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LAT != 0 &
+                                   GSOD_stations$LON != 0, ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LAT > -90 &
+                                   GSOD_stations$LAT < 90, ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LON > -180 &
+                                   GSOD_stations$LON < 180, ]
+  GSOD_stations$STNID <- as.character(paste(GSOD_stations$USAF,
+                                            GSOD_stations$WBAN, sep = "-"))
   
   SRTM_GSOD_elevation <- data.table::setkey(GSODR::SRTM_GSOD_elevation, STNID)
   data.table::setDT(GSOD_stations)
@@ -75,7 +81,7 @@ Using [ggplot2](https://CRAN.R-project.org/package=ggplot2) and the
 the station locations using alpha transparency to see the densest part of the
 network and use the Robinson projection for the map.
 
-```{r plot_stations, fig.width=7, fig.height=7, fig.cap="GSOD Station Locations"}
+```{r plot_stations, fig.width=7, fig.height=7,fig.cap="GSOD Station Locations"}
 library(ggplot2)
 library(ggalt)
 
@@ -94,7 +100,8 @@ in Australia.
 ```{r australia_stations, message=FALSE, warning=FALSE}
 library(dplyr)
 # left_join the the station data with the country list
[…]
