### 2016-10-16 3216892
Build vignette faster
Restructure vignette so that it builds faster when checking R package

Rename `GSOD_country_list` to `country_list` to save character space

Correct spelling error on line 285 of `get_GSOD.R` in error message
