### 2016-10-16 902d128
Correct issues with package check
Package check now passes.
 *Vignette builds
 *Testthat tests now pass
