### 2016-10-16 #79 @adamhsparks
I've updated the DESCRIPTION file in the package and here in this issue as suggested.
