### 2016-10-16 #79 @adamhsparks
Vignette builds faster now, as suggested as well.
