### 2016-10-16 463fa56
Added examples with spacetime plotting and data set with climate data
