+++ R/data.R
@@ -68,3 +68,94 @@
 #' @source \url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/country-list.txt}
 #'
 "GSOD_country_list"
+
+#' GSOD_clim
+#' @format A data frame with 23927 observations of  73 variables:
+#' \describe{
+#' \item{STNID}{Unique station ID}
+#' \item{LON}{Longitude in WGS84 system}
+#' \item{LAT}{Latitude in WGS84 system}
+#' \item{CHELSA_bio1_1979-2013_V1_1}{Annual mean temperature [degree C]}
+#' \item{CHELSA_bio2_1979-2013_V1_1}{Mean diurnal range [degree C]}
+#' \item{CHELSA_bio3_1979-2013_V1_1}{Isothermality}
+#' \item{CHELSA_bio4_1979-2013_V1_1}{Temperature seasonality}
+#' \item{CHELSA_bio5_1979-2013_V1_1}{Maximum Temperature of warmest month [degree C]}
+#' \item{CHELSA_bio6_1979-2013_V1_1}{Minimum Temperature of coldest month [degree C]}
+#' \item{CHELSA_bio7_1979-2013_V1_1}{Temperature Annual Range [degree C]}
+#' \item{CHELSA_bio8_1979-2013_V1_1}{Mean Temperature of wettest quarter [degree C]}
+#' \item{CHELSA_bio9_1979-2013_V1_1}{Mean Temperature of driest quarter [degree C]}
+#' \item{CHELSA_bio10_1979-2013_V1_1}{Mean Temperature of warmest quarter [degree C]}
+#' \item{CHELSA_bio11_1979-2013_V1_1}{Mean Temperature of coldest quarter [degree C]}
+#' \item{CHELSA_bio12_1979-2013_V1_1}{Annual precipitation amount [mm]}
+#' \item{CHELSA_bio13_1979-2013_V1_1}{Precipitation of wettest month [mm]}
+#' \item{CHELSA_bio14_1979-2013_V1_1}{Precipitation of driest month [mm]}
+#' \item{CHELSA_bio15_1979-2013_V1_1}{Precipitation Seasonality}
+#' \item{CHELSA_bio16_1979-2013_V1_1}{Precipitation of wettest quarter [mm]}
+#' \item{CHELSA_bio17_1979-2013_V1_1}{Precipitation of driest quarter [mm]}
+#' \item{CHELSA_bio18_1979-2013_V1_1}{Precipitation of warmest quarter [mm]}
+#' \item{CHELSA_bio19_1979-2013_V1_1}{Precipitation of coldest quarter [mm]}
+#' \item{CHELSA_prec_1_1979-2013}{Mean january precipitation}
+#' \item{CHELSA_prec_2_1979-2013}{Mean february precipitation}
+#' \item{CHELSA_prec_3_1979-2013}{Mean march precipitation}
+#' \item{CHELSA_prec_4_1979-2013}{Mean april precipitation}
+#' \item{CHELSA_prec_5_1979-2013}{Mean may precipitation}
+#' \item{CHELSA_prec_6_1979-2013}{Mean june precipitation}
+#' \item{CHELSA_prec_7_1979-2013}{Mean july precipitation}
+#' \item{CHELSA_prec_8_1979-2013}{Mean august precipitation}
+#' \item{CHELSA_prec_9_1979-2013}{Mean september precipitation}
+#' \item{CHELSA_prec_10_1979-2013}{Mean october precipitation}
+#' \item{CHELSA_prec_11_1979-2013}{Mean november precipitation}
+#' \item{CHELSA_prec_12_1979-2013}{Mean december precipitation}
[…]
