### 2016-10-16 #79 @adamhsparks
Tomislav Hengl has contributed to the package and is now an author. I've updated the Description.
