### 2016-10-16 6eeedc9
Revert "Merge branch 'devel'"
This reverts commit 85e46e27fe3ee6692be7c0ad751eaf5d7e9625d5, reversing
changes made to 62343aba51991f7681348539e374f86ee2ab39b8.
