### 2016-10-19 fdf751b
Added visualization of GSOD values vs CHELSA temperatures
