### 2016-10-20 #79 @jeffreyhanson
## Package Review

- [:white_check_mark:] As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).

#### Documentation

- [:white_check_mark:] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [:white_check_mark:] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [:x:] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [:white_check_mark:] **Function Documentation:** for all exported functions in R help
- [:white_check_mark:] **Examples** for all exported functions in R Help that run successfully locally
- [:white_check_mark:] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION

##### Paper (for packages co-submitting to JOSS)

The package contains a `paper.md` with:

- [:white_check_mark:] **A short summary** describing the high-level functionality of the software
- [:x:] **Authors:**  A list of authors with their affiliations
- [:white_check_mark:] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
- [:white_check_mark:] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [:white_check_mark:] **Installation:** Installation succeeds as documented.
- [:white_check_mark:] **Functionality:** Any functional claims of the software been confirmed.
- [:white_check_mark:] **Performance:** Any performance claims of the software been confirmed.
- [:x:] **Automated tests:** Unit tests cover essential functions of the package
- [:x:] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

Estimated hours spent reviewing: 3

---

### Review Comments
#### Overview
The [Global Summary Of The Day](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod) dataset contains weather data collected from meteorological stations all over the planet. However, accessing and formatting this data for use in R can be rather tricky. **The GSODR R package will make this a dream**. This package will be incredibly useful for new R users who want easy access to this data set. It will also be useful for experienced R users who want to download data over large extents and do post-processing later in R. I very much look forward to using this package in the future.

I have several major concerns before I can recommend this package for release. First, the main function `get_GSOD` fails when the user wants to download data for an entire country. Personally, I think this is the biggest feature of this package. Second, the package could benefit from more automated testing. None of the examples are actually tested, and the unit tests only cover cases that are expected to fail. Third, the package only downloads data. It would be brilliant if the `get_GSOD` function returned data after completion. I have outlined these main issues and other issues below.

#### Major issues
1. The `get_GSOD` function cannot download all data for an entire country. For instance, running the code `get_GSOD(year=2015, country='NZ', dsn=tempdir())` returns this error:
```
Error in col[[i, exact = exact]] :
 attempt to select more than one element in vectorIndex
```
2. All examples are wrapped in `\dontrun{}` so they are not actually tested.
3. The main function `get_GSOD` just saves data to disk and does not actually return data. It would be brilliant if this package returned GSOD data so the user does not have to then load it back into R. To that end, I would recommend the function returns a single `tibble` or `SpatialPointsDataFrame` object (perhaps depending on an argument?), and let the user then re-save the data where they want it stored.
4. As far as I can tell, the vignette cannot be accessed within R. After installation, the following code `vignette(package='GSODR')` yields `'no vignettes found'`. Additionally, the rmarkdown file for the vignette has an obscure name, so even if it could be opened within R, this would be tricky given its obscure name. I recommend naming the vignette after the package so the following code will open it `vignette('GSODR')`.
5. Unit testing covers cases where `get_GSOD` is expected to fail but it does not actually test cases that are expected to succeed.
6. Missing data is represented using -9999. Why not use NA values? The main aim of this package is to load GSOD data into R.
7. There is no help page for the package (ie. `?GSODR`). It would be really useful to include this, even if it it just a copy-paste from the README. [See here for more details.](http://r-pkgs.had.co.nz/man.html#man-packages)
8. The `nearest_station` function will be incredibly useful for a user. It would be even more useful if--instead of printing the stations closest to a user-specified point--it returned the stations. Then this function could be used programmatically to find the closest stations to a series of points, and download data accordingly.
9. The accompanying paper does not list authors in the html document.

#### Minor issues
1. This package needs a rerun of `devtools::document()` before resubmission. When I ran `devtools::run_examples()`, this automatically re-compiled some Rd files.
2. get_GSOD.R:20 & data.R:29: Both lines refer to a document available on the github repository. I recommend making this document an additional vignette.
3. get_GSOD.R:279: I am unsure why `foreach` iterators are used here. It would be great to do offer some parallelisation and options for progress bars here. If you wrapped this in a `plyr::**ply` function, you could easily make it such that: (1) years can easily be processed in parallel, (2) if not processed in parallel, then a progress-bar can be shown.
4. get_GSOD.R:322 ROpensci guidelines recommend using the httr package instead of the RCurl package.
5. get_GSOD.R:392--393: Options are restored to the default state but this is not necessarily the previous state. These values could be cached at the beginning of the function. For example: use `orginal_options <- options()` at the start of the function, and at the end of the function use `options(original_options)`.
6. get_GSOD.R:345: the user should have explicit control over thread use to avoid accidently fork-bombing their own computer. I recommend providing a `threads` argument in the `get_GSOD` function and defaulting it to 1.
7. get_GSOD.R:561: Here and elsewhere `return(0)` is used after a call to stop. Please consider removing the `return(0)` as this is unnecessary .
8. nearest_stations.R:28--29: This function changes options() and does not reset them after it finishes. This is against recommended CRAN practices. Please see previous comment on options for a suggestion on how to fix this.
9. nearest_stations.R:34--44: This code could be rewritten for clarity and computational efficiency. I suggest something like this:
```
dists <- fields::rdist.earth(as.matrix(stations[c('LON', 'LAT')]), matrix(c(LON, LAT), ncol=2), miles=FALSE)
nearby <- which(dists[,1] < distance)
```

#### Typographic and stylistic issues
1. get_GSOD.R:83--85: Documentation contains '&lt' which is not rendered as a less than symbol in the html documentation.
3. get_GSOD.R:292: Instead of a call to `utils::glob2rx`, this could be used instead `pattern='^.*\\.gz$'`
2. get_GOSD.R:631: This could rewritten be as `!vs %in% stations[[12]]` for ease of reading.
