### 2016-10-20 #79 @sckott
thanks for the review @jeffreyhanson !   a few notes:

* wrapping examples in `\dontrun{}` is what we suggest in a package like this where examples make web requests, so I'd keep that - though if there's any examples that don't make web requests, definitely do put those outside of `\dontrun`
