### 2016-10-21 #79 @jeffreyhanson
@sckott Ok, thanks for correcting me on that. I just did a quick check and the `nearest_stations` function also has its example wrapped in a `\dontrun{}`. This function doesn't make web requests.
