### 2016-10-22 81fe389
Fix bug when ISO2c/ISO3c is set, query fails
