### 2016-10-22 #79 @adamhsparks
@jeffreyhanson, I've fixed the bug with the country request in the devel branch. Good catch, THANK YOU!

Your comment about the `nearest_stations` is incorrect. It does make a web query with the `.fetch_stations` function, thus is wrapped in a `\dontrun{}`.

I'll work on some of the other issue's you've raised while I wait on my plane now.
