### 2016-10-22 f8ae2dc
Rerun devtools::document()
