### 2016-10-22 5c622af
Replace pattern call on line 292
Replace: `utils::glob2rx`

with: `pattern='^.*\\.gz$’`

Suggested by jefferyhanson
