### 2016-10-22 84c210f
Fix Documentation contains '&lt' which is not rendered
Correct: get_GSOD.R:83--85: Documentation contains '&lt' which is not
rendered as a less than symbol in the html documentation.

Suggested by jefferyhanson
