### 2016-10-22 d404cb4
Add threads option to `get_GSOD` function for parallel processing
get_GSOD.R:345: the user should have explicit control over thread use
to avoid accidently fork-bombing their own computer. I recommend
providing a threads argument in the get_GSOD function and defaulting it
to 1.

Added `thread` option, defaults to 1

Suggested by jeffereyhanson
