### 2016-10-22 #79 @adamhsparks
Add threads option to `get_GSOD` function for parallel processing  …
adamhsparks committed a minute ago
d404cb4

Fix Documentation contains '&lt' which is not rendered  …
adamhsparks committed 19 minutes ago
84c210f

Replace pattern call on line 292  …
adamhsparks committed 21 minutes ago
5c622af

Fix bug when ISO2c/ISO3c is set, query fails
adamhsparks committed 29 minutes ago
81fe389

Thanks, @jeffreyhanson, boarding now.
