### 2016-10-22 434f71e
Fix spelling error in documentation
