### 2016-10-22 26e3717
Update capitalisation and formatting in R/data.R
