### 2016-10-22 ec43e89
Remove date field in DESCRIPTION file
