+++ DESCRIPTION
@@ -2,7 +2,6 @@ Package: GSODR
 Type: Package
 Title: Global Summary Daily Weather Data in R
 Version: 0.3
-Date: 2016-10-16
 Authors@R: c(person("Adam", "Sparks", role = c("aut", "cre"),
     email = "adamhsparks@gmail.com"),
     person("Tomislav", "Hengl", role = "aut",
