### 2016-10-22 59b816e
Shorten Description field in DESCRIPTION file
This field somehow reverted after modifications were made to shorten it
and merged into the Master branch.

Replacing longer text with the shorter version from Master, here in
devel
