### 2016-10-23 12f8683
Add Tom to authors in paper.md
