+++ inst/paper/paper.md
@@ -11,6 +11,9 @@ authors:
   - name: Adam H Sparks
     orcid: 0000-0002-0061-8359
     affiliation: Centre for Crop Health, University of Southern Queensland, Toowoomba Queensland 4350, Australia
+  - name: Tomislav Hengl
+    orcid: 0000-0002-9921-5129
+    affiliation: ISRIC - World Soil Information, ISRIC - World Soil Information P.O. Box 353, 6700 AJ Wageningen, The Netherlands
   - name: Andrew Nelson
     orcid: 0000-0002-7249-3778
     affiliation: Faculty of Geo-Information and Earth Observation (ITC), University of Twente, Enschede 7500 AE, The Netherlands
