### 2016-10-23 4aa35eb
Small updates in the summary and added a ref
