### 2016-10-23 #79 @adamhsparks
Update authors in paper.md
https://github.com/adamhsparks/GSODR/commit/12f8683a9147b41aa7bbf76f88dffdfd90c3631c
