### 2016-10-23 8372844
Edit code for clarity
Replace: vs %in% stations[[12]] == FALSE
With !vs %in% stations[[12]]

Suggested by @jefferyhanson
