### 2016-10-23 #79 @adamhsparks
Edit code for clarity
> get_GOSD.R:631: This could rewritten be as !vs %in% stations[[12]] for clarity.
https://github.com/adamhsparks/GSODR/commit/83728444ac456da8b021032696ed469b0b208ac7
