### 2016-10-23 960bb82
Check and reset original options()
Define original options and reset at end of `get_GSOD` function

suggested by @jefferyhanson
