### 2016-10-23 #79 @adamhsparks
Check and reset original options()

> get_GSOD.R:392--393: Options are restored to the default state but this is not necessarily the previous state. These values could be cached at the beginning of the function. For example: use orginal_options <- options() at the start of the function, and at the end of the function use options(original_options).

https://github.com/adamhsparks/GSODR/commit/960bb82e657b33a175cdac2e10d73fd755885bae
