### 2016-10-24 a0713f6
Rename vignette
Rename vignette to "GSODR"
