+++ vignettes/GSODR.md
@@ -1,0 +1,449 @@
+<!--
+%\VignetteEngine{knitr::knitr}
+%\VignetteIndexEntry{Sea ice vignette}
+%\VignetteEncoding{UTF-8}
+-->
+
+# Introduction
+An R package that provides a function that automates downloading and
+cleaning data from the [Global Surface Summary of the Day (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)
+data provided by the US National Climatic Data Center (NCDC). Stations
+are individually checked for number of missing days to assure data
+quality, those stations with too many missing observations as defined by the
+user are omitted. All units are converted to International System of Units (SI),
+e.g., inches to millimetres and Fahrenheit to Celsius. Output is saved as a
+Comma Separated Value (CSV) file or in a spatial GeoPackage (GPKG) file,
+implemented by most major GIS software, summarising each year by station, which
+also includes vapour pressure and relative humidity variables calculated from
+existing data in GSOD.
+
+This package was largely based on Tomislav Hengl's work,
+[getGSOD.R](http://spatial-analyst.net/book/getGSOD.R)
+in "[A Practical Guide to Geostatistical Mapping](http://spatial-analyst.net)",
+with updates for speed, cross-platform functionality, and more options for data
+retrieval and error correction.
+
+For more information see the description of the data provided by NCDC,
+<http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt>.
+
+# Using the package
+
+## Load GSODR
+
+```r
+library(GSODR)
+```
+
+## Plot Global Station Locations
+
+The `get_GSOD()` function automatically fetches the most recent station data
+from the NCDC website and removes stations that are missing data (as shown
+below). Using this data we can plot the station locations that are included in
+GSOD that provide valid geo-locations after cleaning.
+
+```r
+  GSOD_stations <- readr::read_csv(
+    "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
+    col_types = "ccccccddddd",
+    col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL",
+                  "LAT", "LON", "ELEV_M", "BEGIN", "END"), skip = 1)
+
+  GSOD_stations[GSOD_stations == -999.9] <- NA
+  GSOD_stations[GSOD_stations == -999] <- NA
+
+  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LAT) &
+                                   !is.na(GSOD_stations$LON), ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LAT != 0 &
+                                   GSOD_stations$LON != 0, ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LAT > -90 &
+                                   GSOD_stations$LAT < 90, ]
+  GSOD_stations <- GSOD_stations[GSOD_stations$LON > -180 &
+                                   GSOD_stations$LON < 180, ]
+  GSOD_stations$STNID <- as.character(paste(GSOD_stations$USAF,
+                                            GSOD_stations$WBAN, sep = "-"))
+  
+  SRTM_GSOD_elevation <- data.table::setkey(GSODR::SRTM_GSOD_elevation, STNID)
[…]
