### 2016-10-24 42867e6
Revert "Merge branch 'master' into devel"
This reverts commit b948af7d3084805f86aae212c1f611dc16b68e38, reversing
changes made to 960bb82e657b33a175cdac2e10d73fd755885bae.
