### 2016-10-24 ddbbe82
Correct misspelled "original_options" in `get_GSOD`
