### 2016-10-24 0c75123
Fix vignette index entry
