+++ vignettes/GSODR.Rmd
@@ -1,6 +1,6 @@
 <!--
 %\VignetteEngine{knitr::knitr}
-%\VignetteIndexEntry{Sea ice vignette}
+%\VignetteIndexEntry{GSODR}
 %\VignetteEncoding{UTF-8}
 -->
 
@@ -16,12 +16,6 @@ Comma Separated Value (CSV) file or in a spatial GeoPackage (GPKG) file,
 implemented by most major GIS software, summarising each year by station, which
 also includes vapour pressure and relative humidity variables calculated from
 existing data in GSOD.
-
-This package was largely based on Tomislav Hengl's work,
-[getGSOD.R](http://spatial-analyst.net/book/getGSOD.R)
-in "[A Practical Guide to Geostatistical Mapping](http://spatial-analyst.net)",
-with updates for speed, cross-platform functionality, and more options for data
-retrieval and error correction.
 
 For more information see the description of the data provided by NCDC,
 <http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt>.
