### 2016-10-24 #79 @adamhsparks
Renamed vignette to "GSODR"

https://github.com/adamhsparks/GSODR/commit/a0713f6602ec31981773c7a3d675c0082b5412c7

Additionally, @jeffreyhanson, If you install from github using devtools, `devtools::install(build_vignettes = TRUE)`, will build the vignette, so then using `vignette(package="GSODR")` a vignette is returned.
