### 2016-10-24 #79 @adamhsparks
Added GSODR.R to create helpfile for package

https://github.com/adamhsparks/GSODR/commit/a4044943d284e0cd44f53670b21995c33ee4a1fb
