### 2016-10-24 a404494
Add GSODR.R to create helpfile for package
suggested by @jefferyhanson
