### 2016-10-24 1cc0403
Update README.Rmd file
Fix grammar

Incorporate Tom's edits and documentation
