+++ README.md
@@ -76,8 +76,8 @@ Using GSODR
 
 ### Query the NCDC FTP server for GSOD data
 
-GSODR's main function, `get_GSOD()` downloads and cleans GSOD data from
-the NCDC server. Following are a few examples of it's capabilities.
+GSODR's main function, `get_GSOD` downloads and cleans GSOD data from
+the NCDC server. Following are a few examples of its capabilities.
 
 ``` r
 
@@ -118,10 +118,13 @@ pnts <- readOGR("PHL_2010-2010.gpkg", layers[1])
 # It has 46 fields
 
 # Plot results in Google Earth as a spacetime object:
-pnts$DATE = as.Date(paste(pnts$YEAR, pnts$MONTH, pnts$DAY, sep="-"))
-tmp_ST <- STIDF(sp=as(pnts, "SpatialPoints"), time=pnts$DATE-0.5, data=pnts@data[,c("TEMP","STNID")], endTime=pnts$DATE+0.5)
+pnts$DATE = as.Date(paste(pnts$YEAR, pnts$MONTH, pnts$DAY, sep = "-"))
+tmp_ST <- STIDF(sp = as(pnts, "SpatialPoints"), time = pnts$DATE - 0.5,
+                data = pnts@data[, c("TEMP", "STNID")],
+                endTime = pnts$DATE + 0.5)
 shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
-kml(tmp_ST, dtime = 24*3600, colour = TEMP, shape = shape, labels = TEMP, file.name="Temperatures_PHL_2010-2010.kml", folder.name="TEMP")
+kml(tmp_ST, dtime = 24*3600, colour = TEMP, shape = shape, labels = TEMP,
+    file.name = "Temperatures_PHL_2010-2010.kml", folder.name = "TEMP")
 system("zip -m Temperatures_PHL_2010-2010.kmz Temperatures_PHL_2010-2010.kml")
 
 # Compare daily measurements with climatic data:
@@ -132,24 +135,26 @@ library(reshape2)
 data(GSOD_clim)
 cnames = paste0("CHELSA_temp_", 1:12, "_1979-2013")
 clim.temp = GSOD_clim[GSOD_clim$STNID %in% pnts$STNID, paste(c("STNID", cnames))]
-clim.temp.df = data.frame(STNID=rep(clim.temp$STNID, 12),
-   MONTHC=as.vector(sapply(1:12, rep, times=nrow(clim.temp))), 
-   TEMPC=as.vector(unlist(clim.temp[,cnames])))
+clim.temp.df = data.frame(STNID = rep(clim.temp$STNID, 12),
+   MONTHC = as.vector(sapply(1:12, rep, times = nrow(clim.temp))), 
+   TEMPC = as.vector(unlist(clim.temp[,cnames])))
 clim.temp.df$YEAR = 2010
 pnts$MONTHC = as.numeric(paste(pnts$MONTH))
-temp <- plyr::join(pnts@data, clim.temp.df, type="left")
+temp <- plyr::join(pnts@data, clim.temp.df, type = "left")
 df_melt <- melt(temp[, c("STNID", "DATE", "TEMPC", "TEMP")],
-   id=c("DATE","STNID"))
-gg = ggplot(df_melt, aes(x=DATE, y=value))
-gg1 = gg + geom_point(aes(color=variable))
+   id = c("DATE","STNID"))
+gg = ggplot(df_melt, aes(x = DATE, y = value))
+gg1 = gg + geom_point(aes(color = variable))
 gg1 + facet_wrap( ~ STNID)
 ```
[…]
