### 2016-10-24 78b643e
Corrections to README.md
