### 2016-10-24 #79 @adamhsparks
Removed `return(0)` after `stop` command

https://github.com/adamhsparks/GSODR/commit/9c4c24ea4d1221768b2083d86404fb29ce744959
