### 2016-10-24 9c4c24e
Remove `return(0)` after `stop`
Suggested by @jefferyhanson
