### 2016-10-24 57e8803
Update SRTM_GSOD_elevation data
