+++ data-raw/fetch_isd-history.md
@@ -1,12 +1,12 @@
 Fetch, clean and correct altitude in GSOD isd\_history.csv Data
 ================
 Adam H. Sparks
-2016-10-11
+2016-10-24
 
 Introduction
 ============
 
-This document details how the GSOD station history data file, ["isd-history.csv"](ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv), is fetched from the NCDC ftp server, error checked and new elevation values generated. The new elevation values are then saved for inclusion in package as /data/SRTM\_GSOD\_elevation.rda. The resulting values are merged with the most recent station history data file from the NCDC when the user runs the `get_GSOD()` function. The resulting data frame of station information, based on the merging of the `SRTM_GSOD_elevation` data frame with the most recently available "isd-history.csv" file will result in the following changes to the data:
+This document details how the GSOD station history data file, ["isd-history.csv"](ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv), is fetched from the NCDC ftp server, error checked and new elevation values generated. The new elevation values are then saved for inclusion in package as /data/SRTM\_GSOD\_elevation.rda. The resulting values are merged with the most recent station history data file from the NCDC when the user runs the `get_GSOD` function. The resulting data frame of station information, based on the merging of the `SRTM_GSOD_elevation` data frame with the most recently available "isd-history.csv" file will result in the following changes to the data:
 
 -   Stations where latitude or longitude are NA or both 0 are removed
 
@@ -177,7 +177,7 @@ summary(SRTM_GSOD_elevation)
 ```
 
     ##      USAF               WBAN             STN_NAME        
-    ##  Length:27855       Length:27855       Length:27855      
+    ##  Length:27861       Length:27861       Length:27861      
     ##  Class :character   Class :character   Class :character  
     ##  Mode  :character   Mode  :character   Mode  :character  
     ##                                                          
@@ -185,8 +185,8 @@ summary(SRTM_GSOD_elevation)
     ##                                                          
     ##                                                          
     ##      CTRY              STATE               CALL                LAT        
-    ##  Length:27855       Length:27855       Length:27855       Min.   :-89.00  
-    ##  Class :character   Class :character   Class :character   1st Qu.: 24.09  
+    ##  Length:27861       Length:27861       Length:27861       Min.   :-89.00  
+    ##  Class :character   Class :character   Class :character   1st Qu.: 24.08  
     ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.67  
     ##                                                           Mean   : 31.86  
     ##                                                           3rd Qu.: 50.02  
@@ -194,20 +194,20 @@ summary(SRTM_GSOD_elevation)
     ##                                                                           
     ##       LON               ELEV_M           BEGIN               END          
     ##  Min.   :-179.983   Min.   :-350.0   Min.   :19010101   Min.   :19051231  
-    ##  1st Qu.: -83.845   1st Qu.:  22.3   1st Qu.:19570630   1st Qu.:20020207  
-    ##  Median :   7.828   Median : 137.0   Median :19750729   Median :20150808  
-    ##  Mean   :  -2.743   Mean   : 359.6   Mean   :19775921   Mean   :20040896  
[…]
