### 2016-10-24 9020dc8
Revert "Update SRTM_GSOD_elevation data"
This reverts commit 57e8803f14aa63b13407d9ca7682d495a6e4cab8.
