+++ data-raw/fetch_isd-history.md
@@ -1,7 +1,7 @@
 Fetch, clean and correct altitude in GSOD isd\_history.csv Data
 ================
 Adam H. Sparks
-2016-10-11
+2016-10-24
 
 Introduction
 ============
@@ -177,7 +177,7 @@ summary(SRTM_GSOD_elevation)
 ```
 
     ##      USAF               WBAN             STN_NAME        
-    ##  Length:27855       Length:27855       Length:27855      
+    ##  Length:27861       Length:27861       Length:27861      
     ##  Class :character   Class :character   Class :character  
     ##  Mode  :character   Mode  :character   Mode  :character  
     ##                                                          
@@ -185,8 +185,8 @@ summary(SRTM_GSOD_elevation)
     ##                                                          
     ##                                                          
     ##      CTRY              STATE               CALL                LAT        
-    ##  Length:27855       Length:27855       Length:27855       Min.   :-89.00  
-    ##  Class :character   Class :character   Class :character   1st Qu.: 24.09  
+    ##  Length:27861       Length:27861       Length:27861       Min.   :-89.00  
+    ##  Class :character   Class :character   Class :character   1st Qu.: 24.08  
     ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.67  
     ##                                                           Mean   : 31.86  
     ##                                                           3rd Qu.: 50.02  
@@ -194,20 +194,20 @@ summary(SRTM_GSOD_elevation)
     ##                                                                           
     ##       LON               ELEV_M           BEGIN               END          
     ##  Min.   :-179.983   Min.   :-350.0   Min.   :19010101   Min.   :19051231  
-    ##  1st Qu.: -83.845   1st Qu.:  22.3   1st Qu.:19570630   1st Qu.:20020207  
-    ##  Median :   7.828   Median : 137.0   Median :19750729   Median :20150808  
-    ##  Mean   :  -2.743   Mean   : 359.6   Mean   :19775921   Mean   :20040896  
-    ##  3rd Qu.:  64.585   3rd Qu.: 428.0   3rd Qu.:20010915   3rd Qu.:20161007  
-    ##  Max.   : 179.750   Max.   :5304.0   Max.   :20160927   Max.   :20161009  
+    ##  1st Qu.: -83.840   1st Qu.:  22.3   1st Qu.:19570630   1st Qu.:20020207  
+    ##  Median :   7.833   Median : 137.0   Median :19750803   Median :20150808  
+    ##  Mean   :  -2.744   Mean   : 359.7   Mean   :19776028   Mean   :20040953  
+    ##  3rd Qu.:  64.588   3rd Qu.: 428.0   3rd Qu.:20010915   3rd Qu.:20161020  
+    ##  Max.   : 179.750   Max.   :5304.0   Max.   :20161018   Max.   :20161021  
     ##                     NA's   :218                                           
-    ##     STNID           ELEV_M_SRTM_90m  
-    ##  Length:27855       Min.   :-361.00  
-    ##  Class :character   1st Qu.:  24.75  
-    ##  Mode  :character   Median : 153.00  
-    ##                     Mean   : 379.32  
-    ##                     3rd Qu.: 456.00  
-    ##                     Max.   :5273.00  
-    ##                     NA's   :3011
+    ##     STNID           ELEV_M_SRTM_90m 
+    ##  Length:27861       Min.   :-361.0  
+    ##  Class :character   1st Qu.:  24.0  
+    ##  Mode  :character   Median : 153.0  
[…]
