### 2016-10-24 77a9c3f
`get_GSOD` creates data.frame; Replace -9999 with NA; Documentation
* `get_GSOD` will now return a data frame object within R. No need to
save a file and reimport;

* Missing values are now `NA`, previously -9999;

* Update documentation to reflect these changes.
