+++ R/get_GSOD.R
@@ -63,7 +63,7 @@
 #'stations, all years queried and stations queried will be merged into one final
 #'ouptut file (CSV or GeoPackage).
 #'
-#'All missing values in resulting files are represented as -9999
+#'All missing values in resulting files are represented as NA
 #'regardless of which field they occur in.
 #'
 #'Be sure to have disk space free and allocate the proper time for this to run.
@@ -95,36 +95,36 @@
 #' \item{DAY}{The day (dd)}
 #' \item{YDAY}{Sequential day of year (not in original GSOD)}
 #' \item{TEMP}{Mean daily temperature converted to degrees C to tenths.
-#' Missing = -9999}
+#' Missing = NA}
 #' \item{TEMP_CNT}{Number of observations used in calculating mean daily
 #' temperature}
 #' \item{DEWP}{Mean daily dew point converted to degrees C to tenths. Missing
-#' = -9999}
+#' = NA}
 #' \item{DEWP_CNT}{Number of observations used in calculating mean daily dew
 #' point}
-#' \item{SLP}{Mean sea level pressure in millibars to tenths. Missing =   -9999}
+#' \item{SLP}{Mean sea level pressure in millibars to tenths. Missing =   NA}
 #' \item{SLP_CNT}{Number of observations used in calculating mean sea level
 #' pressure}
 #' \item{STP}{Mean station pressure for the day in millibars to tenths.
-#' Missing = -9999}
+#' Missing = NA}
 #' \item{STP_CNT}{Number of observations used in calculating mean station
 #' pressure}
 #' \item{VISIB}{Mean visibility for the day converted to kilometres to
-#' tenths Missing = -9999}
+#' tenths Missing = NA}
 #' \item{VISIB_CNT}{Number of observations used in calculating mean daily
 #' visibility}
 #' \item{WDSP}{Mean daily wind speed value converted to metres/second to
-#' tenths Missing = -9999}
+#' tenths Missing = NA}
 #' \item{WDSP_CNT}{Number of observations used in calculating mean daily
 #' wind speed}
 #' \item{MXSPD}{Maximum sustained wind speed reported for the day converted
-#' to metres/second to tenths. Missing = -9999}
+#' to metres/second to tenths. Missing = NA}
 #' \item{GUST}{Maximum wind gust reported for the day converted to
-#' metres/second to tenths. Missing = -9999}
+#' metres/second to tenths. Missing = NA}
 #' \item{MAX}{Maximum temperature reported during the day converted to
 #' Celsius to tenths--time of max temp report varies by country and region,
 #' so this will sometimes not be the max for the calendar day. Missing =
-#' -9999}
+#' NA}
 #' \item{MAX_FLAG}{Blank indicates max temp was taken from the explicit max
 #' temp report and not from the 'hourly' data. An "\*" indicates max temp was
 #' derived from the hourly data (i.e., highest hourly or synoptic-reported
@@ -132,7 +132,7 @@
 #' \item{MIN}{Minimum temperature reported during the day converted to
 #' Celsius to tenths--time of min temp report varies by country and region,
 #' so this will sometimes not be the max for the calendar day. Missing =
-#' -9999}
+#' NA}
 #' \item{MIN_FLAG}{Blank indicates max temp was taken from the explicit max
 #' temp report and not from the 'hourly' data. An "\*" indicates min temp was
 #' derived from the hourly data (i.e., highest hourly or synoptic-reported
@@ -141,8 +141,8 @@
 #' the day converted to millimetres to hundredths; will usually not end
 #' with the midnight observation, i.e., may include latter part of previous
[…]
