+++ vignettes/GSODR.Rmd
@@ -30,7 +30,7 @@ library(GSODR)
 
 ## Plot Global Station Locations
 
-The `get_GSOD()` function automatically fetches the most recent station data
+The `get_GSOD` function automatically fetches the most recent station data
 from the NCDC website and removes stations that are missing data (as shown
 below). Using this data we can plot the station locations that are included in
 GSOD that provide valid geo-locations after cleaning.
@@ -125,23 +125,62 @@ filter(Oz, STN_NAME == "TOOWOOMBA")
 #> 1 20120503 945510-99999             670    AUSTRALIA    AU   AUS
 ```
 
-## Using the `get_GSOD()` Function in GSODR to Download a Single Station and
+## Using the `get_GSOD` Function in GSODR to Download a Single Station and
 Year
 
 Download weather data from the station Toowoomba, Queensland, Australia for 2010
-and save it in the user's home directory using the STNID in the `station`
-parameter of `get_GSOD()`.
+by using the STNID in the `station` parameter of `get_GSOD`.
 
 ```r
-get_GSOD(years = 2010, station = "955510-99999", dsn = "~/",
-         filename = "Toowoomba_Airport")
+Tbar <- get_GSOD(years = 2010, station = "955510-99999")
+
+#> Downloading the station file(s) now.
+
+#> Finished downloading file. Parsing the station file(s) now.
+
+head(Tbar)
+#>    USAF  WBAN        STNID          STN_NAME CTRY STATE CALL    LAT     LON
+#>1 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>2 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>3 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>4 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>5 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>6 955510 99999 955510-99999 TOOWOOMBA AIRPORT   AS  <NA> <NA> -27.55 151.917
+#>  ELEV_M ELEV_M_SRTM_90m    BEGIN      END YEARMODA YEAR MONTH DAY YDAY TEMP
+#>1    642             635 19980301 20161020 20100101 2010    01  01    1 21.2
+#>2    642             635 19980301 20161020 20100102 2010    01  02    2 23.2
+#>3    642             635 19980301 20161020 20100103 2010    01  03    3 21.4
+#>4    642             635 19980301 20161020 20100104 2010    01  04    4 18.9
+#>5    642             635 19980301 20161020 20100105 2010    01  05    5 20.5
+#>6    642             635 19980301 20161020 20100106 2010    01  06    6 21.9
+#>  TEMP_CNT DEWP DEWP_CNT    SLP SLP_CNT   STP STP_CNT VISIB VISIB_CNT WDSP
+#>1        8 17.9        8 1013.4       8 942.0       8    NA         0  2.2
+#>2        8 19.4        8 1010.5       8 939.3       8    NA         0  1.9
+#>3        8 18.9        8 1012.3       8 940.9       8  14.3         6  3.9
+#>4        8 16.4        8 1015.7       8 944.1       8  23.3         4  4.5
+#>5        8 16.4        8 1015.5       8 944.0       8    NA         0  3.9
+#>6        8 18.7        8 1013.7       8 942.3       8    NA         0  3.2
[…]

[…]
