### 2016-10-24 f8ae4ca
Update README.Rmd/md
Add new ability to create data.frame in R
