### 2016-10-24 2979f81
Install libgeos-dev on Travis
rgeos is suggested
