### 2016-10-24 dfc1938
Second attempt at install rgeos on Travis
