+++ .travis.yml
@@ -22,8 +22,10 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
-  - install_aptget libgeos-dev
-  - install_aptget rgeos
+  - sudo apt-get -qq update
+  - sudo apt-get install libgeos
+  - sudo apt-get install libgeos-dev
+  - sudo apt-get install rgeos
   - cd /home/travis/build/adamhsparks/GSODR
 
 install:
