### 2016-10-24 622aa33
There is no  libgeos in Ubuntu
