+++ .travis.yml
@@ -6,7 +6,6 @@ sudo: required
 
 before_install:
   - sudo apt-get -qq update
-  - sudo apt-get install libgeos
   - sudo apt-get install libgeos-dev
   - sudo apt-get install rgeos
   - cd /home/travis/build/adamhsparks/GSODR
