### 2016-10-24 f6ab372
Still trying to install rgeos on Travis
