+++ .travis.yml
@@ -26,6 +26,7 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
+  - cd /home/travis/build/adamhsparks/GSODR
 
 install:
   - ./pkg-build.sh install_deps
