### 2016-10-24 e155cb7
Revert "Still trying to install rgeos on Travis"
This reverts commit f6ab3727cce55a1cad997fb54c12d60f897aece9.
