### 2016-10-24 0e36479
Revert "Travis Install libgeos-c1"
This reverts commit 548a44e06cd443ac79cd9aa0ea343b9ab86a12c9.
