### 2016-10-24 d26f6c4
Revert "There is no  libgeos in Ubuntu"
This reverts commit 622aa338e13c695645ca02e2106220c711f0a1e9.
