### 2016-10-24 ab32706
Revert "Install geos libraries first"
This reverts commit 15f02601a6bd032ef2960ac071771369e622f527.
