+++ .travis.yml
@@ -5,11 +5,6 @@ language: c
 sudo: required
 
 before_install:
-  - sudo apt-get -qq update
-  - sudo apt-get install libgeos
-  - sudo apt-get install libgeos-dev
-  - sudo apt-get install rgeos
-  - cd /home/travis/build/adamhsparks/GSODR
   - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
   - chmod 755 pkg-build.sh
   - ./pkg-build.sh bootstrap
@@ -27,6 +22,11 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
+  - sudo apt-get -qq update
+  - sudo apt-get install libgeos
+  - sudo apt-get install libgeos-dev
+  - sudo apt-get install rgeos
+  - cd /home/travis/build/adamhsparks/GSODR
 
 install:
   - ./pkg-build.sh install_deps
