### 2016-10-24 f813359
Revert "Second attempt at install rgeos on Travis"
This reverts commit dfc1938f0fd21bf17e6bfe00deda39b07c7f2030.
