+++ .travis.yml
@@ -22,10 +22,8 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
-  - sudo apt-get -qq update
-  - sudo apt-get install libgeos
-  - sudo apt-get install libgeos-dev
-  - sudo apt-get install rgeos
+  - install_aptget libgeos-dev
+  - install_aptget rgeos
   - cd /home/travis/build/adamhsparks/GSODR
 
 install:
