### 2016-10-24 f269841
Revert "Install libgeos-dev on Travis"
This reverts commit 2979f81229eed1cb3b83b4087972ea3b45fc7172.
