### 2016-10-24 5131945
Install rgeos from RStudio
