### 2016-10-24 c55633c
Set language to R in travis.yml add CRAN environment
