+++ .travis.yml
@@ -1,28 +1,33 @@
 ## Sample  .travis.yml file for use with metacran/r-builder
 ## See https://github.com/metacran/r-builder for details.
 
-language: c
+language: r
 sudo: required
+cache: packages
 
-before_install:
-  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
-  - chmod 755 pkg-build.sh
-  - ./pkg-build.sh bootstrap
-  - cd
-  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
-  - tar zxf gdal-2.0.1.tar.gz
-  - cd gdal-2.0.1
-  - ./configure --prefix=/usr
-  - make
-  - sudo make install
-  - cd
-  - wget http://download.osgeo.org/proj/proj-4.9.2.tar.gz
-  - tar zxvf proj-4.9.2.tar.gz
-  - cd proj-4.9.2
-  - ./configure --prefix=/usr
-  - make
-  - sudo make install
-  - cd /home/travis/build/adamhsparks/GSODR
+env:
+ global:
+   - CRAN: http://cran.rstudio.com
+
+#before_install:
+#  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
+#  - chmod 755 pkg-build.sh
+#  - ./pkg-build.sh bootstrap
+#  - cd
+#  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
+#  - tar zxf gdal-2.0.1.tar.gz
+#  - cd gdal-2.0.1
+#  - ./configure --prefix=/usr
+#  - make
+#  - sudo make install
+#  - cd
+#  - wget http://download.osgeo.org/proj/proj-4.9.2.tar.gz
+#  - tar zxvf proj-4.9.2.tar.gz
+#  - cd proj-4.9.2
+#  - ./configure --prefix=/usr
+#  - make
+#  - sudo make install
+#  - cd /home/travis/build/adamhsparks/GSODR
 
 install:
   - Rscript -e 'install.packages("rgeos", repos = "http://cran.rstudio.com")'
