### 2016-10-24 e3e47a6
Uncomment pkg-build.sh in travis.yml
