+++ .travis.yml
@@ -9,10 +9,10 @@ env:
  global:
    - CRAN: http://cran.rstudio.com
 
-#before_install:
-#  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
-#  - chmod 755 pkg-build.sh
-#  - ./pkg-build.sh bootstrap
+before_install:
+  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
+  - chmod 755 pkg-build.sh
+  - ./pkg-build.sh bootstrap
 #  - cd
 #  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
 #  - tar zxf gdal-2.0.1.tar.gz
