### 2016-10-24 c389e99
install all dependencies and suggests manually
