+++ .travis.yml
@@ -2,6 +2,7 @@
 ## See https://github.com/metacran/r-builder for details.
 
 language: r
+warnings_are_errors: false
 sudo: required
 cache: packages
 
@@ -10,34 +11,42 @@ env:
    - CRAN: http://cran.rstudio.com
 
 before_install:
-  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
-  - chmod 755 pkg-build.sh
-  - ./pkg-build.sh bootstrap
-#  - cd
-#  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
-#  - tar zxf gdal-2.0.1.tar.gz
-#  - cd gdal-2.0.1
-#  - ./configure --prefix=/usr
-#  - make
-#  - sudo make install
-#  - cd
-#  - wget http://download.osgeo.org/proj/proj-4.9.2.tar.gz
-#  - tar zxvf proj-4.9.2.tar.gz
-#  - cd proj-4.9.2
-#  - ./configure --prefix=/usr
-#  - make
-#  - sudo make install
-#  - cd /home/travis/build/adamhsparks/GSODR
+  - sudo apt-get install libgeos-dev
+  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
+  - tar zxf gdal-2.0.1.tar.gz
+  - cd gdal-2.0.1
+  - ./configure --prefix=/usr
+  - make
+  - sudo make install
+  - cd
+  - wget http://download.osgeo.org/proj/proj-4.9.2.tar.gz
+  - tar zxvf proj-4.9.2.tar.gz
+  - cd proj-4.9.2
+  - ./configure --prefix=/usr
+  - make
+  - sudo make install
 
 install:
+  - Rscript -e 'install.packages("data.table", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("doParallel", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("foreach", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("iterators", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("parallel", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("RCurl", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("readr", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("rgdal", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("sp", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("chillR", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("dplyr", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("ggalt", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("knitr", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("lubridate", repos = "http://cran.rstudio.com")'
+  - Rscript -e 'install.packages("plotKML", repos = "http://cran.rstudio.com")'
   - Rscript -e 'install.packages("rgeos", repos = "http://cran.rstudio.com")'
-  - ./pkg-build.sh install_deps
-
-script:
-  - ./pkg-build.sh run_tests
-
-after_failure:
-  - ./pkg-build.sh dump_logs
+  - Rscript -e 'install.packages("rmarkdown", repos = "http://cran.rstudio.com")'
[…]
