+++ .travis.yml
@@ -11,6 +11,10 @@ env:
    - CRAN: http://cran.rstudio.com
 
 before_install:
+  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
+  - chmod 755 pkg-build.sh
+  - ./pkg-build.sh bootstrap
+  - cd
   - sudo apt-get install libgeos-dev
   - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
   - tar zxf gdal-2.0.1.tar.gz
@@ -25,6 +29,7 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
+  - cd /home/travis/build/adamhsparks/GSODR
 
 install:
   - Rscript -e 'install.packages("data.table", repos = "http://cran.rstudio.com")'
@@ -46,7 +51,15 @@ install:
   - Rscript -e 'install.packages("rmarkdown", repos = "http://cran.rstudio.com")'
   - Rscript -e 'install.packages("spacetime", repos = "http://cran.rstudio.com")'
   - Rscript -e 'install.packages("testthat", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("tidyr", repos = "http://cran.rstudio.com")' 
+  - Rscript -e 'install.packages("tidyr", repos = "http://cran.rstudio.com")'
+
+  - ./pkg-build.sh install_deps
+
+  script:
+  - ./pkg-build.sh run_tests
+
+after_failure:
+  - ./pkg-build.sh dump_logs
 
 notifications:
   email:
