### 2016-10-25 2b4dfc4
Lint .travis.yml
