+++ .travis.yml
@@ -52,7 +52,6 @@ install:
   - Rscript -e 'install.packages("spacetime", repos = "http://cran.rstudio.com")'
   - Rscript -e 'install.packages("testthat", repos = "http://cran.rstudio.com")'
   - Rscript -e 'install.packages("tidyr", repos = "http://cran.rstudio.com")'
-
   - ./pkg-build.sh install_deps
 
   script:
