### 2016-10-25 f12ab77
Fix issues with .travis.yml
