+++ .travis.yml
@@ -1,18 +1,19 @@
 ## Sample  .travis.yml file for use with metacran/r-builder
 ## See https://github.com/metacran/r-builder for details.
 
-language: r
-warnings_are_errors: false
+language: c
 sudo: required
-cache: packages
 
 env:
  global:
    - CRAN: http://cran.rstudio.com
 
 before_install:
+before_install:
   - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
   - chmod 755 pkg-build.sh
+  - cp pkg-build.sh rbuildertest
+  - cd rbuildertest
   - ./pkg-build.sh bootstrap
   - cd
   - sudo apt-get install libgeos-dev
@@ -32,26 +33,6 @@ before_install:
   - cd /home/travis/build/adamhsparks/GSODR
 
 install:
-  - Rscript -e 'install.packages("data.table", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("doParallel", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("foreach", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("iterators", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("parallel", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("RCurl", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("readr", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("rgdal", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("sp", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("chillR", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("dplyr", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("ggalt", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("knitr", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("lubridate", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("plotKML", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("rgeos", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("rmarkdown", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("spacetime", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("testthat", repos = "http://cran.rstudio.com")'
-  - Rscript -e 'install.packages("tidyr", repos = "http://cran.rstudio.com")'
   - ./pkg-build.sh install_deps
 
   script:
