### 2016-10-25 99d0f56
Cleanup .travis.yml, still trying to install rgeos
