### 2016-10-25 1851ee1
Cleanup in .travis.yml
