+++ .travis.yml
@@ -1,6 +1,3 @@
-## Sample  .travis.yml file for use with metacran/r-builder
-## See https://github.com/metacran/r-builder for details.
-
 language: c
 sudo: required
 
@@ -43,4 +40,4 @@ env:
   matrix:
     - RVERSION=oldrel
     - RVERSION=release
-- RVERSION=devel
+    - RVERSION=devel
