### 2016-10-25 3275131
Download and install GEOS (travis) and update documentation
