+++ R/get_GSOD.R
@@ -7,16 +7,16 @@
 #'and calculates three new variables; Saturation Vapor Pressure (ES) – Actual
 #'Vapor Pressure (EA) and relative humidity (RH). Stations are individually
 #'checked for number of missing days to assure data quality, stations with too
-#'many missing observations are omitted, stations with a latitude of < -90 or >
-#'90 or longitude of < -180 or > 180 are removed. All units are converted to
-#'International System of Units (SI), e.g., Fahrenheit to Celsius and inches to
-#'millimetres. Alternative elevation measurements are supplied for missing
-#'values or values found to be questionable based on the Consultative Group
-#'for International Agricultural Research's Consortium for Spatial Information
-#'group's (CGIAR-CSI) Shuttle Radar Topography Mission 90 metre (SRTM 90m)
-#'digital elevation data based on NASA's original SRTM 90m data. Further
-#'information on these data and methods can be found on GSODR's GitHub
-#'repository here:
+#'many missing observations are omitted. Additionally, stations reporting a
+#'latitude of < -90 or > 90 or longitude of < -180 or > 180 are removed. All
+#'units are converted to International System of Units (SI), e.g., Fahrenheit to
+#'Celsius and inches to millimetres. Alternative elevation measurements are
+#'supplied for missing values or values found to be questionable based on the
+#'Consultative Group for International Agricultural Research's Consortium for
+#'Spatial Information group's (CGIAR-CSI) Shuttle Radar Topography Mission 90
+#'metre (SRTM 90m) digital elevation data based on NASA's original SRTM 90m
+#'data. Further information on these data and methods can be found on GSODR's
+#'GitHub repository here:
 #'\url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
 #'
 #' @param years Year(s) of weather data to download.
@@ -33,10 +33,10 @@
 #' having an ISO code 2 or 3 letter ISO code can also be used if known. See
 #' \code{\link{country_list}} for a full list of country names and ISO
 #' codes available.
-#' @param dsn Path to file write to.
+#' @param dsn Path to file write to. (Optional)
 #' @param filename The filename for resulting file(s) to be written with no
 #' file extension. Year and file extension will be automatically appended to
-#' file outputs. Defaults to "GSOD-year".
+#' file outputs. Defaults to "GSOD-year". (Optional)
 #' @param max_missing The maximum number of days allowed to be missing from a
 #' station's data before it is excluded from final file output. Defaults to five
 #' days. If a single station is specified, this option is ignored and any data
@@ -46,31 +46,27 @@
 #' TRUE to include only stations within the confines of these
 #' latitudes.
 #' @param CSV Logical. If set to TRUE, create a comma separated value (CSV)
-#' file of data, defaults to TRUE, a CSV file is created.
-#' @param GPKG Logical. If set to TRUE, create a GeoPackage file, if
-#' set to FALSE, no GPKG file is created. Defaults to FALSE, no GPKG file is
-#' created.
+#' file ile and save it locally in a user specified location. Depends on
+#' \code{dsn} being specified.
+#' @param GPKG Logical. If set to TRUE, create a GeoPackage file and save it
+#' locally in a user specified location. Depends on \code{dsn} being specified.
 #' @param threads The number of computing threads to use for parallel processing
[…]
