+++ .travis.yml
@@ -6,7 +6,12 @@ before_install:
   - chmod 755 pkg-build.sh
   - ./pkg-build.sh bootstrap
   - cd
-  - sudo apt-get install libgeos-dev
+  - wget http://download.osgeo.org/geos/geos-3.5.0.tar.bz2
+  - tar xjf geos-3.5.0.tar.bz2
+  - cd geos-3.5.0
+  - ./configure
+  - make
+  - sudo make install
   - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
   - tar zxf gdal-2.0.1.tar.gz
   - cd gdal-2.0.1
