### 2016-10-25 1a4d785
Add prefix for geos installation in .travis.ym.
