+++ .travis.yml
@@ -9,7 +9,7 @@ before_install:
   - wget http://download.osgeo.org/geos/geos-3.5.0.tar.bz2
   - tar xjf geos-3.5.0.tar.bz2
   - cd geos-3.5.0
-  - ./configure
+  - ./configure --prefix=/usr
   - make
   - sudo make install
   - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
