### 2016-10-25 3a07757
Updates to `nearest_stations` function
* `nearest_stations` function now returns a list of stations for use in
the `get_GSOD` function's `station` parameter.

* the `nearest_stations` function now is more computationally efficient
thanks to @jefferyhanson's suggestion

* Updates to documentation for clarity and to reflect new functionality
of this function

* Add fields package to list of depends in DESCRIPTION for this function
