### 2016-10-25 #79 @adamhsparks
Updated `nearest_stations` function to return a list of stations for use in the `get_GSOD` function's `stations` parameter.

> The nearest_station function will be incredibly useful for a user. It would be even more useful if--instead of printing the stations closest to a user-specified point--it returned the stations. Then this function could be used programmatically to find the closest stations to a series of points, and download data accordingly.

Also rewrote the function for computational efficiency.

> nearest_stations.R:34--44: This code could be rewritten for clarity and computational efficiency.

https://github.com/adamhsparks/GSODR/commit/3a077575aaee54543ee4ee3aa93b9950c5355d80
