### 2016-10-26 3e3e30e
Updates to paper.md/.bib and documentation files
