+++ inst/paper/paper.bib
@@ -5,7 +5,7 @@
     address = {Vienna, Austria},
     year = {2016},
     url = {https://www.R-project.org/}
-    }
+  }
 
   @Misc{NCDC,
     author = {{United States National Oceanic and Atmospheric Administration National Climatic Data Center}},
@@ -14,7 +14,7 @@
     year = {2016},
     howpublished = {\url{https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod}},
     lastchecked = {10/08/2016}},
-    }
+  }
 
   @Manual{rnoaa,
     title = {rnoaa: 'NOAA' Weather Data from R},
@@ -22,22 +22,13 @@
     year = {2016},
     note = {R package version 0.5.6},
     url = {https://CRAN.R-project.org/package=rnoaa},
-    }
-
-  @Book{Hengl2009,
-    title = {A Practical Guide to Geostatistical Mapping},
-    author = {Tomislav Hengl},
-    year = {2009},
-    publisher = {University of Amsterdam},
-    isbn = {ISBN 978-90-9024981-0},
-    url = {http://spatial-analyst.net/book/},
-    }
+  }
 
   @Misc{Jarvis2008,
-  title = {{Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database}},
-  author = {Andy Jarvis and Hannes I Reuter and Andy Nelson and Edward Guevara},
-  year = {2008},
-  url = {http://srtm.csi.cgiar.org}
+    title = {{Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database}},
+    author = {Andy Jarvis and Hannes I Reuter and Andy Nelson and Edward Guevara},
+    year = {2008},
+    url = {http://srtm.csi.cgiar.org}
   }
 
   @Manual{GSODR,
@@ -49,13 +40,13 @@
   }
 
   @Misc{geopackage,
-  author = {{Open Geospatial Consortium}},
-  title = {GeoPackage Encoding Standard},
-  url = {http://www.opengeospatial.org/standards/geopackage},
-  year = {2014}
+    author = {{Open Geospatial Consortium}},
+    title = {GeoPackage Encoding Standard},
+    url = {http://www.opengeospatial.org/standards/geopackage},
+    year = {2014}
   }
 
-   @Manual{readr,
+  @Manual{readr,
     title = {readr: Read Tabular Data},
     author = {Hadley Wickham and Jim Hester and Romain Francois},
     year = {2016},
@@ -63,7 +54,7 @@
     url = {https://CRAN.R-project.org/package=readr},
   }
 
-    @Manual{data.table,
+  @Manual{data.table,
     title = {data.table: Extension of Data.frame},
     author = {M Dowle and A Srinivasan and T Short and S Lianoglou with contributions from R Saporta and E Antonyan},
     year = {2015},
@@ -71,7 +62,7 @@
     url = {https://CRAN.R-project.org/package=data.table},
   }
 
-   @Manual{foreach,
+  @Manual{foreach,
     title = {foreach: Provides Foreach Looping Construct for R},
     author = {{Revolution Analytics} and Steve Weston},
     year = {2015},
@@ -79,3 +70,12 @@
     url = {https://CRAN.R-project.org/package=foreach},
   }
 
+  @article{Kilibarda2014,
+    title = {Spatio‐temporal interpolation of daily temperatures for global land areas at 1 km resolution},
+    author = {Kilibarda, M and Hengl, T and Heuvelink, G and Gräler, B and Pebesma, E and Perčec Tadić, M and Bajat, B},
+    year = {2014},
+    journal = {Journal of Geophysical Research: Atmospheres},
+    volume = {119},
+    issue = {5},
+    pages = {2294-2313},
+  }

[…]
