### 2016-10-26 f26d11a
Update .gitignore, Update README.Rmd file
ignore .kml and .kmz files generated during README.md knitting

README.Rmd example figure now has average monthly temperature from
GSOD, daily temperature and CHELSA temperature plotted.
