+++ README.md
@@ -13,19 +13,27 @@ downloads](http://cranlogs.r-pkg.org/badges/GSODR?color=blue)](https://github.co
 version](http://www.r-pkg.org/badges/version/GSODR)](https://cran.r-project.org/package=GSODR)
 [![DOI](https://zenodo.org/badge/32764550.svg)](https://zenodo.org/badge/latestdoi/32764550)
 
+Introduction to GSODR
+=====================
+
 The GSODR package is an R package that provides a function that
-automates downloading and cleaning data from the "[Global Surface
+automates downloading and cleaning of data from the "[Global Surface
 Summary of the Day
 (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)"
 weather station data provided by the US National Climatic Data Center
-(NCDC). Station files are individually checked for number of missing
-days to assure data quality, stations with too many missing observations
-are omitted. All units are converted to International System of Units
-(SI), e.g., inches to millimetres and Fahrenheit to Celsius. Output is
-saved as a Comma Separated Value (CSV) file or in a spatial GeoPackage
-(GPKG) file, implemented by most major GIS software, summarising each
-year by station, which also includes vapour pressure and relative
-humidity variables calculated from existing data in GSOD.
+(NCDC). Data are formatted for easy use in R, returned as a `data.frame`
+object in R summarising each year by station, with options to save as a
+Comma Separated Value (CSV) file or as a spatial GeoPackage (GPKG) file,
+implemented by most major GIS softwares.
+
+Station files are individually checked for number of missing days to
+assure data quality, stations with too many missing observations are
+omitted. All units are converted to International System of Units (SI),
+e.g., inches to millimetres and Fahrenheit to Celsius.
+
+Additional data are calculated by this R package using the original data
+and included in the final data. These include vapour pressure (ea and
+es) and relative humidity.
 
 There are several other sources of weather data and ways of retrieving
 them through R. In particular, the excellent
@@ -61,8 +69,8 @@ A development version is available from from GitHub. If you wish to
 install the development version that may have new features (but also may
 not work properly), install the [devtools
 package](https://CRAN.R-project.org/package=devtools), available from
-CRAN. I strive to keep the master branch on GitHub functional and
-working properly this may not always happen.
+CRAN. We strive to keep the master branch on GitHub functional and
+working properly, although this may not always happen.
 
 If you find bugs, please file a report as an issue.
 
@@ -79,11 +87,11 @@ Using GSODR
 GSODR's main function, `get_GSOD`, downloads and cleans GSOD data from
 the NCDC server. Following are a few examples of its capabilities.
 
+#### Example 1 - Download weather station for Toowoomba, Queensland for 2010
+
 ``` r
 
 library(GSODR)
-
-# Download weather station for Toowoomba, Queensland for 2010
 
[…]
