### 2016-10-26 f388f99
Lint package
Shorten line lengths
