### 2016-10-27 3959617
Fix station/year file list for fetching specific stations
