### 2016-10-28 e25788a
Major rewrite of package structure
Remove foreach %do% loops, use ply::ldply instead

remove many extra functions from get_GSOD.R and put into files for file
handling, file_handling.R; validation, validation_functions.R; and file
processing, process_files.R

Add plyr to list of depends
