+++ R/get_GSOD.R
@@ -214,7 +214,6 @@
 #' \url{http://srtm.csi.cgiar.org}}
 #'
 #' @importFrom foreach %dopar%
-#' @importFrom foreach %do%
 #' @importFrom data.table :=
 #'
 #' @export
@@ -223,22 +222,22 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                      agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE,
                      threads = 1) {
 
-  # Set up options, creating objects, check variables entered by user-----------
+  # Set up options, create objects, fetch most recent station metadata ---------
   original_options <- options()
   options(warn = 2)
   options(timeout = 300)
 
+  s <- yr <- LON <- LAT <- NULL
 
-  # Set up tempfile and directory for downloading data from server
   tf <- tempfile()
   td <- tempdir()
 
-  # Create objects for use later
-  s <- j <- yr <- LON <- LAT <-  NULL
+  ftp_site <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
 
   # fetch most recent station history file
-  stations <- .fetch_stations()
+  stations <- .fetch_station_list()
 
+  # Validate -------------------------------------------------------------------
   # Check data path given by user, does it exist? Is it properly formatted?
   if (!is.null(dsn)) {
     dsn <- .validate_dsn(dsn)
@@ -246,8 +245,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   # Check years given by the user, are they valid?
   .validate_years(years)
 
-  # Check station given by user, is it valid, are the years for this station
-  # valid?
+  # Check station given by user and check that years for this station are valid
   if (!is.null(station)) {
     .validate_station(station, stations)
     .validate_station_years(station, stations, years)
@@ -264,70 +262,49 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     max_missing <- 366
   }
 
-  ftp_site <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
+  # Download data --------------------------------------------------------------
 
-  # Year loop ------------------------------------------------------------------
-  ity <- iterators::iter(years)
-  foreach::foreach(yr = ity) %do% {
-    if (is.null(station)) {
+  # Global or Agroclimatology
+  if (is.null(station)) {
+    message("\nDownloading the data file(s) now.")
+    s <- paste0(ftp_site, years, "/")
+    GSOD_XY <- plyr::ldply(.data = s, .fun = .dl_global_files)
 
+  } else {
+    # Individual stations
+    message("\nDownloading the station file(s) now.")
+    s <- paste0(ftp_site, years, "/")
+    s <- do.call(paste0, c(expand.grid(s, stations)))
+    s <- paste0(s, "-", years, ".op.gz")
+    plyr::ldply(.data = s, .fun = .dl_specified_stations)
 
-    # THIS WORKS NEED TO MODIFY ALL OTHER INSTANCES/REMOVE YEAR LOOP------------
-    if (!is.null(station)) {
-      message("\nDownloading the station file(s) now.")
+  }
 
-      s <- paste0(ftp_site, years, "/")
-      s <- do.call(paste0, c(expand.grid(s, stations)))
-      s <- paste0(s, "-", years, ".op.gz")
+  #### Write to disk ---------------------------------------------------------
+  if (!is.null(dsn)) {
+    outfile <- paste0(dsn, filename)
 
[…]
