+++ DESCRIPTION
@@ -40,6 +40,7 @@ Imports:
     foreach,
     iterators,
     parallel,
+    plyr,
     RCurl,
     readr,
     rgdal,

[…]
