+++ NAMESPACE
@@ -3,5 +3,4 @@
 export(get_GSOD)
 export(nearest_stations)
 importFrom(data.table,":=")
-importFrom(foreach,"%do%")
 importFrom(foreach,"%dopar%")
