### 2016-10-28 d09e362
Fix bug when pasting file names on line 277
