### 2016-10-30 c2a2770
Updates and bug fixes
Downloading single or specific stations is fixed, now using plyr

Specify to use parallel processing

Check if stations file has been retrieved from server, if not then
fetch it

Streamline file handling and downloading using plyr
