### 2016-10-30 ef6985c
Fix errors and warnings with package check
