### 2016-10-31 e6e820d
Remove parallel processing options and plyr
Use data.table only to reduce dependencies

Parallelisation doesn't work with ldply from issues listed on plyr's
GitHub repository
