+++ DESCRIPTION
@@ -35,11 +35,8 @@ Depends:
 License: GPL (>= 3)
 Imports:
     data.table,
-    doParallel,
     fields,
-    foreach,
     iterators,
-    parallel,
     plyr,
     RCurl,
     readr,
