### 2016-10-31 2283057
Put ply::ldply back in
This seems to be the most reasonable route for working with this for
now, even without the option of running in parallel.
