### 2016-10-31 b7cd021
Fix bug when checking for stations
In nearest_stations.R fix bug when checking to see if stations list
already has been retrieved or not
