### 2016-10-31 91f0cfd
Add functionality to check record length
Not yet working, but adding back the check for `max_missing`
