+++ R/file_handling.R
@@ -1,11 +1,12 @@
 # Functions used in GSODR for handling files -----------------------------------
 
 #'@noRd
-.dl_global_files <- function(agroclimatology, country, s, stations, td, years) {
+.dl_global_files <- function(agroclimatology, country, file_list, stations, td,
+                             years) {
 
   tryCatch(Map(function(ftp, dest)
     utils::download.file(url = ftp, destfile = dest),
-    s, file.path(td, basename(s))), error = function(x) stop(
+    file_list, file.path(td, basename(file_list))), error = function(x) stop(
       "\nThe file downloads have failed. Please restart.\n"))
 
   tar_files <- list.files(td, pattern = "^gsod.*\\.tar$", full.names = TRUE)
@@ -13,6 +14,8 @@
   plyr::ldply(.data = tar_files, .fun = utils::untar, exdir = td)
 
   GSOD_list <- list.files(td, pattern = "^.*\\.op.gz$", full.names = FALSE)
+  
+  GSOD_list <- plyr::llply(.data = GSOD_list, .function = .check_missing)
 
   # If agroclimatology == TRUE, subset list of stations to process--------------
   if (agroclimatology == TRUE) {
@@ -34,23 +37,23 @@
                             c(expand.grid(station_list, "-", years, ".op.gz")))
     GSOD_list <- paste0(td, "/", GSOD_list[GSOD_list %in% station_list == TRUE])
   }
-  data.table::data.table(.fun = .process_gz(GSOD_list, stations = stations))
+  plyr::ldply(.data = GSOD_list, .fun = .process_gz, stations = stations)
 }
 
 
 #' @noRd
-.dl_specified_stations <- function(s, stations, td) {
+.dl_specified_stations <- function(file_list, stations, td) {
 
-  filenames <- paste0(substr(s, 1, 43),
-                      strsplit(RCurl::getURL(substr(s, 1, 43),
+  filenames <- paste0(substr(file_list, 1, 43),
+                      strsplit(RCurl::getURL(substr(file_list, 1, 43),
                                              ftp.use.epsv = FALSE,
                                              ftplistonly = TRUE, crlf = TRUE),
                                "\r*\n")[[1]])[-c(1:2)]
-  s <- filenames[which(s %in% filenames)]
+  file_list <- filenames[which(file_list %in% filenames)]
 
   tryCatch(Map(function(ftp, dest)
     utils::download.file(url = ftp, destfile = dest),
-    s, file.path(td, basename(s))),
+    file_list, file.path(td, basename(file_list))),
     error = function(x) message(paste0(
       "\nThe file downloads have failed. Please restart.\n")))
 
