### 2016-10-31 b25c8b0
Lint data_manipulation.R file
