+++ R/data_manipulation.R
@@ -80,17 +80,17 @@
                                                         tmp$DAY,
                                                         sep = "-")),
                                           format = "%j"))]
-  tmp[, (TEMP)  := round( ((5 / 9) * ((tmp$TEMP) - 32)), 1)]
-  tmp[, (DEWP)  := round( ((5 / 9) * ((tmp$DEWP) - 32)), 1)]
-  tmp[, (WDSP)  := round( (tmp$WDSP) * 0.514444444, 1)]
-  tmp[, (MXSPD) := round( (tmp$MXSPD) * 0.514444444, 1)]
-  tmp[, (VISIB) := round( (tmp$VISIB) * 1.60934, 1)]
-  tmp[, (WDSP)  := round( (tmp$WDSP) * 0.514444444, 1)]
-  tmp[, (GUST)  := round( (tmp$GUST) * 0.514444444, 1)]
-  tmp[, (MAX)   := round( ((tmp$MAX) - 32) * (5 / 9), 2)]
-  tmp[, (MIN)   := round( ((tmp$MIN) - 32) * (5 / 9), 2)]
-  tmp[, (PRCP)  := round( ((tmp$PRCP) * 25.4), 1)]
-  tmp[, (SNDP)  := round( ((tmp$SNDP) * 25.4), 1)]
+  tmp[, (TEMP)  := round(((5 / 9) * ((tmp$TEMP) - 32)), 1)]
+  tmp[, (DEWP)  := round(((5 / 9) * ((tmp$DEWP) - 32)), 1)]
+  tmp[, (WDSP)  := round((tmp$WDSP) * 0.514444444, 1)]
+  tmp[, (MXSPD) := round((tmp$MXSPD) * 0.514444444, 1)]
+  tmp[, (VISIB) := round((tmp$VISIB) * 1.60934, 1)]
+  tmp[, (WDSP)  := round((tmp$WDSP) * 0.514444444, 1)]
+  tmp[, (GUST)  := round((tmp$GUST) * 0.514444444, 1)]
+  tmp[, (MAX)   := round(((tmp$MAX) - 32) * (5 / 9), 2)]
+  tmp[, (MIN)   := round(((tmp$MIN) - 32) * (5 / 9), 2)]
+  tmp[, (PRCP)  := round(((tmp$PRCP) * 25.4), 1)]
+  tmp[, (SNDP)  := round(((tmp$SNDP) * 25.4), 1)]
 
   # Compute other weather vars--------------------------------------------------
 
@@ -99,10 +99,10 @@
   #   Edward Arnold, London
 
   # EA derived from dew point
-  tmp[, (EA) := round(0.61078 * exp( (17.2694 * (tmp$DEWP)) /
+  tmp[, (EA) := round(0.61078 * exp((17.2694 * (tmp$DEWP)) /
                                        ((tmp$DEWP) + 237.3)), 1)]
   # ES derived from average temperature
-  tmp[, (ES) := round(0.61078 * exp( (17.2694 * (tmp$TEMP)) /
+  tmp[, (ES) := round(0.61078 * exp((17.2694 * (tmp$TEMP)) /
                                        ((tmp$TEMP) + 237.3)), 1)]
   # Calculate relative humidity
   tmp[, (RH) := round(tmp$EA / tmp$ES * 100, 1)]
