### 2016-11-01 9a7323b
Parallel option and bug fixes
Still not completely functional

Parallel processing is optional again

max_missing check function works
