+++ NAMESPACE
@@ -3,3 +3,4 @@
 export(get_GSOD)
 export(nearest_stations)
 importFrom(data.table,":=")
+importFrom(foreach,"%dopar%")
