### 2016-11-01 e6d96e0
Bug fixes
Single station still fails when processing files. Need to convert to
data.frame
