### 2016-11-02 #79 @deniederhut
Estimated hours spent reviewing: Maybe 2? This was conducted in ~20 minute chunks over the last two weeks, which makes it a little hard to guesstimate.

---

- `vignettes` and `browseVignettes` function called on 'GSODR' returns `no vignettes found`
- :+1: for including a code of conduct
- it seems odd to me the enforce a missingness limit at download time -- I think I would either let the user do this in memory, or set the default limit to 0
- the test suite only seems to be concerned with handling bad input - am I missing the blocks where it is testing that data are returned correctly?
- allowing multiple threads for calculation doesn't speed up the execution time as much as I would expect:

```
> system.time(get_GSOD(years = 2010, country = "Philippines", dsn = "~/", filename = "PHL",
+          GPKG = TRUE, CSV = FALSE, threads = 2))
trying URL 'ftp://ftp.ncdc.noaa.gov/pub/data/gsod/2010/gsod_2010.tar'
Content type 'unknown' length 98222080 bytes (93.7 MB)
==================================================
downloaded 93.7 MB

Finished downloading file.

Parsing the indivdual station files now.

   user  system elapsed
  4.391  11.714 147.535
>
> system.time(get_GSOD(years = 2010, country = "Philippines", dsn = "~/", filename = "PHL",
+          GPKG = TRUE, CSV = FALSE, threads = 1))
trying URL 'ftp://ftp.ncdc.noaa.gov/pub/data/gsod/2010/gsod_2010.tar'
Content type 'unknown' length 98222080 bytes (93.7 MB)
==================================================
downloaded 93.7 MB

Finished downloading file.

Parsing the indivdual station files now.

   user  system elapsed
  4.350  18.334 147.915
```
