### 2016-11-02 #79 @sckott
Thanks @jeffreyhanson and @deniederhut for your reviews!  both reviews in now, @adamhsparks continue on, and let us know if you have any questions, want any feedback
