### 2016-11-02 #79 @adamhsparks
Thanks folks. Excellent reviews and points raised in them.

I've been working through the review of @jeffreyhanson and will work on @deniederhut's as well.

I think this process is well worth it, especially for someone like me that is just learning how to put a package together. This will make this a much better package than it otherwise would have been.
