### 2016-11-08 94f78b1
Revert branch back to working state
Discard most changes with ldply, etc. and revert to working state
