+++ R/get_GSOD.R
@@ -281,6 +281,119 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     file_list <- paste0(file_list, "-", years, ".op.gz")
     GSOD_XY <- plyr::ldply(.data = file_list, .fun = .dl_specified_stations,
                            stations = stations, td = td, .progress = "text")
+  ftp_site <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
+
+  # Year loop ------------------------------------------------------------------
+  ity <- iterators::iter(years)
+  foreach::foreach(yr = ity) %do% {
+    if (is.null(station)) {
+      tryCatch(utils::download.file(url = paste0(ftp_site, yr, "/gsod_", yr,
+                                                 ".tar"),
+                                    destfile = tf, mode = "wb"),
+               error = function(x) message(paste0(
+                 "\nThe download stopped at year: ", yr, ".
+                \nPlease restart the 'get_GSOD()' function starting here.\n")))
+      utils::untar(tarfile = tf, exdir  = paste0(td, "/", yr, "/"))
+
+      message("\nFinished downloading file.
+              \nParsing the indivdual station files now.\n")
+      GSOD_list <- list.files(paste0(td, "/", yr, "/"),
+                              pattern = "^.*\\.gz$",
+                              full.names = FALSE)
+
+      # If agroclimatology == TRUE, subset list of stations to clean -----------
+      if (agroclimatology == TRUE) {
+        station_list <- stations[stations$LAT >= -60 &
+                                   stations$LAT <= 60, ]$STNID
+        station_list <- vapply(station_list,
+                               function(x) rep(paste0(x, "-", yr, ".op.gz")),
+                               "")
+        GSOD_list <- GSOD_list[GSOD_list %in% station_list == TRUE]
+        rm(station_list)
+      }
+
+      # If country is set, subset list of stations to clean --------------------
+      if (!is.null(country)) {
+        country_FIPS <- unlist(as.character(stats::na.omit(
+          GSODR::country_list[GSODR::country_list$FIPS == country, ][[1]]),
+          use.names = FALSE))
+        station_list <- stations[stations$CTRY == country_FIPS, ]$STNID
+        station_list <- vapply(station_list,
+                               function(x) rep(paste0(x, "-", yr, ".op.gz")),
+                               "")
+        GSOD_list <- GSOD_list[GSOD_list %in% station_list == TRUE]
+      }
+    }
+
+    # Stations specified --------------------------- ---------------------------
+    if (!is.null(station)) {
+      message("\nDownloading the station file(s) now.")
+      filenames <- RCurl::getURL(paste0(ftp_site, yr, "/"),
+                                 ftp.use.epsv = FALSE, ftplistonly = TRUE,
+                                 crlf = TRUE)
+      filenames <- paste0(ftp_site, yr, "/",
+                          strsplit(filenames, "\r*\n")[[1]])[-c(1:2)]
+      itw <- (iterators::iter(station))
+      message("\nFinished downloading file. Parsing the station file(s) now.\n")
+      GSOD_XY <- as.data.frame(
+        data.table::rbindlist(
+          foreach::foreach(s = itw) %do% {
+            s <- paste0(ftp_site, yr, "/", s, "-", yr, ".op.gz")
[…]
