### 2016-11-08 52e303a
Update documentation, remove parallelisation
Still not functional
