+++ R/get_GSOD.R
@@ -20,38 +20,34 @@
 #'\url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
 #'
 #' @param years Year(s) of weather data to download.
-#' @param station Specify a station or multiple stations for which to retrieve,
-#' check and clean weather data. The NCDC reports years for which the data are
-#' available. This function checks against these years. However, not all cases
-#' are properly documented and in some cases files may not exist on the ftp
-#' server even though it is indicated that data was recorded for the station for
-#' a particular year. If a station is specified that does not have an
-#' existing file on the server, this function will silently fail and move on
+#' @param station Optional. Specify a station or multiple stations for which to
+#' retrieve, check and clean weather data. The NCDC reports years for which the
+#' data are available. This function checks against these years. However, not
+#' all cases are properly documented and in some cases files may not exist on
+#' the ftp server even though it is indicated that data was recorded for the
+#' station for a particular year. If a station is specified that does not have
+#' an existing file on the server, this function will silently fail and move on
 #' to existing files for download and cleaning from the ftp server.
-#' @param country Specify a country of interest for which to retrieve weather
-#' data; full name. For stations located in locales
-#' having an ISO code 2 or 3 letter ISO code can also be used if known. See
-#' \code{\link{country_list}} for a full list of country names and ISO
-#' codes available.
-#' @param dsn Path to file write to. (Optional)
-#' @param filename The filename for resulting file(s) to be written with no
-#' file extension. Year and file extension will be automatically appended to
-#' file outputs. Defaults to "GSOD-year". (Optional)
-#' @param max_missing The maximum number of days allowed to be missing from a
-#' station's data before it is excluded from final file output. Defaults to five
-#' days. If a single station is specified, this option is ignored and any data
-#' available, even an empty file, from NCDC will be returned.
-#' @param agroclimatology Logical. Only clean data for stations between
-#' latitudes 60 and -60 for agroclimatology work, defaults to FALSE. Set to
-#' TRUE to include only stations within the confines of these
-#' latitudes.
-#' @param CSV Logical. If set to TRUE, create a comma separated value (CSV)
-#' file ile and save it locally in a user specified location. Depends on
+#' @param country Optional. Specify a country of interest for which to retrieve
+#' weather data; full name. For stations located in locales having an ISO code 2
+#' or 3 letter ISO code can also be used if known. See
+#' \code{\link{country_list}} for a full list of country names and ISO codes
+#' available.
+#' @param dsn Optional. Path to file write to if saving a local file is desired.
+#' @param filename Optional. The filename for resulting file(s) to be written
+#' with no file extension. Year and file extension will be automatically
+#' appended to file outputs. Defaults to "GSOD-year".
+#' @param max_missing Optional. The maximum number of days allowed to be missing
+#' from a station's data before it is excluded from final file output.
+#' @param agroclimatology Optional. Logical. Only clean data for stations
[…]
