### 2016-11-08 9c5871a
Update documentation for get_GSOD
