### 2016-11-08 c39996d
Update get_GSOD documentation
