+++ R/get_GSOD.R
@@ -5,18 +5,18 @@
 #'Climatic Data Center (NCDC),
 #'\url{https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod},
 #'and calculates three new variables; Saturation Vapor Pressure (ES) – Actual
-#'Vapor Pressure (EA) and relative humidity (RH). Stations are individually
-#'checked for number of missing days to assure data quality, stations with too
-#'many missing observations are omitted. Additionally, stations reporting a
-#'latitude of < -90 or > 90 or longitude of < -180 or > 180 are removed. All
-#'units are converted to International System of Units (SI), e.g., Fahrenheit to
-#'Celsius and inches to millimetres. Alternative elevation measurements are
-#'supplied for missing values or values found to be questionable based on the
-#'Consultative Group for International Agricultural Research's Consortium for
-#'Spatial Information group's (CGIAR-CSI) Shuttle Radar Topography Mission 90
-#'metre (SRTM 90m) digital elevation data based on NASA's original SRTM 90m
-#'data. Further information on these data and methods can be found on GSODR's
-#'GitHub repository here:
+#'Vapor Pressure (EA) and relative humidity (RH). Stations reporting a latitude
+#'of < -90 or > 90 or longitude of < -180 or > 180 are removed. Stations may be
+#'individually checked for number of missing days to assure data quality and
+#'omitting stations with too many missing observations. All units are converted
+#'to International System of Units (SI), e.g., Fahrenheit to Celsius and inches
+#'to millimetres. Alternative elevation measurements are supplied for missing
+#'values or values found to be questionable based on the Consultative Group for
+#'International Agricultural Research's Consortium for Spatial Information
+#'group's (CGIAR-CSI) Shuttle Radar Topography Mission 90 metre (SRTM 90m)
+#'digital elevation data based on NASA's original SRTM 90m data. Further
+#'information on these data and methods can be found on GSODR's GitHub
+#'repository here:
 #'\url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
 #'
 #' @param years Year(s) of weather data to download.
@@ -29,9 +29,9 @@
 #' an existing file on the server, this function will silently fail and move on
 #' to existing files for download and cleaning from the FTP server.
 #' @param country Optional. Specify a country of interest for which to retrieve
-#' weather data; full name. For stations located in locales having an ISO code,
+#' weather data; full name. For stations located in locales having an ISO code
 #' the 2 or 3 letter ISO code can also be used if known.
-#' @param dsn Optional. Path to file write to.
+#' @param dsn Optional. Local file path to file write to.
 #' @param filename Optional. The filename for resulting file(s) to be written
 #' with no file extension. Year and file extension will be automatically
 #' appended to file outputs. Defaults to "GSOD-year".
@@ -41,17 +41,18 @@
 #' latitudes 60 and -60 for agroclimatology work, defaults to FALSE. Set to
 #' TRUE to include only stations within the confines of these
 #' latitudes.
-#' @param CSV Optional. Logical. If set to TRUE, create a comma separated value (CSV)
-#' file ile and save it locally in a user specified location. Depends on
+#' @param CSV Optional. Logical. If set to TRUE, create a comma separated value
+#' (CSV) file and save it locally in a user specified location. Depends on
 #' \code{dsn} being specified.
-#' @param GPKG Optional. Logical. If set to TRUE, create a GeoPackage file and save it
[…]
