### 2016-11-09 651d8b4
Fix bugs and update documentation for get_GSOD function
