+++ R/get_GSOD.R
@@ -5,9 +5,9 @@
 #'Climatic Data Center (NCDC),
 #'\url{https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod},
 #'and calculates three new variables; Saturation Vapor Pressure (ES) – Actual
-#'Vapor Pressure (EA) and relative humidity (RH). Stations are individually
-#'checked for number of missing days to assure data quality, stations with too
-#'many missing observations are omitted. Additionally, stations reporting a
+#'Vapor Pressure (EA) and relative humidity (RH). Stations may be individually
+#'checked for number of missing days to assure data quality, with stations
+#'having too many missing observations being omitted. Any stations reporting a
 #'latitude of < -90 or > 90 or longitude of < -180 or > 180 are removed. All
 #'units are converted to International System of Units (SI), e.g., Fahrenheit to
 #'Celsius and inches to millimetres. Alternative elevation measurements are
@@ -20,46 +20,43 @@
 #'\url{https://github.com/adamhsparks/GSODR/blob/master/data-raw/fetch_isd-history.md}
 #'
 #' @param years Year(s) of weather data to download.
-#' @param station Specify a station or multiple stations for which to retrieve,
-#' check and clean weather data. The NCDC reports years for which the data are
-#' available. This function checks against these years. However, not all cases
-#' are properly documented and in some cases files may not exist on the ftp
-#' server even though it is indicated that data was recorded for the station for
-#' a particular year. If a station is specified that does not have an
-#' existing file on the server, this function will silently fail and move on
+#' @param station Optional. Specify a station or multiple stations for which to
+#' retrieve, check and clean weather data. The NCDC reports years for which the
+#' data are available. This function checks against these years. However, not
+#' all cases are properly documented and in some cases files may not exist on
+#' the ftp server even though it is indicated that data was recorded for the
+#' station for a particular year. If a station is specified that does not have
+#' an existing file on the server, this function will silently fail and move on
 #' to existing files for download and cleaning from the ftp server.
-#' @param country Specify a country of interest for which to retrieve weather
-#' data; full name. For stations located in locales
-#' having an ISO code 2 or 3 letter ISO code can also be used if known. See
+#' @param country Optional. Specify a country for which to retrieve weather
+#' data; full name or ISO codes can be used. See
 #' \code{\link{country_list}} for a full list of country names and ISO
 #' codes available.
-#' @param dsn Path to file write to. (Optional)
-#' @param filename The filename for resulting file(s) to be written with no
-#' file extension. Year and file extension will be automatically appended to
-#' file outputs. Defaults to "GSOD-year". (Optional)
-#' @param max_missing The maximum number of days allowed to be missing from a
-#' station's data before it is excluded from final file output. Defaults to five
-#' days. If a single station is specified, this option is ignored and any data
-#' available, even an empty file, from NCDC will be returned.
-#' @param agroclimatology Logical. Only clean data for stations between
-#' latitudes 60 and -60 for agroclimatology work, defaults to FALSE. Set to
-#' TRUE to include only stations within the confines of these
-#' latitudes.
-#' @param CSV Logical. If set to TRUE, create a comma separated value (CSV)
-#' file ile and save it locally in a user specified location. Depends on
+#' @param CSV Optional. Logical. If set to TRUE, create a comma separated value
+#' (CSV) file ile and save it locally in a user specified location. Depends on
[…]
