+++ man/get_GSOD.Rd
@@ -5,52 +5,48 @@
 \title{Download, Clean, Reformat and Generate New Variables From GSOD Weather Data}
 \usage{
 get_GSOD(years = NULL, station = NULL, country = NULL, dsn = NULL,
-  filename = "GSOD", max_missing = 5, agroclimatology = FALSE,
-  CSV = FALSE, GPKG = FALSE, threads = 1)
+  filename = "GSOD", max_missing = NULL, agroclimatology = FALSE,
+  CSV = FALSE, GPKG = FALSE)
 }
 \arguments{
 \item{years}{Year(s) of weather data to download.}
 
-\item{station}{Specify a station or multiple stations for which to retrieve,
-check and clean weather data. The NCDC reports years for which the data are
-available. This function checks against these years. However, not all cases
-are properly documented and in some cases files may not exist on the ftp
-server even though it is indicated that data was recorded for the station for
-a particular year. If a station is specified that does not have an
-existing file on the server, this function will silently fail and move on
+\item{station}{Optional. Specify a station or multiple stations for which to
+retrieve, check and clean weather data. The NCDC reports years for which the
+data are available. This function checks against these years. However, not
+all cases are properly documented and in some cases files may not exist on
+the ftp server even though it is indicated that data was recorded for the
+station for a particular year. If a station is specified that does not have
+an existing file on the server, this function will silently fail and move on
 to existing files for download and cleaning from the ftp server.}
 
-\item{country}{Specify a country of interest for which to retrieve weather
-data; full name. For stations located in locales
-having an ISO code 2 or 3 letter ISO code can also be used if known. See
+\item{country}{Optional. Specify a country for which to retrieve weather
+data; full name or ISO codes can be used. See
 \code{\link{country_list}} for a full list of country names and ISO
 codes available.}
 
-\item{dsn}{Path to file write to. (Optional)}
+\item{dsn}{Optional. Local file path to write file out to. Must be specified
+if CSV or GPKG parameters are selected.}
 
-\item{filename}{The filename for resulting file(s) to be written with no
-file extension. Year and file extension will be automatically appended to
-file outputs. Defaults to "GSOD-year". (Optional)}
+\item{filename}{Optional. The filename for resulting file(s) to be written
+with no file extension. Year and file extension will be automatically
+appended to file outputs. Defaults to "GSOD-year". (Optional)}
 
-\item{max_missing}{The maximum number of days allowed to be missing from a
-station's data before it is excluded from final file output. Defaults to five
-days. If a single station is specified, this option is ignored and any data
-available, even an empty file, from NCDC will be returned.}
+\item{max_missing}{Optional. The maximum number of days allowed to be missing
+from a station's data before it is excluded from final file output.}
 
-\item{agroclimatology}{Logical. Only clean data for stations between
-latitudes 60 and -60 for agroclimatology work, defaults to FALSE. Set to
-TRUE to include only stations within the confines of these
+\item{agroclimatology}{Optional. Logical. Only clean data for stations
+between latitudes 60 and -60 for agroclimatology work, defaults to FALSE.
+Set to TRUE to include only stations within the confines of these
 latitudes.}
 
-\item{CSV}{Logical. If set to TRUE, create a comma separated value (CSV)
-file ile and save it locally in a user specified location. Depends on
+\item{CSV}{Optional. Logical. If set to TRUE, create a comma separated value
+(CSV) file ile and save it locally in a user specified location. Depends on
 \code{dsn} being specified.}
 
-\item{GPKG}{Logical. If set to TRUE, create a GeoPackage file and save it
[…]
