### 2016-11-11 #79 @adamhsparks
@deniederhut, I did some checking on the parallelisation just today. It seems to be much faster in my testing. This is for two years, 2010:2011 and all stations in the data.

Using plyr
```r
  system.time(GSOD_XY <- plyr::ldply(.data = GSOD_list, .fun = .process_gz,
                         stations = stations))

  #user   system  elapsed
  #1055.835   71.492 2921.688
```

In parallel using foreach dopar loop:

```r
  ity <- iterators::iter(GSOD_list)
  system.time(GSODY_XY2 <- foreach::foreach(i = ity) %dopar% {
    .process_gz(i, stations = stations)
  }
  )
   user  system elapsed
 24.395   9.373 380.866
```

This is strictly the data formatting, cleaning, etc. not the whole process, so I'll do more checking through the whole package, but for now I think it's useful.
