### 2016-11-11 260c755
Implement parallel processing for local file manipulation
