### 2016-11-11 a265736
Cleanup validation_functions.R a bit
