### 2016-11-11 e195d93
Return GSOD_XY as a data.frame in .dl_global_files
