### 2016-11-11 71dff67
Fix bug in .process_gz and call parallel package
