### 2016-11-13 134f94f
Major revisions to package structure
Drop parallel. Use plyr functions.

max_missing defaults to NULL, only used if user-specified.

Spelling corrections in documentation

Use less internal functions in get_GSOD

Reduced package dependencies
