+++ R/process_gz.R
@@ -1,0 +1,131 @@
+# Main function used by GSODR to process data files ----- ----------------------
+
+#' @noRd
+#' @importFrom data.table :=
+.process_gz <- function(gz_file, stations) {
+  
+  GSOD_XY <- data.table::data.table()
+  
+  YEARMODA <- "YEARMODA"
+  MONTH <- "MONTH"
+  DAY <- "DAY"
+  YDAY <- "YDAY"
+  DEWP <- "DEWP"
+  EA <- "EA"
+  ES <- "ES"
+  GUST <- "GUST"
+  MAX <- "MAX"
+  MIN <- "MIN"
+  MODA <- "MODA"
+  MXSPD <- "MXSPD"
+  PRCP <- "PRCP"
+  RH <- "RH"
+  SNDP <- "SNDP"
+  STN <- "STN"
+  STNID <- "STNID"
+  TEMP <- "TEMP"
+  VISIB <- "VISIB"
+  WBAN <- "WBAN"
+  WDSP <- "WDSP"
+  
+  tmp <- data.table::setDT(
+    readr::read_fwf(
+      file = gz_file, skip = 1,
+      readr::fwf_positions(c(1, 8, 15, 19, 25, 32, 36, 43, 47, 54,
+                             58, 65, 69, 75, 79, 85, 89, 96, 103,
+                             109, 111, 117, 119, 124, 126, 133,
+                             134, 135, 136, 137, 138),
+                           c(6, 12, 18, 22, 30, 33, 41, 44, 52,
+                             55, 63, 66, 73, 76, 83, 86, 93, 100,
+                             108, 109, 116, 117, 123, 124, 130,
+                             133, 134, 135, 136, 137, 138),
+                           col_names = c("STN", "WBAN", "YEAR",
+                                         "MODA", "TEMP",
+                                         "TEMP_CNT", "DEWP",
+                                         "DEWP_CNT", "SLP",
+                                         "SLP_CNT", "STP",
+                                         "STP_CNT", "VISIB",
+                                         "VISIB_CNT", "WDSP",
+                                         "WDSP_CNT", "MXSPD",
+                                         "GUST", "MAX",
+                                         "MAX_FLAG", "MIN",
+                                         "MIN_FLAG",
+                                         "PRCP", "PRCP_FLAG",
+                                         "SNDP", "I_FOG",
+                                         "I_RAIN_DRIZZLE",
+                                         "I_SNOW_ICE", "I_HAIL",
+                                         "I_THUNDER",
+                                         "I_TORNADO_FUNNEL")),
+      col_types = c("ccccdididididididddcdcdcdiiiiii"),
+      na = c("9999.9", "999.9", "99.99")))
+  
+  # add names to columns in data table
+  data.table::setnames(tmp, c("STN", "WBAN", "YEAR", "MODA", "TEMP",
+                              "TEMP_CNT", "DEWP", "DEWP_CNT", "SLP",
+                              "SLP_CNT", "STP", "STP_CNT", "VISIB",
+                              "VISIB_CNT", "WDSP", "WDSP_CNT", "MXSPD",
+                              "GUST", "MAX", "MAX_FLAG", "MIN", "MIN_FLAG",
+                              "PRCP", "PRCP_FLAG", "SNDP", "I_FOG",
+                              "I_RAIN_DRIZZLE", "I_SNOW_ICE", "I_HAIL",
+                              "I_THUNDER", "I_TORNADO_FUNNEL"))
+  
[…]
