### 2016-11-13 8ab73e9
Fix bugs with error check for file save options
