+++ R/get_GSOD.R
@@ -217,7 +217,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                      dsn = NULL, filename = NULL, max_missing = NULL,
                      agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE) {
   
-  # set up options, create objects, fetch most recent station metadata ---------
+  # Set up options, create objects, fetch most recent station metadata ---------
   original_options <- options()
   options(warn = 2)
   options(timeout = 300)
@@ -226,7 +226,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   LON <- LAT <- NULL
   ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
   
-  # validate years -------------------------------------------------------------
+  # Validate years -------------------------------------------------------------
   this_year <- 1900 + as.POSIXlt(Sys.Date())$year
   if (is.null(years) | is.character(years)) {
     stop("\nYou must provide at least one year of data to download in a numeric
@@ -244,7 +244,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     }
   }
   
-  # if file outs are specified, check that everything is in place --------------
+  # If file outs are specified, check that everything is in place --------------
   if (isTRUE(CSV) | isTRUE(GPKG)) {
     if (is.null(dsn)) {
       stop("\nYou must supply a valid file path (dsn) for storing the resulting
@@ -252,37 +252,39 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     } else {
         dsn <- trimws(dsn)
     }
-    
-    if (is.null(CSV) | is.null(GPKG)) {
-      stop("\nYou must supply a valid file format (CSV or GPKG) for storing the
-           resulting file(s).\n")
-    } else {
-      if (substr(dsn, nchar(dsn) - 1, nchar(dsn)) == "//") {
-        p <- substr(dsn, 1, nchar(dsn) - 2)
-      } else if (substr(dsn, nchar(dsn), nchar(dsn)) == "/" |
-                 substr(dsn, nchar(dsn), nchar(dsn)) == "\\") {
-        p <- substr(dsn, 1, nchar(dsn) - 1)
-      } else {
-        p <- dsn
-      }
-      if (!file.exists(p) & !file.exists(dsn)) {
-        stop("\nFile path does not exist: ", dsn, ".\n")
-      }
-    }
-    if (substr(dsn, nchar(dsn), nchar(dsn)) != "/" &
-        substr(dsn, nchar(dsn), nchar(dsn)) != "\\") {
-      dsn <- paste0(dsn, "/")
-    }
-    outfile <- paste0(dsn, filename)
   }
   
+  if (!isTRUE(CSV) | !isTRUE(GPKG)) {
+    stop("\nYou must supply a valid file format (CSV or GPKG) for storing the
+           resulting file(s).\n")
+  } else {
+    if (substr(dsn, nchar(dsn) - 1, nchar(dsn)) == "//") {
+      p <- substr(dsn, 1, nchar(dsn) - 2)
+    } else if (substr(dsn, nchar(dsn), nchar(dsn)) == "/" |
+               substr(dsn, nchar(dsn), nchar(dsn)) == "\\") {
+      p <- substr(dsn, 1, nchar(dsn) - 1)
+    } else {
+      p <- dsn
+    }
+    if (!file.exists(p) & !file.exists(dsn)) {
+      stop("\nFile path does not exist: ", dsn, ".\n")
+    }
+  }
+  if (substr(dsn, nchar(dsn), nchar(dsn)) != "/" &
+      substr(dsn, nchar(dsn), nchar(dsn)) != "\\") {
[…]
