### 2016-11-13 dc2423a
Rename function .fetch_stations to .fetch_station_list
Rename for clarity
