### 2016-11-14 2d0f095
Update SRTM_GSOD_elevation and country list
