+++ data-raw/fetch_isd-history.md
@@ -1,7 +1,7 @@
 Fetch, clean and correct altitude in GSOD isd\_history.csv Data
 ================
 Adam H. Sparks
-2016-10-24
+2016-11-14
 
 Introduction
 ============
@@ -122,24 +122,26 @@ doParallel::registerDoParallel(cl)
 
 corrected_elev <- tibble::as_tibble(
   data.table::rbindlist(foreach(i = dem_tiles) %dopar% {
-  
- # Load the DEM tile
-  dem <- raster::raster(i)
-  sub_stations <- raster::crop(stations, dem)
-
-  # in some cases the DEM represents areas where there is no station
-  # check for that here and if no stations, go on to next iteration
-  if (!is.null(sub_stations)) {
-
-    # use a 200m buffer to extract elevation from the DEM
-    sub_stations$ELEV_M_SRTM_90m <- raster::extract(dem, sub_stations, buffer = 200, fun = mean)
-
-    # convert spatial object back to normal data frame and add new fields
-    sub_stations <- as.data.frame(sub_stations)
-    return(sub_stations)
+    
+    # Load the DEM tile
+    dem <- raster::raster(i)
+    sub_stations <- raster::crop(stations, dem)
+    
+    # in some cases the DEM represents areas where there is no station
+    # check for that here and if no stations, go on to next iteration
+    if (!is.null(sub_stations)) {
+      
+      # use a 200m buffer to extract elevation from the DEM
+      sub_stations$ELEV_M_SRTM_90m <- raster::extract(dem, sub_stations,
+                                                      buffer = 200,
+                                                      fun = mean)
+      
+      # convert spatial object back to normal data frame and add new fields
+      sub_stations <- as.data.frame(sub_stations)
+      return(sub_stations)
+    }
   }
-}
-)
+  )
 )
 # stop cluster
 parallel::stopCluster(cl)
@@ -177,7 +179,7 @@ summary(SRTM_GSOD_elevation)
 ```
 
     ##      USAF               WBAN             STN_NAME        
-    ##  Length:27861       Length:27861       Length:27861      
+    ##  Length:28310       Length:28310       Length:28310      
     ##  Class :character   Class :character   Class :character  
     ##  Mode  :character   Mode  :character   Mode  :character  
     ##                                                          
@@ -185,27 +187,27 @@ summary(SRTM_GSOD_elevation)
     ##                                                          
     ##                                                          
     ##      CTRY              STATE               CALL                LAT        
-    ##  Length:27861       Length:27861       Length:27861       Min.   :-89.00  
-    ##  Class :character   Class :character   Class :character   1st Qu.: 24.08  
-    ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.67  
-    ##                                                           Mean   : 31.86  
-    ##                                                           3rd Qu.: 50.02  
+    ##  Length:28310       Length:28310       Length:28310       Min.   :-89.00  
+    ##  Class :character   Class :character   Class :character   1st Qu.: 22.47  
+    ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.25  
+    ##                                                           Mean   : 31.13  
+    ##                                                           3rd Qu.: 49.85  
[…]
