### 2016-11-14 2b58c21
Fix bug in checking whether CSV or GPKG is specified for filetype
