### 2016-11-14 4fdfa00
Add messages regarding file io when creating file outputs
