### 2016-11-14 5302a5f
Local file handling and validation functions
Create R file for validation functions in preparation for unit testing

For file handling. If CSV or GPKG is set to TRUE but no dsn is
specified, defaults to current working directory

If no filename is specified, defaults to "GSOD" followed by the file
extension.
