### 2016-11-14 023d36a
Consolidate functions into one file; update documentation
Move validation functions into get_GSOD.R

Move .process_gz into get_GSOD.R

Move fetch_stations.R to fetch_station_list.R to match function name.
This is used by two different exported functions so will remain as a
separate R file

Edit Description field of DESCRIPTION file to reflect the packages
current capabilities
