+++ R/get_GSOD.R
@@ -177,6 +177,8 @@
 #' \item{RH}{Mean daily relative humidity}
 #' }
 #'
+#' @author Adam Sparks \email{adamhsparks@gmail.com}
+#'
 #' @note Some of these data are redistributed with this R package. Originally
 #' from these data come from the US NCDC which states that users of these data
 #' should take into account the following: \dQuote{The following data and
@@ -225,133 +227,33 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   LON <- LAT <- NULL
   ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
 
+  # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   outfile <- .validate_fileout(CSV, dsn, filename, GPKG)
 
-  # Fetch latest station metadata from NCDC server -----------------------------
+  # Fetch latest station metadata from NCDC server
   if (!exists("stations")) {
     stations <- .fetch_station_list()
   }
+  .check_stations(station, stations, years)
+  .check_country(country)
 
-  # Check station integrity ----------------------------------------------------
-  if (!is.null(station)) {
-    if (!station %in% stations[[12]]) {
-      stop("\nThis is not a valid station ID number, please check your entry.
-         \nStation IDs are provided as a part of the GSODR package in the
-         'stations' data\nin the STNID column.\n")
-    }
-    # check station years in station listing
-    for (vsy in station) {
-      BEGIN <- as.numeric(substr(stations[stations[[12]] == vsy]$BEGIN, 1, 4))
-      END <- as.numeric(substr(stations[stations[[12]] == vsy]$END, 1, 4))
-      if (min(years) < BEGIN | max(years) > END)
-        message("This station, ", vsy, ", only provides data for years ", BEGIN,
-                " to ", END, ".\n")
-    }
-  }
+  GSOD_list <- .download_files(ftp, station, years, td)
 
-
-  # If country supplied, check and return letter code for filtering data ------
-  if (!is.null(country)) {
-    country <- toupper(trimws(country[1]))
-    nc <- nchar(country)
-    if (nc == 3) {
-      if (country %in% GSODR::country_list$iso3c) {
-        c <- which(country == GSODR::country_list$iso3c)
-        country <- GSODR::country_list[[c, 1]]
-      } else {
-        stop("\nPlease provide a valid name or 2 or 3 letter ISO country code;
-              you can view the entire list of valid countries in this data by
-              typing, 'country_list'.\n")
-      }
-    } else if (nc == 2) {
-      if (country %in% GSODR::country_list$iso2c) {
-        c <- which(country == GSODR::country_list$iso2c)
-        country <- GSODR::country_list[[c, 1]]
-      } else {
-        stop("\nPlease provide a valid name or 2 or 3 letter ISO country code;
-              you can view the entire list of valid countries in this data by
-              typing, 'country_list'.\n")
-      }
-    } else if (country %in% GSODR::country_list$COUNTRY_NAME) {
-      c <- which(country == GSODR::country_list$COUNTRY_NAME)
-      country <- GSODR::country_list[[c, 1]]
-    } else {
-      stop("\nPlease provide a valid name or 2 or 3 letter ISO country code;
-              you can view the entire list of valid countries in this data by
-              typing, 'country_list'.\n")
-    }
-  }
-
-
-  # Download complete tar files ------------------------------------------------
-  if (is.null(station)) {
[…]
