+++ man/get_GSOD.Rd
@@ -228,6 +228,9 @@ filename = "GSOD_agroclimatology_2010-2011", agroclimatology = TRUE)
 }
 
 }
+\author{
+Adam Sparks \email{adamhsparks@gmail.com}
+}
 \references{
 {Jarvis, A, HI Reuter, A Nelson, E Guevara, 2008, Hole-filled
 SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database

[…]
