### 2016-11-14 2650e8f
Use return in get_GSOD function
Fixes a bug where I'd deleted the return(GSOD_XY) in the get_GSOD
function, so no data was returned in the R session
