### 2016-11-14 7df1c2a
Correct issues with file out naming; update documentation
