### 2016-11-14 69a02e0
Fix issues with checking filenames and not specifying file format for saving
