+++ R/get_GSOD.R
@@ -242,13 +242,13 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
 
   # Check stations for missing days --------------------------------------------
   if (!is.null(max_missing)) {
-   GSOD_list <- .check_missing_days(max_missing, GSOD_list, td)
+    GSOD_list <- .check_missing_days(max_missing, GSOD_list, td)
   }
 
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
     GSOD_list <- .agroclimatology_list(agroclimatology, GSOD_list, stations, td,
-                                  years)
+                                       years)
   }
 
   # Subset GSOD_list for specified country -------------------------------------
@@ -260,11 +260,11 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
 
   message("Starting data file processing")
   GSOD_XY <- as.data.frame(
-        try(
-          plyr::ldply(.data = GSOD_list, .fun = .process_gz,
-                      stations = stations, .progress = "text")
-        )
+    try(
+      plyr::ldply(.data = GSOD_list, .fun = .process_gz,
+                  stations = stations, .progress = "text")
     )
+  )
 
 
   # Write files to disk --------------------------------------------------------
@@ -318,14 +318,17 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
         stop("\nThe year cannot be greater than current year.\n")
       }
     }
-    }
   }
+}
 
 
 #' @noRd
 .validate_fileout <- function(CSV, dsn, filename, GPKG) {
-  if (isTRUE(CSV) | isTRUE(GPKG)) {
+  if (!is.null(filename) & !isTRUE(CSV) & !isTRUE(GPKG)) {
+    stop("\nYou need to specify a filetype, CSV or GPKG.\n")
+  }
 
+  if (isTRUE(CSV) | isTRUE(GPKG)) {
     if (is.null(dsn)) {
       dsn <- getwd()
     } else {
@@ -346,11 +349,14 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
           substr(dsn, nchar(dsn), nchar(dsn)) != "\\") {
         dsn <- paste0(dsn)
       }
-      if (is.null(filename)) {
-        filename <- "GSOD"
-      }
     }
-    outfile <- paste0(dsn, "/", filename)
+
+    if (is.null(filename)) {
+      filename_out <- "GSOD"
+    } else {
+      filename_out <- filename
+    }
+    outfile <- paste0(dsn, "/", filename_out)
     return(outfile)
   }
 }
@@ -371,8 +377,8 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
         message("This station, ", vsy, ", only provides data for years ", BEGIN,
                 " to ", END, ".\n")
     }
-    }
   }
+}
 
 #' @noRd
 .check_country <- function(country) {
@@ -388,25 +394,25 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
              you can view the entire list of valid countries in this data by
              typing, 'country_list'.\n")
       }
-      } else if (nc == 2) {
-        if (country %in% GSODR::country_list$iso2c) {
-          c <- which(country == GSODR::country_list$iso2c)
-          country <- GSODR::country_list[[c, 1]]
-        } else {
-          stop("\nPlease provide a valid name or 2 or 3 letter ISO country code;
+    } else if (nc == 2) {
+      if (country %in% GSODR::country_list$iso2c) {
+        c <- which(country == GSODR::country_list$iso2c)
+        country <- GSODR::country_list[[c, 1]]
+      } else {
+        stop("\nPlease provide a valid name or 2 or 3 letter ISO country code;
                you can view the entire list of valid countries in this data by
[…]
