### 2016-11-14 c2e54ed
Updates to unit testing; correct documentation; revdep check
