+++ R/get_GSOD.R
@@ -27,7 +27,7 @@
 #' the ftp server even though it is indicated that data was recorded for the
 #' station for a particular year. If a station is specified that does not have
 #' an existing file on the server, this function will silently fail and move on
-#' to existing files for download and cleaning from the ftp server.
+#' to existing files for download and cleaning from the FTP server.
 #' @param country Optional. Specify a country for which to retrieve weather
 #' data; full name or ISO codes can be used. See
 #' \code{\link{country_list}} for a full list of country names and ISO
