+++ man/get_GSOD.Rd
@@ -18,7 +18,7 @@ all cases are properly documented and in some cases files may not exist on
 the ftp server even though it is indicated that data was recorded for the
 station for a particular year. If a station is specified that does not have
 an existing file on the server, this function will silently fail and move on
-to existing files for download and cleaning from the ftp server.}
+to existing files for download and cleaning from the FTP server.}
 
 \item{country}{Optional. Specify a country for which to retrieve weather
 data; full name or ISO codes can be used. See
