### 2016-11-14 ba23a6e
Fix error checking and unit tests for years validation
