### 2016-11-14 e257776
Fix bugs in validation functions; fix errors in unit tests; expand unit test capabilities
