### 2016-11-14 770ab15
Fix issues in testing
