### 2016-11-14 4231097
Add capabilities to check missing days in test_that; clean up .check_missing_days function
