### 2016-11-15 1384b3c
Rename ".check" functions to be ".validate" functions for consistency and clarity
