+++ R/get_GSOD.R
@@ -218,7 +218,7 @@
 get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                      dsn = NULL, filename = NULL, max_missing = NULL,
                      agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE) {
-  
+
   # Set up options, create objects, fetch most recent station metadata ---------
   original_options <- options()
   options(warn = 2)
@@ -226,39 +226,39 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   td <- tempdir()
   LON <- LAT <- NULL
   ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
-  
+
   # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   outfile <- .validate_fileout(CSV, dsn, filename, GPKG)
-  
+
   # Fetch latest station metadata from NCDC server
   if (!exists("stations")) {
     stations <- .fetch_station_list()
   }
-  .check_stations(station, stations, years)
-  .check_country(country)
-  
+  .validate_stations(station, stations, years)
+  .validate_country(country)
+
   # Download files from server -------------------------------------------------
   GSOD_list <- .download_files(ftp, station, years, td)
-  
-  # Check stations for missing days --------------------------------------------
+
+  # Validate stations for missing days --------------------------------------------
   if (!is.null(max_missing)) {
-    GSOD_list <- .check_missing_days(max_missing, GSOD_list, td)
+    GSOD_list <- .validate_missing_days(max_missing, GSOD_list)
   }
-  
+
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
     GSOD_list <- .agroclimatology_list(agroclimatology, GSOD_list, stations, td,
                                        years)
   }
-  
+
   # Subset GSOD_list for specified country -------------------------------------
   if (!is.null(country)) {
     GSOD_list <- .country_list(country, GSOD_list, stations, td, years)
   }
-  
+
   # Clean and reformat list of station files from local disk in tempdir --------
-  
+
   message("Starting data file processing")
   GSOD_XY <- as.data.frame(
     try(
@@ -266,16 +266,16 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                   stations = stations, .progress = "text")
     )
   )
-  
-  
+
+
   # Write files to disk --------------------------------------------------------
-  
+
   if (isTRUE(CSV)) {
     message("\nWriting CSV file to disk.\n")
     outfile <- paste0(outfile, ".csv")
     readr::write_csv(GSOD_XY, path = paste0(outfile))
   }
-  
+
   if (isTRUE(GPKG)) {
     message("\nWriting GeoPackage File to Disk.\n")
     outfile <- paste0(outfile, ".gpkg")
@@ -283,7 +283,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     GSOD_XY <- as.data.frame(GSOD_XY)
     sp::coordinates(GSOD_XY) <- ~LON + LAT
     sp::proj4string(GSOD_XY) <- sp::CRS("+proj=longlat +datum=WGS84")
-    
+
     # If the filename specified exists, remove it and create new
     if (file.exists(path.expand(outfile))) {
       file.remove(outfile)
@@ -292,9 +292,9 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     rgdal::writeOGR(GSOD_XY, dsn = path.expand(outfile), layer = "GSOD",
                     driver = "GPKG")
   }
-  
+
   return(GSOD_XY)
-  
+
[…]
