### 2016-11-15 6960104
Fix bug in .validate_missing_days; test .validate_missing_days in tests
