### 2016-11-15 1bb1b05
test for valid list returned, not just length in .validate_missing_days test
