### 2016-11-15 b93a198
Test .validate_country handles proper and improper entries
