### 2016-11-15 dc6294a
Comment test_get_GSOD.R
