### 2016-11-15 f6c3a67
Cleanup in some get_GSOD functions; more testing
Test subsetting for agro and country

Test that downloading works and uncompresses files as it should
