+++ tests/testthat/test-get_GSOD.R
@@ -4,22 +4,22 @@ test_that(".validate_years handles invalid years", {
   skip_on_cran()
 
   expect_error(.validate_years(years = NULL),
-    "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
+               "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
   expect_error( .validate_years(years = "nineteen ninety two"),
-    "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
+                "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
   expect_error(.validate_years(years = 1923),
-    "\nThe GSOD data files start at 1929, you have entered a year prior\n             to 1929.\n")
+               "\nThe GSOD data files start at 1929, you have entered a year prior\n             to 1929.\n")
   expect_error(.validate_years(years = 1901 + as.POSIXlt(Sys.Date())$year),
-    "\nThe year cannot be greater than current year.\n")
+               "\nThe year cannot be greater than current year.\n")
 
 })
 
 # Check that .validate_years handles valid years -------------------------------
 test_that(".validate_years handles valid years", {
   skip_on_cran()
-expect_error(.validate_years(years = 1929:2016), regexp = NA)
+  expect_error(.validate_years(years = 1929:2016), regexp = NA)
 
-expect_error(.validate_years(years = 2016), regexp = NA)
+  expect_error(.validate_years(years = 2016), regexp = NA)
 
 })
 
@@ -70,26 +70,27 @@ test_that("stations list and associated metatdata", {
 # Check missing days in non-leap years -----------------------------------------
 test_that("missing days check allows stations with permissible days missing,
           non-leap year", {
-  skip_on_cran()
-  max_missing <- 5
-  td <- tempdir()
-  just_right_2015 <- data.frame(c(rep(12, 360)), c(rep("X", 360)))
-  too_short_2015 <- data.frame(c(rep(12, 300)), c(rep("X", 300)))
-  df_list <- list(just_right_2015, too_short_2015)
+            skip_on_cran()
+            max_missing <- 5
+            td <- tempdir()
+            just_right_2015 <- data.frame(c(rep(12, 360)), c(rep("X", 360)))
+            too_short_2015 <- data.frame(c(rep(12, 300)), c(rep("X", 300)))
+            df_list <- list(just_right_2015, too_short_2015)
 
-  filenames <- c("just_right_2015", "too_short_2015")
-  sapply(1:length(df_list),
-         function(x) write.csv(df_list[[x]],
-                               file = gzfile(
-                                 paste0(td, "/", filenames[x], ".csv.gz"))
-                               )
-  )
-  GSOD_list <- as.list(list.files(td, pattern = "2015.csv.gz$"))
-  GSOD_list_filtered <- .validate_missing_days(max_missing, GSOD_list, td)
+            filenames <- c("just_right_2015", "too_short_2015")
+            sapply(1:length(df_list),
+                   function(x) write.csv(df_list[[x]],
+                                         file = gzfile(
+                                           paste0(td, "/", filenames[x], ".csv.gz"))
+                   )
+            )
+            GSOD_list <- as.list(list.files(td, pattern = "2015.csv.gz$"))
+            GSOD_list_filtered <- .validate_missing_days(max_missing, GSOD_list, td)
 
-  expect_length(GSOD_list, 2)
-  expect_match(GSOD_list_filtered, "just_right_2015.csv.gz")
-})
+            expect_length(GSOD_list, 2)
+            expect_match(GSOD_list_filtered, "just_right_2015.csv.gz")
+            unlink(td)
+          })
 
 # Check missing days in leap years ---------------------------------------------
[…]
