### 2016-11-15 cc0ce87
Fix bugs in test-get_GSOD.R
Fix leap year check bug
