### 2016-11-15 c53e45f
Fix bug in get_GSOD for agroclimatology; new test in testthat
Test for properly formatted data frame from .process_gz function
