+++ tests/testthat/test-get_GSOD.R
@@ -2,7 +2,7 @@ context("get_GSOD")
 # Check that .validate_years handles invalid years -----------------------------
 test_that(".validate_years handles invalid years", {
   skip_on_cran()
-
+  
   expect_error(.validate_years(years = NULL),
                "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
   expect_error( .validate_years(years = "nineteen ninety two"),
@@ -11,19 +11,19 @@ test_that(".validate_years handles invalid years", {
                "\nThe GSOD data files start at 1929, you have entered a year prior\n             to 1929.\n")
   expect_error(.validate_years(years = 1901 + as.POSIXlt(Sys.Date())$year),
                "\nThe year cannot be greater than current year.\n")
-
+  
 })
 
 # Check that .validate_years handles valid years -------------------------------
 test_that(".validate_years handles valid years", {
   skip_on_cran()
   expect_error(.validate_years(years = 1929:2016), regexp = NA)
-
+  
   expect_error(.validate_years(years = 2016), regexp = NA)
-
+  
 })
 
-# invalid stations are handled -------------------------------------------------
+# Check that invalid stations are handled --------------------------------------
 test_that("invalid stations are handled", {
   skip_on_cran()
   stations <- .fetch_station_list()
@@ -34,7 +34,7 @@ test_that("invalid stations are handled", {
 # Check that invalid dsn is handled --------------------------------------------
 test_that("invalid dsn is handled", {
   skip_on_cran()
-
+  
   expect_error(.validate_fileout(CSV = FALSE, dsn = "~/R", filename = NULL,
                                  GPKG = FALSE),
                "\nFile dsn does not exist: ~/R.\n")
@@ -46,10 +46,11 @@ test_that("invalid dsn is handled", {
 # Check stations list and associated metadata for validity ---------------------
 test_that("stations list and associated metatdata", {
   skip_on_cran()
-
+  
   stations <- .fetch_station_list()
-
-  expect_equal(ncol(stations), 13)
+  
+  expect_length(stations, 13)
+  
   expect_is(stations, "data.table")
   expect_is(stations$USAF, "character")
   expect_is(stations$WBAN , "character")
@@ -64,6 +65,7 @@ test_that("stations list and associated metatdata", {
   expect_is(stations$END, "numeric")
   expect_is(stations$STNID, "character")
   expect_is(stations$ELEV_M_SRTM_90m, "numeric")
+  
   expect_gt(nrow(stations), 2300)
 })
 
@@ -76,17 +78,19 @@ test_that("missing days check allows stations with permissible days missing,
             just_right_2015 <- data.frame(c(rep(12, 360)), c(rep("X", 360)))
             too_short_2015 <- data.frame(c(rep(12, 300)), c(rep("X", 300)))
             df_list <- list(just_right_2015, too_short_2015)
-
+            
             filenames <- c("just_right_2015", "too_short_2015")
             sapply(1:length(df_list),
                    function(x) write.csv(df_list[[x]],
                                          file = gzfile(
-                                           paste0(td, "/", filenames[x], ".csv.gz"))
+                                           paste0(td, "/", filenames[x],
+                                                  ".csv.gz"))
                    )
             )
             GSOD_list <- as.list(list.files(td, pattern = "2015.csv.gz$"))
-            GSOD_list_filtered <- .validate_missing_days(max_missing, GSOD_list, td)
-
+            GSOD_list_filtered <- .validate_missing_days(max_missing, GSOD_list,
+                                                         td)
+            
[…]
