### 2016-11-15 fddb2f7
Fix bug in get_GSOD
Fix missing parameter in the .validate_missing_days function
