### 2016-11-15 91b1285
Test nearest_stations function
