+++ tests/testthat/test-nearest_stations.R
@@ -1,0 +1,12 @@
+context("nearest_stations")
+# Check that .validate_years handles invalid years -----------------------------
+test_that("nearest stations returns numeric station IDs", {
+  skip_on_cran()
+  
+  n <- nearest_stations(LAT = -27.5598, LON = 151.9507, distance = 100)
+  expect_length(n, 15)
+  expect_type(n, "character")
+  expect_equal(n, c("749459-99999", "945420-99999", "945490-99999", "945510-99999", "945520-99999", "945550-99999", "945620-99999", "945951-99999", "949999-00170", "949999-00172", "949999-00179", "949999-00183", "949999-00186", "955510-99999", "955550-99999"))
+  
+})
+  
