### 2016-11-15 00bb089
Comments in test-nearest_stations.R
