### 2016-11-16 2a294c7
Minor edits in tests
