### 2016-11-16 ee7f846
Fix bug in tests
