+++ R/get_GSOD.R
@@ -224,7 +224,6 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   options(warn = 2)
   options(timeout = 300)
   td <- tempdir()
-  LON <- LAT <- NULL
   ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
 
   # Validate user inputs -------------------------------------------------------
@@ -241,7 +240,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   # Download files from server -------------------------------------------------
   GSOD_list <- .download_files(ftp, station, years, td)
 
-  # Validate stations for missing days --------------------------------------------
+  # Validate stations for missing days -----------------------------------------
   if (!is.null(max_missing)) {
     GSOD_list <- .validate_missing_days(max_missing, GSOD_list, td)
   }
@@ -576,15 +575,15 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                                                     tmp$DAY,
                                                     sep = "-")),
                                       format = "%j"))]
-  tmp[, (TEMP)  := round((5 / 9) * (tmp$TEMP - 32), 1)]
-  tmp[, (DEWP)  := round((5 / 9) * (tmp$DEWP - 32), 1)]
+  tmp[, (TEMP)  := round( (5 / 9) * (tmp$TEMP - 32), 1)]
+  tmp[, (DEWP)  := round( (5 / 9) * (tmp$DEWP - 32), 1)]
   tmp[, (WDSP)  := round(tmp$WDSP * 0.514444444, 1)]
   tmp[, (MXSPD) := round(tmp$MXSPD * 0.514444444, 1)]
   tmp[, (VISIB) := round(tmp$VISIB * 1.60934, 1)]
   tmp[, (WDSP)  := round(tmp$WDSP * 0.514444444, 1)]
   tmp[, (GUST)  := round(tmp$GUST * 0.514444444, 1)]
-  tmp[, (MAX)   := round((tmp$MAX - 32) * (5 / 9), 2)]
-  tmp[, (MIN)   := round((tmp$MIN - 32) * (5 / 9), 2)]
+  tmp[, (MAX)   := round( (tmp$MAX - 32) * (5 / 9), 2)]
+  tmp[, (MIN)   := round( (tmp$MIN - 32) * (5 / 9), 2)]
   tmp[, (PRCP)  := round(tmp$PRCP * 25.4, 1)]
   tmp[, (SNDP)  := round(tmp$SNDP * 25.4, 1)]
   # Compute other weather vars--------------------------------------------------
@@ -592,11 +591,11 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   # Monteith JL (1973) Principles of environmental physics.
   #   Edward Arnold, London
   # EA derived from dew point
-  tmp[, (EA) := round(0.61078 * exp((17.2694 * (tmp$DEWP)) /
-                                      ((tmp$DEWP) + 237.3)), 1)]
+  tmp[, (EA) := round(0.61078 * exp( (17.2694 * (tmp$DEWP)) /
+                                      ( (tmp$DEWP) + 237.3)), 1)]
   # ES derived from average temperature
-  tmp[, (ES) := round(0.61078 * exp((17.2694 * (tmp$TEMP)) /
-                                      ((tmp$TEMP) + 237.3)), 1)]
+  tmp[, (ES) := round(0.61078 * exp( (17.2694 * (tmp$TEMP)) /
+                                      ( (tmp$TEMP) + 237.3)), 1)]
   # Calculate relative humidity
   tmp[, (RH) := round(tmp$EA / tmp$ES * 100, 1)]
   # Join to the station and SRTM data-------------------------------------------
