+++ tests/testthat/test-get_GSOD.R
@@ -5,7 +5,7 @@ test_that(".validate_years handles invalid years", {
 
   expect_error(.validate_years(years = NULL),
                "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
-  expect_error( .validate_years(years = "nineteen ninety two"),
+  expect_error(.validate_years(years = "nineteen ninety two"),
                 "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
   expect_error(.validate_years(years = 1923),
                "\nThe GSOD data files start at 1929, you have entered a year prior\n             to 1929.\n")
@@ -35,9 +35,9 @@ test_that("invalid stations are handled", {
 test_that("invalid dsn is handled", {
   skip_on_cran()
 
-  expect_error(.validate_fileout(CSV = FALSE, dsn = "~/R", filename = NULL,
+  expect_error(.validate_fileout(CSV = FALSE, dsn = "~/NULL", filename = NULL,
                                  GPKG = FALSE),
-               "\nFile dsn does not exist: ~/R.\n")
+               "\nFile dsn does not exist: ~/NULL.\n")
   expect_error(.validate_fileout(CSV = FALSE, dsn = NULL, filename = "test",
                                  GPKG = FALSE),
                "\nYou need to specify a filetype, CSV or GPKG.")
@@ -53,7 +53,7 @@ test_that("stations list and associated metatdata", {
 
   expect_is(stations, "data.table")
   expect_is(stations$USAF, "character")
-  expect_is(stations$WBAN , "character")
+  expect_is(stations$WBAN, "character")
   expect_is(stations$STN_NAME, "character")
   expect_is(stations$CTRY, "character")
   expect_is(stations$STATE, "character")
@@ -188,7 +188,7 @@ test_that(".download_files properly works, subsetting for country and
             expect_is(gz_out, "data.table")
 
             expect_is(gz_out$USAF, "character")
-            expect_is(gz_out$WBAN , "character")
+            expect_is(gz_out$WBAN, "character")
             expect_is(gz_out$STNID, "character")
             expect_is(gz_out$STN_NAME, "character")
             expect_is(gz_out$CTRY, "character")
