### 2016-11-16 dcaa760
Update documentation and NEWS.md; new version number 1.0
