+++ R/GSODR.R
@@ -17,6 +17,9 @@
 #'name, country, latitude, longitude, elevation) and weather observations and
 #'associated flags. Data may be automatically saved to disk. File output may
 #'be returned as a comma-separated values (CSV) or GeoPackage (GPKG) file.
+#'For information on the original data from NCDC, please see the GSOD
+#'readme.txt file available from,
+#'<http://www1.ncdc.noaa.gov/pub/data/gsod/readme.txt>.
 #'
 #' @section GSODR functions:
 #' \code{\link{get_GSOD}}
