### 2016-11-16 ada810c
Update NEWS.md
