### 2016-11-16 a6cfadd
Spelling corrections
