+++ R/nearest_stations.R
@@ -1,6 +1,6 @@
 #' Select Nearest Stations to Specified Latitude and Longitude
 #'
-#'Given a latitude and longitiude value entered as decimal degrees (DD),
+#'Given a latitude and longitude value entered as decimal degrees (DD),
 #'this function returns a list of STNID values, which can be used in the
 #'\code{get_GSOD} function to query for specific stations as an argument in the
 #'\code{station} parameter of that function.
