### 2016-11-17 d4bae25
Add new vignette for IRRI loop survey stations; Check server file list to verify before downloading stations
