### 2016-11-17 a06d83f
Fix bug validating DSN and fileout
