+++ vignettes/Fetch_specified_stations_for_a_range_of_years.Rmd
@@ -79,6 +79,8 @@ loop_stations <- station_meta[station_meta$STNID %in% loop_stations, ]
 loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
                                  loop_stations$END >= 20151231, ]
 
+stations_list <- c(loop_stations[, 12])
+
 ```
 
 ## Fetch GSOD Weather Data for These Stations and Save to Disk
@@ -88,7 +90,7 @@ of interest from 1960 until 2016. Then write the resulting data to a file called
 (Mac/Linux) or "My Documents" directory (Windows).
 
 ```{r, fetch_weather, message=FALSE, eval=FALSE}
-get_GSOD(station = loop_stations[, 12], years = 1960:2016, CSV = TRUE,
+get_GSOD(station = stations_list, years = 1960:2016, CSV = TRUE,
          dsn = "~/", filename = "Loop_Survey_Weather_1960-2016")
 ```
 
