### 2016-11-17 e0174b9
Vignette, fix bug in survey loop station list in stations parameter for get_GSOD
