### 2016-11-17 926af55
Bug fixes in get_GSOD for checking files exist on server
