### 2016-11-18 1c15d93
Tracking down bugs related to individual station file downloads
When multiple stations and years are selected the download fails.

Using the `curl` library's `curl_fetch_disk` function to try to
alleviate these issues.

Also some minor cleanup in the IRRI loop station vignette

Add `curl` to dependencies
