+++ R/get_GSOD.R
@@ -218,14 +218,14 @@
 get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                      dsn = NULL, filename = NULL, max_missing = NULL,
                      agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE) {
-  
+
   # Set up options, create objects, fetch most recent station metadata ---------
   original_options <- options()
   options(warn = 2)
   options(timeout = 300)
   td <- tempdir()
   ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
-  
+
   # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   if (!is.null(dsn)) {
@@ -235,34 +235,33 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   if (!exists("stations")) {
     stations <- .fetch_station_list()
   }
+  # Validate user entered stations for existence in stations list from NCDC
   plyr::l_ply(.data = station, .fun = .validate_station, stations = stations,
               years = years)
-  
+
   .validate_country(country)
-  
+
   # Download files from server -------------------------------------------------
-  station <- plyr::llply(.data = station, .fun = .validate_server_files,
-                         ftp = ftp, years = years)
-  
+
   GSOD_list <- .download_files(ftp, station, years, td)
-  
+
   # Validate stations for missing days -----------------------------------------
   if (!is.null(max_missing)) {
     GSOD_list <- .validate_missing_days(max_missing, GSOD_list, td)
   }
-  
+
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
     GSOD_list <- .agroclimatology_list(GSOD_list, stations, td, years)
   }
-  
+
   # Subset GSOD_list for specified country -------------------------------------
   if (!is.null(country)) {
     GSOD_list <- .country_list(country, GSOD_list, stations, td, years)
   }
-  
+
   # Clean and reformat list of station files from local disk in tempdir --------
-  
+
   message("Starting data file processing")
   GSOD_XY <- as.data.frame(
     try(
@@ -270,16 +269,16 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                   stations = stations, .progress = "text")
     )
   )
-  
-  
+
+
   # Write files to disk --------------------------------------------------------
-  
+
   if (isTRUE(CSV)) {
     message("\nWriting CSV file to disk.\n")
     outfile <- paste0(outfile, ".csv")
     readr::write_csv(GSOD_XY, path = paste0(outfile))
   }
-  
+
   if (isTRUE(GPKG)) {
     message("\nWriting GeoPackage File to Disk.\n")
     outfile <- paste0(outfile, ".gpkg")
@@ -287,7 +286,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     GSOD_XY <- as.data.frame(GSOD_XY)
     sp::coordinates(GSOD_XY) <- ~LON + LAT
     sp::proj4string(GSOD_XY) <- sp::CRS("+proj=longlat +datum=WGS84")
-    
+
     # If the filename specified exists, remove it and create new
     if (file.exists(path.expand(outfile))) {
       file.remove(outfile)
@@ -296,9 +295,9 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     rgdal::writeOGR(GSOD_XY, dsn = path.expand(outfile), layer = "GSOD",
                     driver = "GPKG")
   }
-  
+
   return(GSOD_XY)
-  
+
   # Cleanup and reset to default state -----------------------------------------
   unlink(td)
[…]
