+++ vignettes/Fetch_specified_stations_for_a_range_of_years.Rmd
@@ -59,9 +59,9 @@ centroid <- gCentroid(loop_area)
 
 ```{r, nearest_stations, message=FALSE, eval=FALSE}
 library(GSODR)
-
+library(readr)
 # Fetch station list from NCDC
-station_meta <- readr::read_csv(
+station_meta <- read_csv(
   "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
   col_types = "ccccccddddd",
   col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL", "LAT",
