### 2016-11-18 7884d75
back to curl_download function
