### 2016-11-20 a3042c2
Work on downloading station files from server rather than whole year files
