+++ R/get_GSOD.R
@@ -469,11 +469,11 @@ your entry. Station IDs are provided as a part of the GSODR package in the
       file_list <- paste0(ftp, i, "/", file_list)
 
       tryCatch(Map(function(ftp, dest)
-        curl::curl_fetch_disk(url = ftp,
-                              path = paste0(dest, "/", basename(ftp))),
+       utils::download.file(url = ftp, destfile = dest, mode = "wb"),
         file_list, file.path(td, basename(file_list))),
         error = function(x) stop(
           "\nThe file downloads have failed. Please restart.\n"))
+      Sys.sleep(15) # slow requests to FTP server, otherwise error
     }
   }
 
