### 2016-11-21 1f5c0f3
Still getting errors about not being logged in once downloading station files
