### 2016-11-21 ff08e55
Update loop station vignette; fix DESCRIPTION file; Remove Sys.sleep in get_GSOD
