+++ DESCRIPTION
@@ -56,7 +56,6 @@ Suggests:
     lubridate,
     plotKML,
     raster,
-    rgdal,
     rgeos,
     rmarkdown,
     spacetime,
