### 2016-11-22 18c8438
Add new reformat_GSOD function
Add a new stand-alone function to reformat station files that are
downloaded via FTP to local disk.

Put .process_gz function in it's own process_gz.R file since it is now
called by two different functions that are exported by the package.

Update documentation
