+++ R/reformat_GSOD.R
@@ -1,0 +1,213 @@
+#' Clean, Reformat and Generate New Variables From GSOD Weather Data
+#'
+#'This function automates cleaning and reformatting of GSOD,
+#'\url{https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod},
+#'station files in "WMO-WBAN-YYYY.op.gz" format that have been downloaded from the
+#'United States National Climatic Data Center's (NCDC) FTP server.
+#'
+#'For automated downloading and processing see the \code{\link{get_GSOD}}
+#'function which provides expanded functionality for automatically downloading
+#'and expanding annual GSOD files and cleaning station files.
+#'
+#'This function reformats the data into a more usable form and caculates three
+#'new variables; Saturation Vapor Pressure (ES) – Actual Vapor Pressure (EA) and
+#'relative humidity (RH). All units are converted to International System of
+#'Units (SI), e.g., Fahrenheit to Celsius and inches to millimetres. Alternative
+#'elevation measurements are supplied for missing values or values found to be
+#'questionable based on the Consultative Group for International Agricultural
+#'Research's Consortium for Spatial Information group's (CGIAR-CSI) Shuttle
+#'Radar Topography Mission 90 metre (SRTM 90m) digital elevation data based on
+#'NASA's original SRTM 90m data.
+#'
+#'@param file_path User supplied file path to location of station file data on
+#'local disk for reformatting.
+#'
+#' @details
+#' Data summarise each year by station, which include vapour pressure and
+#' relative humidity variables calculated from existing data in GSOD.
+#'
+#' All missing values in resulting files are represented as NA regardless of
+#' which field they occur in.
+#'
+#' Only station files in .op.gz file format are supported by this function. If
+#' you have downloaded the full annual "gsod_YYYY.tar" file you will need to
+#' extract the individual station files first to use this function.
+#'
+#' The data returned either in a data.frame object that includes the following
+#' fields:
+#' \describe{
+#' \item{STNID}{Station number (WMO/DATSAV3 number) for the location}
+#' \item{WBAN}{Number where applicable--this is the historical "Weather Bureau
+#' Air Force Navy" number - with WBAN being the acronym}
+#' \item{STN_NAME}{Unique text identifier}
+#' \item{CTRY}{Country in which the station is located}
+#' \item{LAT}{Latitude. *Station dropped in cases where values are < -90 or
+#'> 90 degrees or Lat = 0 and Lon = 0* (WGS84)}
+#' \item{LON}{Longitude. *Station dropped in cases where values are < -180 or
+#'> 180 degrees or Lat = 0 and Lon = 0* (WGS84)}
+#' \item{ELEV_M}{Elevation in metres}
+#' \item{ELEV_M_SRTM_90m}{Elevation in metres corrected for possible errors,
+#' derived from the CGIAR-CSI SRTM 90m database (Jarvis et al. 2008)}
+#' \item{YEARMODA}{Date in YYYY-mm-dd format}
+#' \item{YEAR}{The year (YYYY)}
+#' \item{MONTH}{The month (mm)}
+#' \item{DAY}{The day (dd)}
+#' \item{YDAY}{Sequential day of year (not in original GSOD)}
+#' \item{TEMP}{Mean daily temperature converted to degrees C to tenths.
+#' Missing = NA}
+#' \item{TEMP_CNT}{Number of observations used in calculating mean daily
+#' temperature}
+#' \item{DEWP}{Mean daily dew point converted to degrees C to tenths. Missing
+#' = NA}
+#' \item{DEWP_CNT}{Number of observations used in calculating mean daily dew
+#' point}
+#' \item{SLP}{Mean sea level pressure in millibars to tenths. Missing = NA}
[…]
