+++ NAMESPACE
@@ -2,4 +2,5 @@
 
 export(get_GSOD)
 export(nearest_stations)
+export(reformat_GSOD)
 importFrom(data.table,":=")

[…]
