### 2016-11-22 7911873
Mainly fix FTP issues thanks to Bob Rudis, other small changes
Updates in Fetch_specified_stations_for_a_range_of_years;

Add option in reformat_GSOD.R to clean a list of files;
