+++ R/get_GSOD.R
@@ -214,19 +214,16 @@
 #' SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database
 #' \url{http://srtm.csi.cgiar.org}}
 #'
+#' @importFrom dplyr %>%
 #' @export
 get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                      dsn = NULL, filename = NULL, max_missing = NULL,
                      agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE) {
-
-  # Set up options, create objects, fetch most recent station metadata ---------
-  original_options <- options()
-  options(warn = 2)
-  options(timeout = 300)
-  td <- tempdir()
-  ftp <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/"
-
-
+  
+  # Create objects for use in retrieving files ---------------------------------
+  cache_dir <- tempdir()
+  ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/%s/"
+  
   # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   if (!is.null(dsn)) {
@@ -239,30 +236,34 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
   # Validate user entered stations for existence in stations list from NCDC
   plyr::l_ply(.data = station, .fun = .validate_station, stations = stations,
               years = years)
-
+  
   .validate_country(country)
-
+  
   # Download files from server -------------------------------------------------
-
-  GSOD_list <- .download_files(ftp, station, years, td)
-
+  
+  GSOD_list <- .download_files(ftp_base, station, years, cache_dir)
+  
   # Validate stations for missing days -----------------------------------------
   if (!is.null(max_missing)) {
-    GSOD_list <- .validate_missing_days(max_missing, GSOD_list, td)
+    GSOD_list <- .validate_missing_days(max_missing, GSOD_list, cache_dir)
   }
-
+  
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
-    GSOD_list <- .agroclimatology_list(GSOD_list, stations, td, years)
+    GSOD_list <- .agroclimatology_list(GSOD_list, stations, cache_dir, years)
   }
-
+  
   # Subset GSOD_list for specified country -------------------------------------
   if (!is.null(country)) {
-    GSOD_list <- .country_list(country, GSOD_list, stations, td, years)
+    GSOD_list <- .country_list(country, GSOD_list, stations, cache_dir, years)
   }
-
+  
+  # Download files from FTP server ---------------------------------------------
+  
+  .download_station_files(cache_dir, ftp_base, station, years)
+  
   # Clean and reformat list of station files from local disk in tempdir --------
-
+  
   message("Starting data file processing")
   GSOD_XY <- as.data.frame(
     try(
@@ -270,16 +271,16 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
                   stations = stations, .progress = "text")
     )
   )
-
-
+  
+  
   # Write files to disk --------------------------------------------------------
-
+  
   if (isTRUE(CSV)) {
     message("\nWriting CSV file to disk.\n")
     outfile <- paste0(outfile, ".csv")
     readr::write_csv(GSOD_XY, path = paste0(outfile))
   }
-
+  
   if (isTRUE(GPKG)) {
     message("\nWriting GeoPackage File to Disk.\n")
     outfile <- paste0(outfile, ".gpkg")
@@ -287,7 +288,7 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     GSOD_XY <- as.data.frame(GSOD_XY)
[…]
