+++ NAMESPACE
@@ -4,3 +4,4 @@ export(get_GSOD)
 export(nearest_stations)
 export(reformat_GSOD)
 importFrom(data.table,":=")
+importFrom(dplyr,"%>%")

[…]
