### 2016-11-23 0af9c04
Documentation updates
