### 2016-11-23 2362ae0
Incorporate new method for dl stations; Update documentation for get_GSOD
