+++ R/get_GSOD.R
@@ -177,7 +177,6 @@
 #' \item{RH}{Mean daily relative humidity}
 #' }
 #'
-#' @author Adam Sparks \email{adamhsparks@gmail.com}
 #'
 #' @note Some of these data are redistributed with this R package. Originally
 #' from these data come from the US NCDC which states that users of these data
@@ -209,21 +208,28 @@
 #' filename = "GSOD_agroclimatology_2010-2011", agroclimatology = TRUE)
 #'
 #' }
-#'
+#' @author Adam H Sparks \email{adamhsparks@gmail.com}
 #' @references {Jarvis, A, HI Reuter, A Nelson, E Guevara, 2008, Hole-filled
 #' SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database
 #' \url{http://srtm.csi.cgiar.org}}
 #'
+#' @seealso \code{\link{reformat_GSOD}}
+#'
 #' @importFrom dplyr %>%
 #' @export
-get_GSOD <- function(years = NULL, station = NULL, country = NULL,
-                     dsn = NULL, filename = NULL, max_missing = NULL,
-                     agroclimatology = FALSE, CSV = FALSE, GPKG = FALSE) {
-  
+get_GSOD <- function(years = NULL,
+                     station = NULL,
+                     country = NULL,
+                     dsn = NULL,
+                     filename = NULL,
+                     max_missing = NULL,
+                     agroclimatology = FALSE,
+                     CSV = FALSE,
+                     GPKG = FALSE) {
   # Create objects for use in retrieving files ---------------------------------
   cache_dir <- tempdir()
   ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/%s/"
-  
+
   # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   if (!is.null(dsn)) {
@@ -234,72 +240,80 @@ get_GSOD <- function(years = NULL, station = NULL, country = NULL,
     stations <- .fetch_station_list()
   }
   # Validate user entered stations for existence in stations list from NCDC
-  plyr::l_ply(.data = station, .fun = .validate_station, stations = stations,
-              years = years)
-  
+  plyr::l_ply(
+    .data = station,
+    .fun = .validate_station,
+    stations = stations,
+    years = years
+  )
+
   .validate_country(country)
-  
+
   # Download files from server -------------------------------------------------
-  
+
   GSOD_list <- .download_files(ftp_base, station, years, cache_dir)
-  
+
   # Validate stations for missing days -----------------------------------------
   if (!is.null(max_missing)) {
-    GSOD_list <- .validate_missing_days(max_missing, GSOD_list, cache_dir)
+    GSOD_list <-
+      .validate_missing_days(max_missing, GSOD_list, cache_dir)
   }
-  
+
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
-    GSOD_list <- .agroclimatology_list(GSOD_list, stations, cache_dir, years)
+    GSOD_list <-
+      .agroclimatology_list(GSOD_list, stations, cache_dir, years)
   }
-  
+
   # Subset GSOD_list for specified country -------------------------------------
   if (!is.null(country)) {
-    GSOD_list <- .country_list(country, GSOD_list, stations, cache_dir, years)
+    GSOD_list <-
+      .country_list(country, GSOD_list, stations, cache_dir, years)
   }
-  
-  # Download files from FTP server ---------------------------------------------
-  
-  .download_station_files(cache_dir, ftp_base, station, years)
-  
+
   # Clean and reformat list of station files from local disk in tempdir --------
-  
+
[…]
