+++ man/get_GSOD.Rd
@@ -226,14 +226,16 @@ get_GSOD(years = 2010:2011, dsn = "~/",
 filename = "GSOD_agroclimatology_2010-2011", agroclimatology = TRUE)
 
 }
-
 }
 \author{
-Adam Sparks \email{adamhsparks@gmail.com}
+Adam H Sparks \email{adamhsparks@gmail.com}
 }
 \references{
 {Jarvis, A, HI Reuter, A Nelson, E Guevara, 2008, Hole-filled
 SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database
 \url{http://srtm.csi.cgiar.org}}
+}
+\seealso{
+\code{\link{reformat_GSOD}}
 }
 
