### 2016-11-23 89d2883
fix bugs in get_GSOD and update documentation
