+++ R/get_GSOD.R
@@ -197,7 +197,7 @@
 #' home directory with a maximum of five missing days per station allowed.
 #'
 #' get_GSOD(years = 2010, country = "Philippines", dsn = "~/",
-#' filename = "Philippines_GSOD", GPKG = TRUE, CSV = FALSE)
+#' filename = "Philippines_GSOD", GPKG = TRUE, CSV = FALSE, max_missing = 5)
 #'
 #' # Download global GSOD data for agroclimatology work for years 2009 and 2010
 #' # and generate yearly summary files, GSOD-agroclimatology-2010.csv and
@@ -247,17 +247,11 @@ get_GSOD <- function(years = NULL,
     years = years
   )
 
-  .validate_country(country)
+  country <- .validate_country(country)
 
   # Download files from server -------------------------------------------------
 
   GSOD_list <- .download_files(ftp_base, station, years, cache_dir)
-
-  # Validate stations for missing days -----------------------------------------
-  if (!is.null(max_missing)) {
-    GSOD_list <-
-      .validate_missing_days(max_missing, GSOD_list, cache_dir)
-  }
 
   # Subset GSOD_list for agroclimatology only stations -------------------------
   if (isTRUE(agroclimatology)) {
@@ -269,6 +263,13 @@ get_GSOD <- function(years = NULL,
   if (!is.null(country)) {
     GSOD_list <-
       .country_list(country, GSOD_list, stations, cache_dir, years)
+  }
+
+  # Validate stations for missing days -----------------------------------------
+  if (!is.null(max_missing)) {
+    message("Checking stations against max_missing value")
+    GSOD_list <-
+      .validate_missing_days(max_missing, GSOD_list, cache_dir)
   }
 
   # Clean and reformat list of station files from local disk in tempdir --------
@@ -288,6 +289,7 @@ get_GSOD <- function(years = NULL,
     message("\nWriting CSV file to disk.\n")
     outfile <- paste0(outfile, ".csv")
     readr::write_csv(GSOD_XY, path = paste0(outfile))
+    rm(outfile)
   }
 
   if (isTRUE(GPKG)) {
@@ -295,7 +297,7 @@ get_GSOD <- function(years = NULL,
     outfile <- paste0(outfile, ".gpkg")
     # Convert object to standard df and then spatial object
     GSOD_XY <- as.data.frame(GSOD_XY)
-    sp::coordinates(GSOD_XY) <- ~ LON + LAT
+    sp::coordinates(GSOD_XY) <- ~LON+LAT
     sp::proj4string(GSOD_XY) <-
       sp::CRS("+proj=longlat +datum=WGS84")
 
@@ -304,11 +306,10 @@ get_GSOD <- function(years = NULL,
       file.remove(outfile)
     }
     # Create new .gpkg file
-    rgdal::writeOGR(
-      GSOD_XY,
-      dsn = path.expand(outfile),
-      layer = "GSOD",
-      driver = "GPKG"
+    rgdal::writeOGR(GSOD_XY,
+                    dsn = path.expand(outfile),
+                    layer = "GSOD",
+                    driver = "GPKG"
     )
   }
 
@@ -372,13 +373,12 @@ get_GSOD <- function(years = NULL,
       substr(dsn, nchar(dsn), nchar(dsn)) != "\\") {
     dsn <- paste0(dsn, "/")
   }
-
   if (is.null(filename)) {
     filename_out <- "GSOD"
   } else {
     filename_out <- filename
   }
-  outfile <- paste0(dsn, "/", filename_out)
+  outfile <- paste0(dsn, filename_out)
   return(outfile)
 }
 
@@ -457,10 +457,13 @@ get_GSOD <- function(years = NULL,
[…]
