### 2016-11-23 3d7f702
Fix bug when writing Geopackages to disk
