### 2016-11-23 a37caa1
Fixes in tests; remove extra parameter in subfunction
