### 2016-11-23 13bd488
Fix notes in check and errors in tests
