### 2016-11-23 45134f7
Fix tests for valid country entries
