### 2016-11-23 c8c03f3
Clean up loop station vignette
