+++ vignettes/Specified_stations_for_a_range_of_years.Rmd
@@ -82,8 +82,6 @@ loop_stations <- station_meta[station_meta$STNID %in% loop_stations, ]
 loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
                                  loop_stations$END >= 20151231, ]
 
-loop_stations
-
 ```
 
 ## Using get_GSOD to Fetch the Requested Station Files
@@ -92,17 +90,23 @@ combinations. Often the FTP server becomes overwhelmed and stops responding
 to requests.
 
 This example shows how you could construct a query using the `get_GSOD`
-function. Be aware that it may or will likely result in incomplete data and
-an error from the server.
+function. Be aware that it may result in incomplete data and error from the
+server. If it does this, see the following option for using the `format_GSOD`
+function.
 
 ```{r, fetch_weather, message=FALSE, eval=FALSE}
-get_GSOD(station = eval(parse(text = loop_stations[, 12])), years = 1960:2016,
+get_GSOD(station = eval(parse(text = loop_stations[, 12])), years = 1960:2015,
                               CSV = TRUE, dsn = "~/",
                               filename = "Loop_Survey_Weather_1960-1969")
 
 ```
 
-## A Safer Option
+## Another Option
+`GSODR` provides a function for dealing with local files that have been
+transfered from the server already as well, `reformat_GSOD`. If the previous
+example with `get_GSOD` does not work, this is a good alternative that takes a
+bit more intervention but gives the same results.
+
 Using your FTP client, e.g. FileZilla, log into the NCDC FTP server,
 ftp.ncdc.noaa.gov and navigate to /pub/data/gsod/. Manually downloading the
 files for each station listed above from 1960 to 2015 is possible, but tedious.
@@ -126,7 +130,7 @@ local_files <- local_files[basename(local_files) %in% loop_stations]
 
 loop_data <- reformat_GSOD(file_list = local_files)
 
-write.csv(loop_data, file = "loop_data")
+write.csv(loop_data, file = "Loop_Survey_Weather_1960-1969")
 ```
 
 
