### 2016-11-23 337a654
Fix tests, delete extra files
