+++ vignettes/GSODR.md
@@ -1,479 +1,0 @@
-<!--
-%\VignetteEngine{knitr::knitr}
-%\VignetteIndexEntry{GSODR}
-%\VignetteEncoding{UTF-8}
--->
-
-# Introduction
-An R package that provides a function that automates downloading and
-cleaning data from the [Global Surface Summary of the Day (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)
-data provided by the US National Climatic Data Center (NCDC). All units are
-converted to International System of Units (SI), e.g., inches to millimetres and
-Fahrenheit to Celsius. Output is saved as a Comma Separated Value (CSV) file or
-in a spatial GeoPackage (GPKG) file, implemented by most major GIS software,
-summarising each year by station, which also includes vapour pressure and
-relative humidity variables calculated from existing data in GSOD.
-
-For more information see the description of the data provided by NCDC,
-<http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt>.
-
-# Using the package
-
-## Load GSODR
-
-```r
-library(GSODR)
-```
-
-## Plot Global Station Locations
-
-The GSOD data are comprised of a global set of data from weather stations. To
-visualise where these stations are located we can fetch the station metadata and
-plot it in a map. The resulting map shows only stations with valid geo-locations
-after filtering.
-
-```r
-  GSOD_stations <- readr::read_csv(
-    "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
-    col_types = "ccccccddddd",
-    col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL",
-                  "LAT", "LON", "ELEV_M", "BEGIN", "END"), skip = 1)
-
-  GSOD_stations[GSOD_stations == -999.9] <- NA
-  GSOD_stations[GSOD_stations == -999] <- NA
-
-  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LAT) &
-                                   !is.na(GSOD_stations$LON), ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LAT != 0 &
-                                   GSOD_stations$LON != 0, ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LAT > -90 &
-                                   GSOD_stations$LAT < 90, ]
-  GSOD_stations <- GSOD_stations[GSOD_stations$LON > -180 &
-                                   GSOD_stations$LON < 180, ]
-  GSOD_stations$STNID <- as.character(paste(GSOD_stations$USAF,
-                                            GSOD_stations$WBAN, sep = "-"))
-  
-  SRTM_GSOD_elevation <- data.table::setkey(GSODR::SRTM_GSOD_elevation, STNID)
-  data.table::setDT(GSOD_stations)
-  data.table::setkey(GSOD_stations, STNID)
-  GSOD_stations <- GSOD_stations[SRTM_GSOD_elevation, on = "STNID"]
-  
-  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LAT), ]
-  GSOD_stations <- GSOD_stations[!is.na(GSOD_stations$LON), ]
-```
-
-Using [ggplot2](https://CRAN.R-project.org/package=ggplot2) and the 
-[ggalt](https://CRAN.R-project.org/package=ggalt) package it is possible to plot
-the station locations using alpha transparency to see the densest part of the
[…]
