### 2016-11-23 5d1012b
Lint and reformat code
