+++ R/get_GSOD.R
@@ -1,3 +1,5 @@
+
+
 #' Download, Clean, Reformat and Generate New Variables From GSOD Weather Data
 #'
 #'This function automates downloading, cleaning, reformatting of data from
@@ -229,7 +231,6 @@ get_GSOD <- function(years = NULL,
   # Create objects for use in retrieving files ---------------------------------
   cache_dir <- tempdir()
   ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/%s/"
-
   # Validate user inputs -------------------------------------------------------
   .validate_years(years)
   if (!is.null(dsn)) {
@@ -246,81 +247,69 @@ get_GSOD <- function(years = NULL,
     stations = stations,
     years = years
   )
-
   country <- .validate_country(country)
 
-  # Download files from server -------------------------------------------------
-
+    # Download files from server -----------------------------------------------
   GSOD_list <- .download_files(ftp_base, station, years, cache_dir)
 
-  # Subset GSOD_list for agroclimatology only stations -------------------------
+    # Subset GSOD_list for agroclimatology only stations -----------------------
   if (isTRUE(agroclimatology)) {
     GSOD_list <-
       .agroclimatology_list(GSOD_list, stations, cache_dir, years)
   }
-
   # Subset GSOD_list for specified country -------------------------------------
   if (!is.null(country)) {
     GSOD_list <-
       .country_list(country, GSOD_list, stations, cache_dir, years)
   }
-
   # Validate stations for missing days -----------------------------------------
   if (!is.null(max_missing)) {
     message("Checking stations against max_missing value")
     GSOD_list <-
       .validate_missing_days(max_missing, GSOD_list)
   }
-
   # Clean and reformat list of station files from local disk in tempdir --------
-
   message("Starting data file processing")
   GSOD_XY <- as.data.frame(try(plyr::ldply(
     .data = GSOD_list,
     .fun = .process_gz,
     stations = stations,
     .progress = "text"
-  )))
-
-
+  ))
+  )
   # Write files to disk --------------------------------------------------------
-
   if (isTRUE(CSV)) {
     message("\nWriting CSV file to disk.\n")
     outfile <- paste0(outfile, ".csv")
     readr::write_csv(GSOD_XY, path = paste0(outfile))
     rm(outfile)
   }
-
   if (isTRUE(GPKG)) {
     LAT <- LON <- NULL
     message("\nWriting GeoPackage File to Disk.\n")
     outfile <- paste0(outfile, ".gpkg")
     # Convert object to standard df and then spatial object
     GSOD_XY <- as.data.frame(GSOD_XY)
-    sp::coordinates(GSOD_XY) <- ~LON+LAT
-    sp::proj4string(GSOD_XY) <- sp::CRS("+proj=longlat +datum=WGS84")
-
+    sp::coordinates(GSOD_XY) <- ~ LON + LAT
+    sp::proj4string(GSOD_XY) <-
+      sp::CRS("+proj=longlat +datum=WGS84")
     # If the filename specified exists, remove it and create new
     if (file.exists(path.expand(outfile))) {
       file.remove(outfile)
     }
     # Create new .gpkg file
-    rgdal::writeOGR(GSOD_XY,
-                    dsn = path.expand(outfile),
-                    layer = "GSOD",
-                    driver = "GPKG"
+    rgdal::writeOGR(
+      GSOD_XY,
+      dsn = path.expand(outfile),
+      layer = "GSOD",
+      driver = "GPKG"
     )
   }
-
   return(GSOD_XY)
-
   # Cleanup and reset to default state -----------------------------------------
   unlink(cache_dir)
 }
-
[…]
