### 2016-11-23 7e5f34c
Fix tests
