+++ tests/testthat/test-get_GSOD.R
@@ -146,11 +146,11 @@ test_that("Check validate country returns a two letter code", {
 test_that("Check validate country returns an error on invalid entry", {
   country <- "Philipines"
   expect_error(.validate_country(country),
-               "\nPlease provide a valid name or 2 or 3 letter ISO country code;\n            you can view the entire list of valid countries in this data by\n            typing, 'country_list'.\n")
+               "\nPlease provide a valid name or 2 or 3 letter ISO country code;\n              you can view the entire list of valid countries in this data by\n              typing, 'country_list'.\n")
 
   country <- "RP"
   expect_error(.validate_country(country),
-               "\nPlease provide a valid name or 2 or 3 letter ISO country code;\n            you can view the entire list of valid countries in this data by\n            typing, 'country_list'.\n")
+               "\nPlease provide a valid name or 2 or 3 letter ISO country code;\n              you can view the entire list of valid countries in this data by\n              typing, 'country_list'.\n")
 
 })
 
