### 2016-11-23 a7c4add
Update documentation; skip all tests on CRAN and one on appveyor
