+++ vignettes/Specified_stations_for_a_range_of_years.md
@@ -1,0 +1,137 @@
+<!--
+%\VignetteEngine{knitr::knitr}
+%\VignetteIndexEntry{GSODR - Finding and fetching stations for Central Luzon}
+%\VignetteEncoding{UTF-8}
+-->
+# Gather Data for IRRI Central Luzon Survey Loop
+
+The IRRI (International Rice Research Institute) survey loop in Central Luzon
+is a study that aims to monitor the changes in rice farming in the major rice
+producing area of the Philippines - the Central Luzon region, which is called as
+the "rice bowl of the Philippines". Data have been collected in this project
+since the 1960s. See, <http://ricestat.irri.org/fhsd/php/panel.php?page=1&sortBy=title&sortOrder=ascending#>
+for the panel data.
+
+This vignette details how to find and retrieve weather data for the area that
+this survey covers for the time period of 1960-2016. Methods that are detailed
+include: 
+  * retrieving a spatial object of provincial level data;
+  * subsetting this data for the provinces of interest;
+  * merging the polygons into one object;
+  * finding the centroid of this resulting polygon;
+  * using the centroid of the polygon to find stations within 100km of this
+  point;
+  * determining which stations provide data for the specified time-period, 1960-
+  2016; and 
+  * downloading the station files and creating a single CSV file of the data.
+
+## Retrieve PHL Provincial Data and Select Loop Provinces
+First we retrieve data from GADM.org that will provide the provincial spatial
+data for the survey area. We will then use this to find the centroid, which will
+be used to find the nearest stations.
+
+```r
+library(raster)
+library(rgdal)
+
+RP <- getData(country = "Philippines", level = 1)
+```
+Select the provinces involved in the survey and make a new object called
+`Central_Luzon`.
+
+
+```r
+Central_Luzon <- RP[RP@data$NAME_1 == "Pampanga" | 
+           RP@data$NAME_1 == "Tarlac" |
+           RP@data$NAME_1 == "Pangasinan" |
+           RP@data$NAME_1 == "La Union" |
+           RP@data$NAME_1 == "Nueva Ecija" |
+           RP@data$NAME_1 == "Bulacan", ]
+```
+
+## Dissolve Polygons and Find Centroid of Loop Survey Area
+Now that we have the provincial data imported, we will dissolve the polygons and
+find the centroid.
+
+
+```r
+library(rgeos)
+Central_Luzon <- gUnaryUnion(Central_Luzon)
+centroid <- gCentroid(Central_Luzon)
+```
+
+
+```r
+library(GSODR)
+library(readr)
+# Fetch station list from NCDC
+station_meta <- read_csv(
+  "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
+  col_types = "ccccccddddd",
+  col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL", "LAT",
+                "LON", "ELEV_M", "BEGIN", "END"), skip = 1)
+station_meta$STNID <- as.character(paste(station_meta$USAF,
+                                         station_meta$WBAN,
+                                         sep = "-"))
+
+loop_stations <- nearest_stations(LAT = centroid@coords[, 2],
+                                  LON = centroid@coords[, 1], 
+                                  distance = 100)
+
+loop_stations <- station_meta[station_meta$STNID %in% loop_stations, ]
+
+loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
+                                 loop_stations$END >= 20151231, ]
+```
+
+## Using get_GSOD to Fetch the Requested Station Files
+Using the `get_GSOD` function may not work with this many station and year
+combinations. Often the FTP server becomes overwhelmed and stops responding
+to requests.
+
[…]
