### 2016-11-23 dcddc41
Add GSOD station files for testing to data-raw and clean up .gitignore
