+++ R/reformat_GSOD.R
@@ -144,40 +144,6 @@
 #' \item{RH}{Mean daily relative humidity}
 #' }
 #'
-#'
-#' @note Some of these data are redistributed with this R package. Originally
-#' from these data come from the US NCDC which states that users of these data
-#' should take into account the following: \dQuote{The following data and
-#' products may have conditions placed on their international commercial use.
-#' They can be used within the U.S. or for non-commercial international
-#' activities without restriction. The non-U.S. data cannot be redistributed for
-#' commercial purposes. Re-distribution of these data by others must provide
-#' this same notification.}
-#'
-#' @examples
-#' \dontrun{
-#' # Download weather station for Toowoomba, Queensland for 2010
-#' t <- get_GSOD(years = 2010, station = "955510-99999")
-#'
-#' # Download data for Philippines for year 2010 and generate a yearly
-#' # summary GeoPackage file, Philippines_GSOD-2010.gpkg, file in the user's
-#' home directory with a maximum of five missing days per station allowed.
-#'
-#' get_GSOD(years = 2010, country = "Philippines", dsn = "~/",
-#' filename = "Philippines_GSOD", GPKG = TRUE, CSV = FALSE)
-#'
-#' # Download global GSOD data for agroclimatology work for years 2009 and 2010
-#' # and generate yearly summary files, GSOD-agroclimatology-2010.csv and
-#' # GSOD-agroclimatology-2011.csv in the user's home directory with a maximum
-#' # of five missing days per weather station allowed.
-#'
-#' get_GSOD(years = 2010:2011, dsn = "~/",
-#' filename = "GSOD_agroclimatology_2010-2011", agroclimatology = TRUE)
-#'
-#' }
-#'
-#'
-#'
 #' @note Some of these data are redistributed with this R package. Originally
 #' from these data come from the US NCDC which states that users of these data
 #' should take into account the following: \dQuote{The following data and
