+++ tests/testthat/test-reformat_GSOD.R
@@ -7,11 +7,11 @@ test_that("reformat_GSOD file_list parameter reformats data properly", {
   skip_on_appveyor()
   skip_on_travis()
 
-  y <- list.files("data-raw/GSOD-files", full.names = TRUE)
+  y <- list.files("~/Development/GSODR/data-raw/GSOD-files", full.names = TRUE)
   x <- reformat_GSOD(file_list = y)
   expect_equal(nrow(x), 1454)
   expect_equal(ncol(x), 48)
-  expect_type(x, "data.frame")
+  expect_type(x, "list")
 })
 
 test_that("reformat_GSOD dsn parameter reformats data properly", {
@@ -19,9 +19,9 @@ test_that("reformat_GSOD dsn parameter reformats data properly", {
   skip_on_appveyor()
   skip_on_travis()
 
-  x <- reformat_GSOD(dsn = "data-raw/GSOD-files")
+  x <- reformat_GSOD(dsn = "~/Development/GSODR/data-raw/GSOD-files")
   expect_equal(nrow(x), 1454)
   expect_equal(ncol(x), 48)
-  expect_type(x, "data.frame")
+  expect_type(x, "list")
 })
 
