+++ README.Rmd
@@ -24,23 +24,36 @@ GSODR: Global Surface Summary Daily Weather Data in R
 
 # Introduction to GSODR
 
-The GSODR package is an R package that provides a function that automates
-downloading and cleaning of data from the "[Global Surface Summary of the Day
-(GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)"
-weather station data provided by the US National Climatic Data Center (NCDC).
-Data are formatted for easy use in R, returned as a `data.frame` object in R
-summarising each year by station, with options to save as a Comma Separated
-Value (CSV) file or as a spatial GeoPackage (GPKG) file, implemented by most
-major GIS softwares.
+The GSOD or [Global Surface Summary of the Day (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)
+data provided by the US National Climatic Data Center (NCDC) are a valuable
+source of weather data with global coverage. However, the data files are
+cumbersome and difficult to work with. The GSODR package aims to make it easy to
+find, tranfer and format the data you need for use in analysis. The GSODR 
+package provides three main functions for facilitating this:
 
-Station files are individually checked for number of missing days to assure data
-quality, stations with too many missing observations are omitted. All
-units are converted to International System of Units (SI), e.g., inches to
-millimetres and Fahrenheit to Celsius.
+* `get_GSOD` - the main function that will query and transfer files from the FTP
+server, reformat them and return a data.frame in R or save a file to disk  
+* `reformat_GSOD` - the workhorse, this function takes individual station files
+on the local disk and reformats them returning a data.frame in R  
+* `nearest_stations` - this function returns a dataframe containing a list of
+stations and their metadata that fall within the given radius of a point\
+specified by the user
+
+When reformatting data either with `get_GSOD` or `reformat_GSOD`, all units are
+converted to International System of Units (SI), e.g., inches to millimetres and
+Fahrenheit to Celsius. File output can be saved as a Comma Separated Value (CSV)
+file or in a spatial GeoPackage (GPKG) file, implemented by most major GIS
+software, summarising each year by station, which also includes vapour pressure
+and relative humidity variables calculated from existing data in GSOD.
 
 Additional data are calculated by this R package using the original data and
 included in the final data. These include vapour pressure (ea and es) and
 relative humidity.
+
+For more information see the description of the data provided by NCDC,
+<http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt>.
+
+## Other Sources of Weather Data
 
 There are several other sources of weather data and ways of retrieving them
 through R. In particular, the excellent 
@@ -101,22 +114,20 @@ head(Tbar)
 For years 2010 and 2011, download data and create the files, 
 GSOD-agroclimatology-2010.csv and GSOD-agroclimatology-2011.csv, in the user's
 home directory with a maximum of five missing days per weather station allowed.
-Use parallel processing to run the process more quickly.
 
 ```{r example_2, eval=FALSE}
 
 get_GSOD(years = 2010:2011, dsn = "~/", filename = "GSOD-agroclimatology",
-         agroclimatology = TRUE, max_missing = 5, threads = 3)
+         agroclimatology = TRUE, max_missing = 5)
 
 ```
-#### Example 3 - Download data for a single country and plot it
+#### Example 3 - Download and plot data for a single country
 Download data for Philippines for year 2010 and generate a spatial, year summary
[…]
