### 2016-11-24 0ae25af
Convert NA to -9999 for Geopackage files
Only for Geopackage files are the NA values converted to -9999 on
writing to disk to fix issues with writing the file to disk
