### 2016-11-24 92c1135
Fix bug with writing GPKG files; fixes graph for Philippine T in README too
