### 2016-11-24 804cb96
Delete extra file in vignettes
