### 2016-11-24 40d09cd
Update NEWS.md
