+++ NEWS.md
@@ -9,7 +9,8 @@
   * A new function, `nearest_stations` is now included to find stations within
   a user specified radius (in kilometres) of a point given as latitude and
   longitude in decimal degrees  
-  * A vignette is now included  
+  * A general use vignette is now included  
+  * New vignette with a detailed use-case  
   * Output files now include fields for State (US only) and Call (International
   Civil Aviation Organization (ICAO) Airport Code)  
   * Use FIPS codes in place of ISO3c for file name and in output files because
@@ -32,8 +33,14 @@
   automatically merge it with the CGIAR-CSI SRTM elevation values provided by
   this package. Previously, the package provided it's own list of station
   information, which was difficult to keep up-to-date  
-
-
+  * A new `reformat_GSOD` function reformats station files in
+  "WMO-WBAN-YYYY.op.gz" format that have been downloaded from the United States
+  National Climatic Data Center's (NCDC) FTP server.  
+  * New data layers are included, CHELSA (climatic surfaces at 1 km resolution),
+  http://chelsa-climate.org/, http://www.earthenv.org/cloud,
+  http://maps.elie.ucl.ac.be/CCI/viewer/index.php
+  * Improved file handling for individual station downloads  
+  
 ## MINOR IMPROVEMENTS
 
   * `get_GSOD` function optimised for speed as best possible after FTPing files
