### 2016-11-24 5d50de9
Update NEWS.md
