### 2016-11-25 2332bb2
Update documentation; Edits to paper.md/paper.bib; New vignette
Fix incorrect entry in documentation for Feb. snow cover fraction;

Edit paper.md for new features and clarity;

Edit paper.bib to remove extra references that are not used and include
new references for climate data;

Add vignette based on Tom's work in README.md to show how to use GSODR
with spatial files and the included climate data;
