+++ inst/paper/paper.md
@@ -1,71 +1,84 @@
 ---
 title: 'GSODR: Global Surface Summary Daily Weather Data in R'
-tags:
-  - Global Surface Summary of the Day
-  - GSOD
-  - meteorology
-  - climatology
-  - weather data
-  - R
 authors:
-  - name: Adam H Sparks
-    orcid: 0000-0002-0061-8359
-    affiliation: 1
-  - name: Tomislav Hengl
-    orcid: 0000-0002-9921-5129
-    affiliation: 2
-  - name: Andrew Nelson
-    orcid: 0000-0002-7249-3778
-    affiliation: 3
-affiliations:
-  - name: Centre for Crop Health, University of Southern Queensland, Toowoomba Queensland 4350, Australia
-    index: 1
-  - name: ISRIC - World Soil Information, P.O. Box 353, 6700 AJ Wageningen, The Netherlands
-    index: 2
-  - name: Faculty of Geo-Information and Earth Observation (ITC), University of Twente, Enschede 7500 AE, The Netherlands
-    index: 3
-date: "16 November 2016"
+- affiliation: 1
+  name: Adam H Sparks
+  orcid: 0000-0002-0061-8359
+- affiliation: 2
+  name: Tomislav Hengl
+  orcid: 0000-0002-9921-5129
+- affiliation: 3
+  name: Andrew Nelson
+  orcid: 0000-0002-7249-3778
+date: "25 November 2016"
+output: pdf_document
 bibliography: paper.bib
+tags:
+- Global Surface Summary of the Day
+- GSOD
+- meteorology
+- climatology
+- weather data
+- R
+affiliations:
+- index: 1
+  name: Centre for Crop Health, University of Southern Queensland, Toowoomba Queensland
+    4350, Australia
+- index: 2
+  name: ISRIC - World Soil Information, P.O. Box 353, 6700 AJ Wageningen, The Netherlands
+- index: 3
+  name: Faculty of Geo-Information and Earth Observation (ITC), University of Twente, Enschede 7500 AE, The Netherlands
 ---
 
 # Summary
 
-The GSODR package [@GSODR] is an R package [@R-base] for automated
+The GSODR package [@GSODR] is an R package [@R-base] providing automated
 downloading, parsing, cleaning and of Global Surface Summary of the
-Day (GSOD) [@NCDC] weather data for use in R or saving into Comma Separated
-Values (CSV) or Geopackage (GPKG) [@geopackage] files. It builds on or
-complements several other scripts and packages. We take advantage of modern
-techniques in R to make more efficient use of available computing resources to
-complete the process, e.g., data.table [@data.table] and readr [@readr], which
-allow the data cleaning, conversions and disk input/output processes to take
-advantage of modern computer hardware. The rnoaa [@rnoaa] package offers an
-excellent suite of tools for interacting with and downloading weather data from
-the United States National Oceanic and Atmospheric Administration, but lacks
-options for GSOD data retrieval. Several other APIs and R packages exist to
-access weather data, but most are region or continent specific, whereas GSOD is
-global. This package was developed to provide:
+Day (GSOD) [@NCDC] weather data for use in R or saving as local files in either
+a Comma Separated Values (CSV) or GeoPackage (GPKG) [@geopackage] file. It
+builds on or complements several other scripts and packages. We take advantage
+of modern techniques in R to make more efficient use of available computing
+resources to complete the process, e.g., data.table [@data.table], plyr [@plyr]
+and readr [@readr], which allow the data cleaning, conversions and disk 
+input/output processes to take advantage of modern computer hardware. The rnoaa 
+[@rnoaa] package already offers an excellent suite of tools for interacting with
+and downloading weather data from the United States National Oceanic and
[…]
