+++ inst/paper/paper.bib
@@ -35,7 +35,7 @@
     title = {{GSODR}: Global Summary Daily Weather Data in R},
     author = {Adam Sparks and Tomislav Hengl and Andrew Nelson},
     year = {2016},
-    note = {R package version 0.2},
+    note = {R package version 1.0},
     url = {https://github.com/adamhsparks/GSODR},
   }
 
@@ -62,20 +62,42 @@
     url = {https://CRAN.R-project.org/package=data.table},
   }
 
-  @Manual{foreach,
-    title = {foreach: Provides Foreach Looping Construct for R},
-    author = {{Revolution Analytics} and Steve Weston},
-    year = {2015},
-    note = {R package version 1.4.3},
-    url = {https://CRAN.R-project.org/package=foreach},
+  @article{Karger2016,
+    title = {Climatologies at high resolution for the Earth land surface areas},
+    author = {Dirk Nikolaus Karger and Olaf Conrad and Jürgen Böhner and
+    Tobias Kawohl and Holger Kreft and Rodrigo Wilber Soria-Auza and
+    Niklaus Zimmermann and H. Peter Linder and Michael Kessler},
+    year = {2016},
+    journal = {arXiv preprint},
+    doi = {arXiv:1607.00217},
+    url = {http://chelsa-climate.org/},
   }
 
-  @article{Kilibarda2014,
-    title = {Spatio‐temporal interpolation of daily temperatures for global land areas at 1 km resolution},
-    author = {Kilibarda, M and Hengl, T and Heuvelink, G and Gräler, B and Pebesma, E and Perčec Tadić, M and Bajat, B},
-    year = {2014},
-    journal = {Journal of Geophysical Research: Atmospheres},
-    volume = {119},
-    issue = {5},
-    pages = {2294-2313},
+  @Article{plyr,
+    title = {The Split-Apply-Combine Strategy for Data Analysis},
+    author = {Hadley Wickham},
+    journal = {Journal of Statistical Software},
+    year = {2011},
+    volume = {40},
+    number = {1},
+    pages = {1--29},
+    url = {http://www.jstatsoft.org/v40/i01/},
   }
+
+  @Article{Wilson2016,
+    title = {Remotely Sensed High-Resolution Global Cloud Dynamics for Predicting Ecosystem and Biodiversity Distributions},
+    author = {Wilson, AM and Jetz, W},
+    journal = {PLoS Biology},
+    volume = {14},
+    issue = {3},
+    pages = {e1002415},
+    doi = {10.1371/journal.pbio.1002415},
+    year = {2016},
+  }
+
+  @Misc{ESA2016,
+    title = {{ESA Climate Change Initiative Land Cover Viewer}},
+    howpublished = {\url{http://maps.elie.ucl.ac.be/CCI/viewer/index.php}},
+    year = {2016},
+    note = {Accessed: 2016-11-25},
+}
