+++ DESCRIPTION
@@ -56,6 +56,7 @@ Suggests:
     lubridate,
     plotKML,
     raster,
+    reshape2,
     rgeos,
     rmarkdown,
     spacetime,

[…]
