### 2016-11-25 4584356
Update vignettes to build faster
