+++ vignettes/Specified_stations_for_a_range_of_years.Rmd
@@ -30,38 +30,39 @@ include:
 First we retrieve data from GADM.org that will provide the provincial spatial
 data for the survey area. We will then use this to find the centroid, which will
 be used to find the nearest stations.
-```{r, RP Data, message=FALSE, eval=FALSE}
+
+```r
 library(raster)
 library(rgdal)
 
 RP <- getData(country = "Philippines", level = 1)
-
 ```
 Select the provinces involved in the survey and make a new object called
 `Central_Luzon`.
 
-```{R, Central_Luzon, eval=FALSE}
 
+```r
 Central_Luzon <- RP[RP@data$NAME_1 == "Pampanga" | 
            RP@data$NAME_1 == "Tarlac" |
            RP@data$NAME_1 == "Pangasinan" |
            RP@data$NAME_1 == "La Union" |
            RP@data$NAME_1 == "Nueva Ecija" |
            RP@data$NAME_1 == "Bulacan", ]
-
 ```
 
 ## Dissolve Polygons and Find Centroid of Loop Survey Area
 Now that we have the provincial data imported, we will dissolve the polygons and
 find the centroid.
 
-```{r, dissolve, message=FALSE, eval=FALSE}
+
+```r
 library(rgeos)
 Central_Luzon <- gUnaryUnion(Central_Luzon)
 centroid <- gCentroid(Central_Luzon)
 ```
 
-```{r, nearest_stations, message=FALSE, eval=FALSE}
+
+```r
 library(GSODR)
 library(readr)
 # Fetch station list from NCDC
@@ -82,7 +83,6 @@ loop_stations <- station_meta[station_meta$STNID %in% loop_stations, ]
 
 loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
                                  loop_stations$END >= 20151231, ]
-
 ```
 
 ## Using get_GSOD to Fetch the Requested Station Files
@@ -95,11 +95,11 @@ function. Be aware that it may result in incomplete data and error from the
 server. If it does this, see the following option for using the `format_GSOD`
 function.
 
-```{r, fetch_weather, message=FALSE, eval=FALSE}
+
+```r
 get_GSOD(station = eval(parse(text = loop_stations[, 12])), years = 1960:2015,
                               CSV = TRUE, dsn = "~/",
                               filename = "Loop_Survey_Weather_1960-1969")
-
 ```
 
 ## Another Option
@@ -116,7 +116,8 @@ directory, "gsod-YYYY.tar" and untar them locally and then use R to
 list the available files and select only the files for the stations of interest.
 Lastly, write the data to disk as a CSV file for saving and later use.
 
-```{r, list_local_files, eval=FALSE}
+
+```r
 years <- 1960:2015
 
 loop_stations <- eval(parse(text = loop_stations[, 12]))
@@ -134,4 +135,44 @@ loop_data <- reformat_GSOD(file_list = local_files)
 write.csv(loop_data, file = "Loop_Survey_Weather_1960-1969")
 ```
 
+# Notes
+
+## Sources
+
+#### CHELSA climate layers
+CHELSA (climatic surfaces at 1 km resolution) is based on a quasi-mechanistical
+statistical downscaling of the ERA interim global circulation model
+(Karger et al. 2016). ESA's CCI-LC cloud probability monthly averages are based
+on the MODIS snow products (MOD10A2). <http://chelsa-climate.org/>
+
+#### EarthEnv MODIS cloud fraction 
+<http://www.earthenv.org/cloud>
+
+#### ESA's CCI-LC cloud probability
+<http://maps.elie.ucl.ac.be/CCI/viewer/index.php>
+
+#### Elevation Values
+
+90m hole-filled SRTM digital elevation (Jarvis *et al.* 2008) was used
+to identify and correct/remove elevation errors in data for station
[…]
