### 2016-11-25 c3f551f
Update SRTM_GSOD_elevation.rda data; update country_list.rda data
