+++ data-raw/fetch_country-list.md
@@ -1,7 +1,7 @@
 Fetch GSOD Country List and Merge with ISO Country Codes
 ================
 Adam H. Sparks
-2016-11-14
+2016-11-25
 
 Introduction
 ============
@@ -142,9 +142,10 @@ R System Information
     ## [1] countrycode_0.18
     ## 
     ## loaded via a namespace (and not attached):
-    ##  [1] Rcpp_0.12.7      withr_1.0.2      digest_0.6.10    dplyr_0.5.0     
-    ##  [5] assertthat_0.1   chron_2.3-47     R6_2.2.0         DBI_0.5-1       
-    ##  [9] magrittr_1.5     evaluate_0.10    stringi_1.1.2    curl_2.2        
-    ## [13] data.table_1.9.6 rmarkdown_1.1    devtools_1.12.0  tools_3.3.2     
-    ## [17] stringr_1.1.0    readr_1.0.0      yaml_2.1.13      memoise_1.0.0   
-    ## [21] htmltools_0.3.5  knitr_1.15       tibble_1.2
+    ##  [1] Rcpp_0.12.8      withr_1.0.2      dplyr_0.5.0      digest_0.6.10   
+    ##  [5] rprojroot_1.1    assertthat_0.1   chron_2.3-47     R6_2.2.0        
+    ##  [9] DBI_0.5-1        backports_1.0.4  magrittr_1.5     evaluate_0.10   
+    ## [13] stringi_1.1.2    curl_2.2         data.table_1.9.6 rmarkdown_1.2   
+    ## [17] devtools_1.12.0  tools_3.3.2      stringr_1.1.0    readr_1.0.0     
+    ## [21] yaml_2.1.14      memoise_1.0.0    htmltools_0.3.5  knitr_1.15.1    
+    ## [25] tibble_1.2
