+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: GSODR
 Type: Package
 Title: Global Summary Daily Weather Data in R
-Version: 1.0
+Version: 1.0.0
 Authors@R: c(person("Adam", "Sparks", role = c("aut", "cre"),
     email = "adamhsparks@gmail.com"),
     person("Tomislav", "Hengl", role = "aut",
@@ -62,7 +62,8 @@ Suggests:
     spacetime,
     testthat,
     tibble,
-    tidyr
+    tidyr,
+    covr
 RoxygenNote: 5.0.1
 Encoding: UTF-8
 NeedsCompilation: no
