### 2016-11-25 d4e469a
More setup for Codecov
