+++ .travis.yml
@@ -36,11 +36,11 @@ script:
 after_failure:
   - ./pkg-build.sh dump_logs
 
-r_github_packages:
-  - jimhester/covr
+r_packages:
+  - covr
 
 after_success:
-  - Rscript -e 'covr::codecov()'
+  - Rscript -e 'library(covr); codecov()'
 
 notifications:
   email:
