### 2016-11-25 e805cd9
Revert "More setup for Codecov"
This reverts commit d4e469a61811698cadaf3a5799f79f27423ba51d.
