### 2016-11-25 a688ad9
Revert "Track code coverage"
This reverts commit f7e13b6e35440fee74109322493229e39ffce59b.
