### 2016-11-25 e7903a8
Set up code coverage
