+++ README.Rmd
@@ -18,6 +18,7 @@ GSODR: Global Surface Summary Daily Weather Data in R
 
 [![Travis-CI Build Status](https://travis-ci.org/adamhsparks/GSODR.svg?branch=master)](https://travis-ci.org/adamhsparks/GSODR)
 [![Build status](https://ci.appveyor.com/api/projects/status/8daqtllo2sg6me07/branch/master?svg=true)](https://ci.appveyor.com/project/adamhsparks/GSODR/branch/master?svg=true)
+[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/GSODR/master.svg)](https://codecov.io/github/adamhsparks/GSODR?branch=master)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/GSODR?color=blue)](https://github.com/metacran/cranlogs.app)
 [![cran version](http://www.r-pkg.org/badges/version/GSODR)](https://cran.r-project.org/package=GSODR)
 [![DOI](https://zenodo.org/badge/32764550.svg)](https://zenodo.org/badge/latestdoi/32764550)
