### 2016-11-25 c43744f
Documentation updates
