+++ DESCRIPTION
@@ -21,15 +21,15 @@ Description: The GSODR package offers automated downloading, parsing, cleaning,
     reported latitude and longitude values are permitted in the final data.
     Additional useful variables, saturation vapour pressure (es), actual vapour
     pressure (ea) and relative humidity are calculated from the original data
-    and included in the final data set. A list of elevation values generated from
-    a 200 metre buffer of 90m elevation data, CGIAR-CSI SRTM hole-filled 90
-    metre data (Jarvis et al. 2008), are included for stations between -60 and
-    60 degrees latitude. The resulting data include station data (e.g., station
-    name, country, latitude, longitude, elevation) and weather observations and
-    associated flags. Data may be automatically saved to disk. File output may
-    be returned as a comma-separated values (CSV) or GeoPackage (GPKG) file.
-    For information on the original data from NCDC, please see the GSOD
-    readme.txt file available from,
+    and included in the final data set. The resulting data include station data
+    (e.g., station name, country, latitude, longitude, elevation) and weather
+    observations and associated flags. Data may be automatically saved to disk.
+    File output may be returned as a comma-separated values (CSV) or GeoPackage
+    (GPKG) file. Additional data are included with this R package: a list of
+    elevation values for stations between -60 and 60 degrees latitude derived
+    from the Shuttle Radar Topography Measuring Mission (SRTM) and three sets of
+    historic climate data. For information on the original GSOD data from NCDC,
+    please see the  GSOD readme.txt file available from,
     <http://www1.ncdc.noaa.gov/pub/data/gsod/readme.txt>.
 Depends:
     R (>= 3.2.0)
