### 2016-11-26 f6beaca
Change fetch_station_list to exported object
This way a user can fetch latest station list for filtering by name,
etc.
