+++ R/fetch_station_list.R
@@ -1,9 +1,18 @@
-
-
-#' @noRd
-# fetch most recent station history file -------------------------------------
-# This function is shared between both get_GSOD and nearest_stations functions
-.fetch_station_list <- function() {
+#' Download the Latest Station List From the NCDC Server
+#'
+#' This function downloads the latest station list from the NCDC FTP server.
+#' This list includes metadata for all stations including unique identifiers,
+#' country, state (if in US), latitude, longitude, elevation and when weather
+#' observations begin and end.
+#'
+#' @examples
+#' \dontrun{
+#' GSOD_stations <- fetch_station_list()
+#' }
+#' @author Adam H Sparks \email{adamhsparks@gmail.com}
+#' @export
+#'
+fetch_station_list <- function() {
   STNID <- NULL
   stations <- readr::read_csv(
     "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
@@ -25,9 +34,9 @@
   )
   stations[stations == -999.9] <- NA
   stations[stations == -999] <- NA
-  stations <- stations[stations$LAT != 0 & stations$LON != 0, ]
-  stations <- stations[stations$LAT > -90 & stations$LAT < 90, ]
-  stations <- stations[stations$LON > -180 & stations$LON < 180, ]
+  stations <- stations[stations$LAT != 0 & stations$LON != 0,]
+  stations <- stations[stations$LAT > -90 & stations$LAT < 90,]
+  stations <- stations[stations$LON > -180 & stations$LON < 180,]
   stations$STNID <-
     as.character(paste(stations$USAF, stations$WBAN, sep = "-"))
   SRTM_GSOD_elevation <-
@@ -35,7 +44,7 @@
   data.table::setDT(stations)
   data.table::setkey(stations, STNID)
   stations <- stations[SRTM_GSOD_elevation, on = "STNID"]
-  stations <- stations[!is.na(stations$LAT), ]
-  stations <- stations[!is.na(stations$LON), ]
+  stations <- stations[!is.na(stations$LAT),]
+  stations <- stations[!is.na(stations$LON),]
   return(stations)
 }
