### 2016-11-27 bd03358
Rename fetch_station_list to get_station_list to be inline with get_GSOD
