### 2016-11-27 1d90d11
Fix documentation; delete extra file
