+++ R/get_station_list.R
@@ -3,7 +3,8 @@
 #' This function downloads the latest station list from the NCDC FTP server.
 #' This list includes metadata for all stations including unique identifiers,
 #' country, state (if in US), latitude, longitude, elevation and when weather
-#' observations begin and end.
+#' observations begin and end. Stations with invalid latitude and longitude
+#' values will not be included.
 #'
 #' @examples
 #' \dontrun{
