### 2016-11-28 b55f247
Updates to README.Rmd
