+++ README.Rmd
@@ -3,141 +3,221 @@ output:
   github_document:
     hard_line_breaks: false
 ---
-
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-```{r, echo = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "README-"
-)
-```
 GSODR: Global Surface Summary Daily Weather Data in R
-================
+=====================================================
 
-[![Travis-CI Build Status](https://travis-ci.org/adamhsparks/GSODR.svg?branch=master)](https://travis-ci.org/adamhsparks/GSODR)
-[![Build status](https://ci.appveyor.com/api/projects/status/8daqtllo2sg6me07/branch/master?svg=true)](https://ci.appveyor.com/project/adamhsparks/GSODR/branch/master?svg=true)
-[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/GSODR/master.svg)](https://codecov.io/github/adamhsparks/GSODR?branch=master)
-[![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/GSODR?color=blue)](https://github.com/metacran/cranlogs.app)
-[![cran version](http://www.r-pkg.org/badges/version/GSODR)](https://cran.r-project.org/package=GSODR)
+[![Travis-CI Build
+Status](https://travis-ci.org/adamhsparks/GSODR.svg?branch=master)](https://travis-ci.org/adamhsparks/GSODR)
+[![Build
+status](https://ci.appveyor.com/api/projects/status/8daqtllo2sg6me07/branch/master?svg=true)](https://ci.appveyor.com/project/adamhsparks/GSODR/branch/master?svg=true)
+[![Coverage
+Status](https://img.shields.io/codecov/c/github/adamhsparks/GSODR/master.svg)](https://codecov.io/github/adamhsparks/GSODR?branch=master)
+[![rstudio mirror
+downloads](http://cranlogs.r-pkg.org/badges/GSODR?color=blue)](https://github.com/metacran/cranlogs.app)
+[![cran
+version](http://www.r-pkg.org/badges/version/GSODR)](https://cran.r-project.org/package=GSODR)
 [![DOI](https://zenodo.org/badge/32764550.svg)](https://zenodo.org/badge/latestdoi/32764550)
 
-# Introduction to GSODR
+Introduction to GSODR
+=====================
 
-The GSOD or [Global Surface Summary of the Day (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)
-data provided by the US National Climatic Data Center (NCDC) are a valuable
-source of weather data with global coverage. However, the data files are
-cumbersome and difficult to work with. The GSODR package aims to make it easy to
-find, tranfer and format the data you need for use in analysis. The GSODR 
-package provides three main functions for facilitating this:
+The GSOD or [Global Surface Summary of the Day
+(GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod)
+data provided by the US National Climatic Data Center (NCDC) are a
+valuable source of weather data with global coverage. However, the data
+files are cumbersome and difficult to work with. The GSODR package aims
+to make it easy to find, tranfer and format the data you need for use in
+analysis. The GSODR package provides three main functions for
+facilitating this:
 
-* `get_GSOD` - the main function that will query and transfer files from the FTP
-server, reformat them and return a data.frame in R or save a file to disk  
-* `reformat_GSOD` - the workhorse, this function takes individual station files
-on the local disk and reformats them returning a data.frame in R  
-* `nearest_stations` - this function returns a dataframe containing a list of
[…]
