### 2016-11-30 edaf09e
Update vignettes and update README.md
Specify `build_vignettes = TRUE` when installing from GitHub

Update vignette YAML headers
