### 2016-11-30 #79 @adamhsparks
I've spent quite a bit of time working on this, I think it's ready for another round of review.

## New features and enhancements included in this version, 1.0.0

I've decided to release this as 1.0.0 when the review is complete. I think it's finally a suitable product that meets the needs of researchers. Hopefully I've addressed the reviewers comments sufficiently.

In addition to addressing the reviewer comments, which I've addressed below:
  * I've included a function to fetch the station list from the server, `get_station_list` so that a user can use it to find stations of interest, and included a vignette that details how to use this functionality;
  * I've included a function which will clean and reformat locally stored weather station files, `reformat_GSOD`. This way users can download files using a FTP client, rather than the `get_GSOD` function, and still take advantage of GSODR's ability to clean and reformat the data;
  * I've included another vignette that details an example of to use the spatial GeoPackage files as well as new climatic data that have been included in the package;
  * New climate data are included in the package now as well to enhance the usefulness of the package.
    * CHELSA (climatic surfaces at 1 km resolution) is based on a quasi-mechanistic statistical downscaling of the ERA interim global circulation model (Karger et al. 2016). ESA's CCI-LC cloud probability monthly averages are based on the MODIS snow products (MOD10A2).
  * References for new data:
    * CHELSA climate layers (http://chelsa-climate.org/) / Karger, D. N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R. W., et al. (2016). Climatologies at high resolution for the Earth land surface areas. arXiv preprint arXiv:1607.00217.
    * EarthEnv MODIS cloud fraction (http://www.earthenv.org/cloud) / Wilson AM, Jetz W (2016). Remotely Sensed High-Resolution Global Cloud Dynamics for Predicting Ecosystem and Biodiversity Distributions. PLoS Biol 14(3): e1002415.
    * ESA's CCI-LC snow cover probability (http://maps.elie.ucl.ac.be/CCI/viewer/index.php)

## Responses to reviewers

### @jeffreyhanson's review

> 1. All examples are wrapped in \dontrun{} so they are not actually tested.

Yes, since they require an active Internet connection and take longer to run than CRAN permits, they are wrapped in a \dontrun{} tag.

> 2. The main function get_GSOD just saves data to disk and does not actually return data. It would be brilliant if this package returned GSOD data so the user does not have to then load it back into R. To that end, I would recommend the function returns a single tibble or SpatialPointsDataFrame object (perhaps depending on an argument?), and let the user then re-save the data where they want it stored.

A data.frame object is now returned as well as the option within the function to define a filename and where to store the file with two options: CSV and a spatial format GeoPackage file.

> 3. Unit testing covers cases where get_GSOD is expected to fail but it does not actually test cases that are expected to succeed.

I've revised the unit testing to cover more of the functions within the package accordingly.

> 4. There is no help page for the package (ie. ?GSODR). It would be really useful to include this, even if it it just a copy-paste from the README. See here for more details.

I've updated the documentation to include a help file that includes the main "description" and links to the GSODR functions and data documentation.

> 5. The nearest_station function will be incredibly useful for a user. It would be even more useful if--instead of printing the stations closest to a user-specified point--it returned the stations. Then this function could be used programmatically to find the closest stations to a series of points, and download data accordingly.

I've updated this accordingly, I've included a more detailed vignette on how someone could use this function to find stations in a specific part of a country of interest.

> 6. The accompanying paper does not list authors in the html document.

All of the author details are complete in the md file, however they don't appear in the HTML version when it is knit.

> 7. This package needs a rerun of devtools::document() before resubmission. When I ran devtools::run_examples(), this automatically re-compiled some Rd files.

The documentation is updated and expanded.

> 8. get_GSOD.R:20 & data.R:29: Both lines refer to a document available on the github repository. I recommend making this document an additional vignette.

There are two Rmd files that document data sources in this package. Technically this isn't any functionality offered by this package or related to using the package. It's documentation that is provided to detail how these elevation data were derived, so they are in the "data-raw" directory, per Hadley's R packages book/webpage. I do not feel that it is necessary to include these as vignettes since the end user does not need to run these functions and they are not functions provided by GSODR.

> 9. As far as I can tell, the vignette cannot be accessed within R. After installation, the following code vignette(package='GSODR') yields 'no vignettes found'. Additionally, the rmarkdown file for the vignette has an obscure name, so even if it could be opened within R, this would be tricky given its obscure name. I recommend naming the vignette after the package so the following code will open it vignette('GSODR').

I've renamed the main vignette to "GSODR" and added two more, one detailing working with GeoPackage files and one with a more detailed look at how to find and download stations for a specific area of interest.

Additionally, when installing from GitHub use, `devtools::install_github("adamhsparks/GSODR", build_vignettes = TRUE)`

> 10a. get_GSOD.R:279: I am unsure why foreach iterators are used here. It would be great to do offer some parallelisation and options for progress bars here. If you wrapped this in a plyr::**ply function, you could easily make it such that: (1) years can easily be processed in parallel, (2) if not processed in parallel, then a progress-bar can be shown.

and

> 10b. get_GSOD.R:345: the user should have explicit control over thread use to avoid accidently fork-bombing their own computer. I recommend providing a threads argument in the get_GSOD function and defaulting it to 1.

Per @deniederhut's comment about parallelisation not improving the speed, I've tested and agree, parallelisation does not render the process significantly faster. I've removed this capability and wrapped these functions in `plyr` functions, which provides progress bars.

> 11a. get_GSOD.R:392--393: Options are restored to the default state but this is not necessarily the previous state. These values could be cached at the beginning of the function. For example: use orginal_options <- options() at the start of the function, and at the end of the function use options(original_options).

and

> 11b. nearest_stations.R:28--29: This function changes options() and does not reset them after it finishes. This is against recommended CRAN practices. Please see previous comment on options for a suggestion on how to fix this.

This has been fixed

> 12. get_GSOD.R:322: ROpensci guidelines recommend using the httr package instead of the RCurl package.

I've updated this with assistance from @hrbrmstr, RCurl is no longer used.

> 13. get_GSOD.R:561: Here and elsewhere return(0) is used after a call to stop(). Please consider removing the return(0) as this is unnecessary .

These have been removed

> 14. nearest_stations.R:34--44: This code could be rewritten for clarity and computational efficiency. I suggest something like this:

This was rewritten as suggested

> 15. get_GSOD.R:83--85: Documentation contains '&lt' which is not rendered as a less than symbol in the html documentation.

This has been corrected

> 16. get_GSOD.R:292: Instead of a call to utils::glob2rx, this could be used instead pattern='^.*\\.gz$'

This has been changed as suggested

> 17. get_GOSD.R:631: This could rewritten be as !vs %in% stations[[12]] for clarity.

This has been changed as suggested

### @deniederhut's review

> 1. vignettes and browseVignettes function called on 'GSODR' returns no vignettes found

This has been fixed, when installing from GitHub use, `devtools::install_github("adamhsparks/GSODR", build_vignettes = TRUE)`

> 2. it seems odd to me the enforce a missingness limit at download time -- I think I would either let the user do this in memory, or set the default limit to 0

This has been changed to not enforce any limit with the user having the option to set missing values as they wish.

> 3. the test suite only seems to be concerned with handling bad input - am I missing the blocks where it is testing that data are returned correctly?

The unit tests have been expanded

> 4. allowing multiple threads for calculation doesn't speed up the execution time as much as I would expect:

I tested this and found the same as well. Parallelisation has been removed to simplify the package dependencies and structure

Thanks to the reviewers for their excellent contributions and comments. It's helped to make this package a much more robust and useful package.
