### 2016-11-30 6018413
add message to individual station download process
