### 2016-12-02 #79 @sckott
Thanks @adamhsparks for the changes and the responses to reviewers

@jeffreyhanson @deniederhut do either or both of you have time to take another look?  if not, i'll do so. let me know
