### 2016-12-02 #79 @deniederhut
@sckott Yes, but not until after Christmas. Is that timeline okay?

On Fri, Dec 2, 2016, 03:22 Scott Chamberlain <notifications@github.com>
wrote:

> Thanks @adamhsparks <https://github.com/adamhsparks> for the changes and
> the responses to reviewers
>
> @jeffreyhanson <https://github.com/jeffreyhanson> @deniederhut
> <https://github.com/deniederhut> do either or both of you have time to
> take another look? if not, i'll do so. let me know
>
