+++ .travis.yml
@@ -47,5 +47,8 @@ env:
     - RVERSION=release
     - RVERSION=devel
 
+r_github_packages:
+  - jimhester/covr
+
 after_success:
   - Rscript -e 'covr::codecov()'
