### 2016-12-03 6010cd3
Update .travis.yml
