+++ .travis.yml
@@ -52,3 +52,4 @@ r_github_packages:
 
 after_success:
   - Rscript -e 'covr::codecov()'
+  - bash <(curl -s https://codecov.io/bash)
