### 2016-12-03 9ea0d7e
Codecov edits
