+++ .travis.yml
@@ -52,4 +52,6 @@ r_github_packages:
 
 after_success:
   - Rscript -e 'covr::codecov()'
+
+after_success:
   - bash <(curl -s https://codecov.io/bash)
