### 2017-01-04 f2098aa
[INCOMPLETE] Add new RMD for fetching CRU data; update vignette index entry
