### 2017-01-05 fe3eeec
Add CRU CL2.0 Data; Update data sources;
Adding climate data from CRU CL2.0:
Mark New (1,*), David Lister (2), Mike Hulme (3), Ian Makin (4)

A high-resolution data set of surface climate over global land areas
Climate Research, 2000, Vol 21, pg 1-25

School of Geography and the Environment, University of Oxford,
Mansfield Road, Oxford OX1 3TB, United Kingdom
Climatic Research Unit, and (3) Tyndall Centre for Climate Change
Research, both at School of Environmental Sciences, University of East
Anglia, Norwich NR4 7TJ, United Kingdom
International Water Management Institute, PO Box 2075, Colomb
[…]
