+++ data-raw/fetch_CRU_data.Rmd
@@ -3,7 +3,7 @@ title: "CRU CL2.0 Data"
 output: github_document
 ---
 
-```{r setup, include=FALSE}
+```{r knitr_setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
 ```
 
@@ -11,12 +11,18 @@ knitr::opts_chunk$set(echo = TRUE)
 
 CRU CL2.0 data are a gridded climatology of 1961-1990 monthly means released in 2002 and cover all land areas (excluding Antarctica) at 10 arc-second resolution. For more information see the description of the data provided by the University of East Anglia Climate Research Unit (CRU), http://www.cru.uea.ac.uk/cru/data/hrg/tmc/readme.txt.
 
-```{r get_data}
-install.packages("devtools")
-devtools::install_github("adamhsparks/getCRUCL2.0")
+# Download, extract and merge CRU data with provided GSOD climate data
 
+## Setup the R session
+
+```{r session_setup}
 library(getCRUCL2.0)
 
+# load existing GSOD_clim data from package
+GSOD_clim <- GSODR::GSOD_clim
+```
+## Get CRU CL2.0 data
+```{r get_data}
 CRU_stack <- create_CRU_stack(pre = TRUE,
                               pre_cv = TRUE,
                               rd0 = TRUE,
@@ -27,11 +33,261 @@ CRU_stack <- create_CRU_stack(pre = TRUE,
                               tmx = TRUE,
                               sunp = TRUE,
                               frs = TRUE,
-                              wnd = TRUE,
-                              elv = TRUE)
+                              wnd = TRUE)
 ```
 
 ## Extract data for station locations
 ```{r extract_data}
-GSOD_CRU <- raster::extract(CRU_stack, stations)
+stations <- readr::read_csv(
+  "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
+  col_types = "ccccccddddd",
+  col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL",
+                "LAT", "LON", "ELEV_M", "BEGIN", "END"), skip = 1)
+
+stations[stations == -999.9] <- NA
+stations[stations == -999] <- NA
+stations <- stations[!is.na(stations$LAT) & !is.na(stations$LON), ]
+stations <- stations[stations$LAT != 0 & stations$LON != 0, ]
+stations <- stations[stations$LAT > -90 & stations$LAT < 90, ]
+stations <- stations[stations$LON > -180 & stations$LON < 180, ]
+stations <- stations[!is.na(stations$STN_NAME), ]
+stations$STNID <- as.character(paste(stations$USAF, stations$WBAN, sep = "-"))
+
+stations <- as.data.frame(stations)
+sp::coordinates(stations) <- ~ LON + LAT
+crs <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+sp::proj4string(stations) <- sp::CRS(crs)
+
+# create a vector of names for the raster layers in new stack
+CRU_stack_names <- c(
+  paste0(
+    "CRU_CL2_0_",
+    names(CRU_stack[1]),
+    "_",
+    c(
+      "Jan",
+      "Feb",
+      "Mar",
+      "Apr",
+      "May",
+      "Jun",
+      "Jul",
+      "Aug",
+      "Sep",
+      "Oct",
+      "Nov",
+      "Dec"
+    )
+  ),
+  paste0(
+    "CRU_CL2_0_",
+    names(CRU_stack[2]),
+    "_",
+    c(
+      "Jan",
+      "Feb",
+      "Mar",
+      "Apr",
+      "May",
[…]
