+++ data-raw/fetch_isd-history.md
@@ -1,7 +1,7 @@
 Fetch, clean and correct altitude in GSOD isd\_history.csv Data
 ================
 Adam H. Sparks
-2016-11-25
+2017-01-05
 
 Introduction
 ============
@@ -179,7 +179,7 @@ summary(SRTM_GSOD_elevation)
 ```
 
     ##      USAF               WBAN             STN_NAME        
-    ##  Length:28322       Length:28322       Length:28322      
+    ##  Length:28326       Length:28326       Length:28326      
     ##  Class :character   Class :character   Class :character  
     ##  Mode  :character   Mode  :character   Mode  :character  
     ##                                                          
@@ -187,23 +187,23 @@ summary(SRTM_GSOD_elevation)
     ##                                                          
     ##                                                          
     ##      CTRY              STATE               CALL                LAT        
-    ##  Length:28322       Length:28322       Length:28322       Min.   :-89.00  
+    ##  Length:28326       Length:28326       Length:28326       Min.   :-89.00  
     ##  Class :character   Class :character   Class :character   1st Qu.: 22.47  
-    ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.26  
-    ##                                                           Mean   : 31.14  
+    ##  Mode  :character   Mode  :character   Mode  :character   Median : 39.25  
+    ##                                                           Mean   : 31.13  
     ##                                                           3rd Qu.: 49.88  
     ##                                                           Max.   : 89.37  
     ##                                                                           
     ##       LON               ELEV_M           BEGIN               END          
     ##  Min.   :-179.983   Min.   :-350.0   Min.   :19010101   Min.   :19051231  
-    ##  1st Qu.: -83.269   1st Qu.:  23.0   1st Qu.:19570701   1st Qu.:20020415  
-    ##  Median :   6.700   Median : 140.0   Median :19760302   Median :20160113  
-    ##  Mean   :  -3.444   Mean   : 360.7   Mean   :19782304   Mean   :20043017  
-    ##  3rd Qu.:  61.899   3rd Qu.: 435.0   3rd Qu.:20020328   3rd Qu.:20161122  
-    ##  Max.   : 179.750   Max.   :5304.0   Max.   :20161113   Max.   :20161124  
+    ##  1st Qu.: -83.269   1st Qu.:  23.0   1st Qu.:19570701   1st Qu.:20020417  
+    ##  Median :   6.706   Median : 140.0   Median :19760302   Median :20160127  
+    ##  Mean   :  -3.439   Mean   : 360.8   Mean   :19782357   Mean   :20043158  
+    ##  3rd Qu.:  61.899   3rd Qu.: 435.0   3rd Qu.:20020384   3rd Qu.:20161230  
+    ##  Max.   : 179.750   Max.   :5304.0   Max.   :20161229   Max.   :20161231  
     ##                     NA's   :218                                           
     ##     STNID           ELEV_M_SRTM_90m 
-    ##  Length:28322       Min.   :-361.0  
+    ##  Length:28326       Min.   :-361.0  
     ##  Class :character   1st Qu.:  25.0  
     ##  Mode  :character   Median : 156.0  
     ##                     Mean   : 380.1  
@@ -243,11 +243,11 @@ SRTM_GSOD_elevation[, c(1:11) := NULL]
     ##     4: 010015-99999              NA
     ##     5: 010016-99999              NA
[…]
