+++ data-raw/fetch_CRU_data.md
@@ -1,0 +1,330 @@
+CRU CL2.0 Data
+================
+
+CRU CL2.0 Data
+--------------
+
+CRU CL2.0 data are a gridded climatology of 1961-1990 monthly means released in 2002 and cover all land areas (excluding Antarctica) at 10 arc-second resolution. For more information see the description of the data provided by the University of East Anglia Climate Research Unit (CRU), <http://www.cru.uea.ac.uk/cru/data/hrg/tmc/readme.txt>.
+
+Download, extract and merge CRU data with provided GSOD climate data
+====================================================================
+
+Setup the R session
+-------------------
+
+``` r
+library(getCRUCL2.0)
+
+# load existing GSOD_clim data from package
+GSOD_clim <- GSODR::GSOD_clim
+```
+
+Get CRU CL2.0 data
+------------------
+
+``` r
+CRU_stack <- create_CRU_stack(pre = TRUE,
+                              pre_cv = TRUE,
+                              rd0 = TRUE,
+                              tmp = TRUE,
+                              dtr = TRUE,
+                              reh = TRUE,
+                              tmn = TRUE,
+                              tmx = TRUE,
+                              sunp = TRUE,
+                              frs = TRUE,
+                              wnd = TRUE)
+```
+
+    ##  
+    ## Downloading requested data files.
+    ## 
+
+    ## 
+      |                                                                       
+      |                                                                 |   0%
+      |                                                                       
+      |========                                                         |  12%
+      |                                                                       
+      |================                                                 |  25%
+      |                                                                       
+      |========================                                         |  38%
+      |                                                                       
+      |================================                                 |  50%
+      |                                                                       
+      |=========================================                        |  62%
+      |                                                                       
+      |=================================================                |  75%
+      |                                                                       
+      |=========================================================        |  88%
+      |                                                                       
+      |=================================================================| 100%
+
+Extract data for station locations
+----------------------------------
+
+``` r
+stations <- readr::read_csv(
+  "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
+  col_types = "ccccccddddd",
+  col_names = c("USAF", "WBAN", "STN_NAME", "CTRY", "STATE", "CALL",
+                "LAT", "LON", "ELEV_M", "BEGIN", "END"), skip = 1)
+
+stations[stations == -999.9] <- NA
+stations[stations == -999] <- NA
+stations <- stations[!is.na(stations$LAT) & !is.na(stations$LON), ]
[…]

[…]
