### 2017-01-05 #79 @sckott
@jeffreyhanson @deniederhut if you get a chance soon have a look at submitter replies above and let us know if there's anything else you'd like changed/improved
