### 2017-01-05 e16804b
Update GSOD_clim.rda and documentation
GSOD_clim.rda filesize is small enough for CRAN again

Update fetch_CRU_data.Rmd/md to fix field names of CRU data in GSOD_clim

Update data documentation file
