+++ R/data.R
@@ -75,27 +75,27 @@
 #' \item{STNID}{Unique station ID}
 #' \item{LON}{Longitude in WGS84 system}
 #' \item{LAT}{Latitude in WGS84 system}
-#' \item{CHELSA_bio1_1979-2013_V1_1}{Annual mean temperature [degrees Celsius]}
-#' \item{CHELSA_bio2_1979-2013_V1_1}{Mean diurnal range [degrees Celsius]}
+#' \item{CHELSA_bio1_1979-2013_V1_1}{Annual mean temperature [degree C]}
+#' \item{CHELSA_bio2_1979-2013_V1_1}{Mean diurnal range [degree C]}
 #' \item{CHELSA_bio3_1979-2013_V1_1}{Isothermality}
 #' \item{CHELSA_bio4_1979-2013_V1_1}{Temperature seasonality}
-#' \item{CHELSA_bio5_1979-2013_V1_1}{Maximum temperature of warmest month
-#' [degrees Celsius]}
-#' \item{CHELSA_bio6_1979-2013_V1_1}{Minimum temperature of coldest month
-#' [degrees Celsius]}
-#' \item{CHELSA_bio7_1979-2013_V1_1}{Temperature annual range [degrees Celsius]}
+#' \item{CHELSA_bio5_1979-2013_V1_1}{Maximum Temperature of warmest month
+#' [degree C]}
+#' \item{CHELSA_bio6_1979-2013_V1_1}{Minimum Temperature of coldest month
+#' [degree C]}
+#' \item{CHELSA_bio7_1979-2013_V1_1}{Temperature Annual Range [degree C]}
 #' \item{CHELSA_bio8_1979-2013_V1_1}{Mean Temperature of wettest quarter
-#' [degrees Celsius]}
+#' [degree C]}
 #' \item{CHELSA_bio9_1979-2013_V1_1}{Mean Temperature of driest quarter
-#' [degrees Celsius]}
+#' [degree C]}
 #' \item{CHELSA_bio10_1979-2013_V1_1}{Mean Temperature of warmest quarter
-#' [degrees Celsius]}
+#' [degree C]}
 #' \item{CHELSA_bio11_1979-2013_V1_1}{Mean Temperature of coldest quarter
-#' [degrees Celsius]}
+#' [degree C]}
 #' \item{CHELSA_bio12_1979-2013_V1_1}{Annual precipitation amount [mm]}
 #' \item{CHELSA_bio13_1979-2013_V1_1}{Precipitation of wettest month [mm]}
 #' \item{CHELSA_bio14_1979-2013_V1_1}{Precipitation of driest month [mm]}
-#' \item{CHELSA_bio15_1979-2013_V1_1}{Precipitation seasonality}
+#' \item{CHELSA_bio15_1979-2013_V1_1}{Precipitation Seasonality}
 #' \item{CHELSA_bio16_1979-2013_V1_1}{Precipitation of wettest quarter [mm]}
 #' \item{CHELSA_bio17_1979-2013_V1_1}{Precipitation of driest quarter [mm]}
 #' \item{CHELSA_bio18_1979-2013_V1_1}{Precipitation of warmest quarter [mm]}
@@ -126,6 +126,33 @@
 #' \item{CHELSA_temp_11_1979-2013}{Mean November temperature}
 #' \item{CHELSA_temp_12_1979-2013}{Mean December temperature}
 #' \item{CHELSA_temp_1979-2013_land}{Mean annual temperature}
+#' \item{MODCF_meanannual}{Mean annual cloud fraction}
+#' \item{MODCF_monthlymean_01}{Mean January cloud fraction}
+#' \item{MODCF_monthlymean_02}{Mean February fraction}
[…]
