### 2017-01-06 d0d40b4
Remove climate data
Climate data formatted for use with GSODR are now in their own package,
GSODR.data, available from GitHub:
https://github.com/adamhsparks/GSODR.data

Update description file

Remove extra data files and associated documentation
