+++ R/GSODR.R
@@ -39,9 +39,6 @@
 #' @seealso \code{\link{get_station_list}} Fetch latest list of stations and
 #' corresponding metadata from the NCDC FTP server
 #'
-#' @seealso \code{\link{GSOD_clim}} GSDOR provides additional climatological
-#' data sets to enhance the use of this weather data
-#'
 #' @seealso \code{\link{SRTM_GSOD_elevation}} GSDOR provides additional
 #' elevation data derived from SRTM90m data
 #'
