### 2017-01-06 06564b6
Fix "suggests" in DESCRIPTION
