### 2017-01-06 #79 @sckott
okay, you do plan to put it on CRAN though since I imagine this will go to CRAN.
