### 2017-01-06 #79 @sckott
Ah, if not required for GSODR, then all good,
