### 2017-01-07 #79 @deniederhut
Okay, vignettes look good! I think I would still like to see some explicit correctness check in the unit tests somewhere, perhaps with a cached result, but I am recommending this package for approval. A few bugs I ran into while testing:

1. Call to `get_GSOD(years = 2010, station = "955510-99999")` raises namespace error -- `Error: 'safely' is not an exported object from 'namespace:purrr'` ... probably in line 490 or 507
2. Not sure if this is GSODR or my machine, but tab completing the function names sometimes throws this:

```
Error in fetch(key) :
  lazy-load database '/Library/Frameworks/R.framework/Versions/3.2/Resources/library/GSODR/help/GSODR.rdb' is corrupt
 ```

3. Not super important, but `max_missing` seems to accept negative values and `NA`
