### 2017-01-07 #79 @adamhsparks
Thanks, I have fixed #3 easily. The other two I'm not so sure of, I've never run into either one with all my testing.

  1. I don't know, it's definitely in purrr 0.2.2. I've never had that issue.
  2. I ran to Google and did a quick search and found another RStudio user reporting this issue with all packages and someone on SO who had an issue with Java, but I can't find any fixes.
