### 2017-01-08 #79 @sckott
its possible you have @deniederhut `purrr` `v0.1`, since `safely` is not in that version https://github.com/hadley/purrr/blob/v0.1.0/NAMESPACE  - @adamhsparks plz add a minimum version of `purrr` to your DESCRIPTION file
