### 2017-01-08 #79 @sckott
> lazy-load database ... is corrupt

this is solved by restarting R usually
