### 2017-01-08 #79 @adamhsparks
@sckott, I've added the minimum version to DESCRIPTION. Thanks for the tip.
