### 2017-01-09 d6a3367
Update spatial and climate data vignette with more information on CHELSA
