### 2017-01-09 244c6c9
Update README.Rmd/md and Vignette
Clarify use of CHELSA data and correct typos
