+++ README.Rmd
@@ -242,6 +242,113 @@ system("zip -m Temperatures_PHL_2010-2010.kmz Temperatures_PHL_2010-2010.kml")
 Compare the GSOD weather data from the Philippines with climatic data
 provided by the GSODR package in the `GSOD_clim` data set.
 
+## Example - Download and plot data for a single country
+Download data for Philippines for year 2010 and generate a spatial, year summary
+file, PHL-2010.gpkg, in the user's home directory.
+
+### Install GSODR.data package
+
+This package is only available from GitHub; due to its large size (5.5Mb
+installed) it is not allowed on CRAN. It provides optional data for use with the
+GSODR package. See https://github.com/adamhsparks/GSODR.data for more.
+
+```r
+install.packages("devtools")
+devtools::install_github("adamhsparks/GSOD.data")
+````
+
+### Working with climate data from GSODR.data
+
+Now that the extra data have been installed, take a look at the CHELSA data.
+CHELSA (Climatologies at high resolution for the earth’s land surface areas) are
+climate data at (30 arc sec) for the earth land surface areas. 
+
+**Description of CHELSA data from CHELSA website**
+> CHELSA is a high resolution (30 arc sec) climate data set for the earth land surface areas currently under development in coorporation with the Department of Geography of the University of Hamburg (Prof. Dr. Jürgen Böhner, Dr. Olaf Conrad, Tobias Kawohl), the Swiss Federal Institute for Forest, Snow and Landscape Research WSL (Prof. Dr. Niklaus Zimmermann), the University of Zurich (Dr. Dirk N. Karger, Dr. Michael Kessler), and the University of Göttingen (Prof. Dr. Holger Kreft).
+It includes monthly mean temperature and precipitation patterns for the time period 1979-2013.
+CHELSA is based on a quasi-mechanistical statistical downscaling of the ERA interim global circulation model (http://www.ecmwf.int/en/research/climate-reanalysis/era-interim) with a GPCC bias correction (https://www.dwd.de/EN/ourservices/gpcc/gpcc.html) and is freely available  in the download section.
+
+See http://chelsa-climate.org for more information on these data.
+
+```
+library(GSODR.data)
+head(CHELSA)
+```
+
+```
+##          STNID    LON    LAT CHELSA_bio10_1979-2013_V1_1 CHELSA_bio11_1979-2013_V1_1 CHELSA_bio1_1979-2013_V1_1
+## 1 008268-99999 65.567 32.950                        30.0                         5.8                       18.3
+## 2 010014-99999  5.341 59.792                        14.3                         1.0                        7.1
+## 3 010015-99999  5.867 61.383                        12.5                        -3.2                        4.2
+## 4 010882-99999 11.342 62.578                        12.6                        -7.0                        2.3
+## 5 010883-99999  6.117 61.833                        15.2                        -0.2                        7.0
+## 6 010884-99999  9.567 59.185                        16.3                        -2.2                        6.6
+##   CHELSA_bio12_1979-2013_V1_1 CHELSA_bio13_1979-2013_V1_1 CHELSA_bio14_1979-2013_V1_1 CHELSA_bio15_1979-2013_V1_1
+## 1                       214.0                        54.6                         0.1                       104.9
+## 2                      1889.3                       223.7                        87.7                        30.3
+## 3                      2209.1                       254.7                       108.9                        30.1
+## 4                       563.0                        73.0                        26.7                        30.7
+## 5                      1710.3                       200.6                        78.5                        32.1
+## 6                       987.0                       120.9                        52.3                        24.4
[…]
