### 2017-01-09 c84bc5f
GSODR.data was renamed to GSODRdata
Fix references to this package
