+++ vignettes/Working_with_spatial_and_climate_data.Rmd
@@ -20,7 +20,7 @@ Following is an example of how you might download and save GSOD annual data for
 a given country, Philippines in this example, and convert it into a KML file for
 viewing in GoogleEarth. The second portion uses the same GeoPackage file to
 import the data back into R and combine the GSOD data with CHELSA data from the
-[GSODR.data package](https://github.com/adamhsparks/GSODR.data) available from
+[GSODRdata package](https://github.com/adamhsparks/GSODRdata) available from
 GitHub and plot the station temperatures for daily GSOD, average monthly GSOD
 and CHELSA temperatures (1979-2013). 
 
@@ -28,18 +28,18 @@ and CHELSA temperatures (1979-2013).
 Download data for Philippines for year 2010 and generate a spatial, year summary
 file, PHL-2010.gpkg, in the user's home directory.
 
-### Install GSODR.data package
+### Install GSODRdata package
 
 This package is only available from GitHub; due to its large size (5.5Mb
 installed) it is not allowed on CRAN. It provides optional data for use with the
-GSODR package.  See https://github.com/adamhsparks/GSODR.data for more.
+GSODR package.  See https://github.com/adamhsparks/GSODRdata for more.
 
 ```r
 install.packages("devtools")
 devtools::install_github("adamhsparks/GSOD.data")
 ````
 
-### Working with climate data from GSODR.data
+### Working with climate data from GSODRdata
 
 Now that the extra data have been installed, take a look at the CHELSA data.
 CHELSA (Climatologies at high resolution for the earth’s land surface areas) are
@@ -54,7 +54,7 @@ See http://chelsa-climate.org for more information on these data.
 
 ```r
 library(GSODR)
-library(GSODR.data)
+library(GSODRdata)
 head(CHELSA)
 ```
 
