+++ README.Rmd
@@ -246,18 +246,18 @@ provided by the GSODR package in the `GSOD_clim` data set.
 Download data for Philippines for year 2010 and generate a spatial, year summary
 file, PHL-2010.gpkg, in the user's home directory.
 
-### Install GSODR.data package
+### Install GSODRdata package
 
 This package is only available from GitHub; due to its large size (5.5Mb
 installed) it is not allowed on CRAN. It provides optional data for use with the
-GSODR package. See https://github.com/adamhsparks/GSODR.data for more.
+GSODR package. See https://github.com/adamhsparks/GSODRdata for more.
 
 ```r
 install.packages("devtools")
 devtools::install_github("adamhsparks/GSOD.data")
 ````
 
-### Working with climate data from GSODR.data
+### Working with climate data from GSODRdata
 
 Now that the extra data have been installed, take a look at the CHELSA data.
 CHELSA (Climatologies at high resolution for the earth’s land surface areas) are
@@ -271,7 +271,7 @@ CHELSA is based on a quasi-mechanistical statistical downscaling of the ERA inte
 See http://chelsa-climate.org for more information on these data.
 
 ```
-library(GSODR.data)
+library(GSODRdata)
 head(CHELSA)
 ```
 
