### 2017-01-09 4376e4b
Fix description of CHELSA data
Was not displayed as quote in rendered Markdown file
