### 2017-01-09 cf9e2fd
Update Working_with_spatial_and_climate_data.Rmd
Fix inappropriately closed code chunk

Fix quote for CHELSA description
