+++ codecov.yml
@@ -1,1 +1,72 @@
 comment: false
+codecov:
+  branch: master
+  bot: null
+
+coverage:
+  precision: 2
+  round: down
+  range: "70...100"
+
+  notify:
+    slack:
+      default:
+        url: null
+        threshold: null
+        branches: null
+        attachments: "sunburst, diff"
+
+    hipchat:
+      default:
+        url: null
+        notify: no
+        threshold: null
+        branches: null
+        card: yes
+        only_pulls: null
+        message: null
+
+    gitter:
+      default:
+        url: null
+        threshold: null
+        branches: null
+        message: null
+
+    webhook:
+      default:
+        url: null
+        threshold: null
+        branches: null
+        only_pulls: null
+
+    irc:
+      default:
+        server: null
+        channel: null
+        branches: null
+        threshold: null
+        message: null
+
+  status:
+    project:
+      default:
+        target: auto
+        threshold: null
+        branches: null
+
+    patch:
+      default:
+        target: auto
+        branches: null
+
+    changes:
+      default:
+        branches: null
+
+  ignore: null
+
+comment:
+  layout: "header, diff, changes, sunburst, uncovered, tree"
+  branches: null
+  behavior: default
