### 2017-01-09 d32c815
Minor update to README
