+++ .travis.yml
@@ -47,11 +47,5 @@ env:
     - RVERSION=release
     - RVERSION=devel
 
-r_github_packages:
-  - jimhester/covr
-
 after_success:
   - Rscript -e 'covr::codecov()'
-
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
