### 2017-01-10 d3fe678
Update tests
