+++ tests/testthat/test-get_GSOD.R
@@ -1,7 +1,6 @@
 context("get_GSOD")
 # Check that .validate_years handles invalid years -----------------------------
 test_that(".validate_years handles invalid years", {
-  skip_on_cran()
 
   expect_error(.validate_years(years = NULL),
                "\nYou must provide at least one year of data to download in a numeric\n         format.\n")
@@ -16,7 +15,7 @@ test_that(".validate_years handles invalid years", {
 
 # Check that .validate_years handles valid years -------------------------------
 test_that(".validate_years handles valid years", {
-  skip_on_cran()
+
   expect_error(.validate_years(years = 1929:2016), regexp = NA)
 
   expect_error(.validate_years(years = 2016), regexp = NA)
@@ -25,7 +24,6 @@ test_that(".validate_years handles valid years", {
 
 # Check that invalid stations are handled --------------------------------------
 test_that("invalid stations are handled", {
-  skip_on_cran()
   stations <- get_station_list()
   expect_error(.validate_station(years = 2015, station = "aaa-bbbbbb", stations),
                "\naaa-bbbbbb is not a valid station ID number, please check\n      your entry. Station IDs are provided as a part of the GSODR package in the\n      'stations' data\nin the STNID column.\n")
@@ -33,7 +31,6 @@ test_that("invalid stations are handled", {
 
 # Check that invalid dsn is handled --------------------------------------------
 test_that("invalid dsn is handled", {
-  skip_on_cran()
 
   expect_error(.validate_fileout(CSV = FALSE, dsn = "~/NULL", filename = NULL,
                                  GPKG = FALSE),
@@ -72,7 +69,7 @@ test_that("stations list and associated metatdata", {
 # Check missing days in non-leap years -----------------------------------------
 test_that("missing days check allows stations with permissible days missing,
           non-leap year", {
-            skip_on_cran()
+
             max_missing <- 5
             td <- tempdir()
             just_right_2015 <- data.frame(c(rep(12, 360)), c(rep("X", 360)))
@@ -101,7 +98,7 @@ test_that("missing days check allows stations with permissible days missing,
 # Check missing days in leap years ---------------------------------------------
 test_that("missing days check allows stations with permissible days missing,
           leap year", {
-            skip_on_cran()
+
             max_missing <- 5
             td <- tempdir()
             just_right_2016 <- data.frame(c(rep(12, 361)), c(rep("X", 361)))
@@ -129,14 +126,14 @@ test_that("missing days check allows stations with permissible days missing,
 
 # Check that max_missing only accepts positive values --------------------------
 test_that("The 'max_missing' parameter will not accept NA values", {
-  skip_on_cran()
+
   expect_error(get_GSOD(years = 2010, max_missing = NA),
 "\nThe 'max_missing' parameter must be a positive value larger than
            1\n")
 })
 
-test_that("The 'max_missing' parameter will not accept NA values", {
-  skip_on_cran()
+test_that("The 'max_missing' parameter will not accept values < 1", {
+
   expect_error(get_GSOD(years = 2010, max_missing = 0.1),
   "\nThe 'max_missing' parameter must be a positive value larger than
            1\n")
@@ -144,7 +141,7 @@ test_that("The 'max_missing' parameter will not accept NA values", {
 
 # Check validate country returns a two letter code -----------------------------
 test_that("Check validate country returns a two letter code", {
-  skip_on_cran()
+
   country <- "Philippines"
   Philippines <- .validate_country(country)
   expect_match(Philippines, "RP")
[…]
