### 2017-01-12 dc2cf81
Add documentation website via `pkgdown`
Also fix version number in DESCRIPTION file and update authors field in
DESCRIPTION
