### 2017-01-12 29c4028
Add docs to rbuildignore and update DESCRIPTION file to be in line with master
