### 2017-01-12 102df84
Cleanup NEWS.md
