### 2017-01-12 a646a29
Update NEWS.md
File formatting and spelling corrections
