### 2017-01-12 117b33b
More reformatting NEWS.md
