### 2017-01-13 f40fac8
Fix NEWS.md so that pkgdown generates news page
