### 2017-01-13 6d9df13
Make NEWS.md and news index a bit nicer
