### 2017-01-13 79a5c5a
Update vignettes and README.Rmd/md
Link to documentation for each function in package

Correct some instructions in vignettes related to CHELSA data
