+++ docs/articles/Specified_stations_for_a_range_of_years.md
@@ -1,7 +1,7 @@
 ---
 title: "GSODR use case: Specified years/stations vignette"
 author: "Tomislav Hengl and Adam H Sparks"
-date: "2017-01-10"
+date: "2017-01-13"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteEngine{knitr::knitr}
@@ -91,13 +91,13 @@ loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
 ```
 
 ## Using get_GSOD to Fetch the Requested Station Files
-Using the `get_GSOD` function may not work with this many station and year
+Using the `get_GSOD()` function may not work with this many station and year
 combinations. Often the FTP server becomes overwhelmed and stops responding
 to requests.
 
-This example shows how you could construct a query using the `get_GSOD`
+This example shows how you could construct a query using the `get_GSOD()`
 function. Be aware that it may result in incomplete data and error from the
-server. If it does this, see the following option for using the `format_GSOD`
+server. If it does this, see the following option for using the `reformat_GSOD()`
 function.
 
 
@@ -109,8 +109,8 @@ get_GSOD(station = eval(parse(text = loop_stations[, 12])), years = 1960:2015,
 
 ## Another Option
 `GSODR` provides a function for dealing with local files that have been
-transfered from the server already as well, `reformat_GSOD`. If the previous
-example with `get_GSOD` does not work, this is a good alternative that takes a
+transfered from the server already as well, `reformat_GSOD()`. If the previous
+example with `get_GSOD()` does not work, this is a good alternative that takes a
 bit more intervention but gives the same results.
 
 Using your FTP client, e.g. FileZilla, log into the NCDC FTP server,

[…]
