### 2017-01-13 4eb4966
Corrections to README.Rmd/md and vignettes
Corrections to documentation and some additional enhancements to
examples provided
