+++ README.md
@@ -235,7 +235,7 @@ system("zip -m Temperatures_PHL_2010-2010.kmz Temperatures_PHL_2010-2010.kml")
 ```
 
 Compare the GSOD weather data from the Philippines with climatic data
-provided by the GSODR package in the `GSOD_clim` data set.
+provided by the GSODR package in the `CHELSA` data set.
 
 Example - Download and plot data for a single country
 -----------------------------------------------------
@@ -251,16 +251,17 @@ with the GSODR package. See <https://github.com/adamhsparks/GSODRdata>
 for more.
 
 ``` r
-install.packages("devtools")
-devtools::install_github("adamhsparks/GSOD.data")
+# install.packages("devtools")
+devtools::install_github("adamhsparks/GSODdata")
 ```
 
-### Working with climate data from GSODRdata
+### Working with climate data from `GSODRdata`
 
 Now that the extra data have been installed, take a look at the CHELSA
-data. CHELSA (Climatologies at high resolution for the earth’s land
-surface areas) are climate data at (30 arc sec) for the earth land
-surface areas.
+data that are one of the data sets included in the `GSODRdata` package.
+CHELSA (Climatologies at high resolution for the earth’s land surface
+areas) are climate data at (30 arc sec) for the earth land surface
+areas.
 
 **Description of CHELSA data from CHELSA website**
 
@@ -282,79 +283,84 @@ surface areas.
 
 See <http://chelsa-climate.org> for more information on these data.
 
-    library(GSODRdata)
-    head(CHELSA)
+``` r
+library(GSODRdata)
 
-    ##          STNID    LON    LAT CHELSA_bio10_1979-2013_V1_1 CHELSA_bio11_1979-2013_V1_1 CHELSA_bio1_1979-2013_V1_1
-    ## 1 008268-99999 65.567 32.950                        30.0                         5.8                       18.3
-    ## 2 010014-99999  5.341 59.792                        14.3                         1.0                        7.1
-    ## 3 010015-99999  5.867 61.383                        12.5                        -3.2                        4.2
-    ## 4 010882-99999 11.342 62.578                        12.6                        -7.0                        2.3
-    ## 5 010883-99999  6.117 61.833                        15.2                        -0.2                        7.0
-    ## 6 010884-99999  9.567 59.185                        16.3                        -2.2                        6.6
-    ##   CHELSA_bio12_1979-2013_V1_1 CHELSA_bio13_1979-2013_V1_1 CHELSA_bio14_1979-2013_V1_1 CHELSA_bio15_1979-2013_V1_1
-    ## 1                       214.0                        54.6                         0.1                       104.9
-    ## 2                      1889.3                       223.7                        87.7                        30.3
-    ## 3                      2209.1                       254.7                       108.9                        30.1
-    ## 4                       563.0                        73.0                        26.7                        30.7
-    ## 5                      1710.3                       200.6                        78.5                        32.1
-    ## 6                       987.0                       120.9                        52.3                        24.4
-    ##   CHELSA_bio16_1979-2013_V1_1 CHELSA_bio17_1979-2013_V1_1 CHELSA_bio18_1979-2013_V1_1 CHELSA_bio19_1979-2013_V1_1
-    ## 1                       155.4                         0.6                         1.7                        96.9
-    ## 2                       650.7                       273.5                       433.3                       490.4
-    ## 3                       752.8                       332.6                       448.3                       621.1
-    ## 4                       216.2                        85.5                       198.7                       122.0
-    ## 5                       590.0                       244.7                       330.9                       494.4
-    ## 6                       338.0                       175.3                       242.3                       182.7
[…]
