+++ README.Rmd
@@ -235,7 +235,7 @@ system("zip -m Temperatures_PHL_2010-2010.kmz Temperatures_PHL_2010-2010.kml")
 ```
 
 Compare the GSOD weather data from the Philippines with climatic data
-provided by the GSODR package in the `GSOD_clim` data set.
+provided by the GSODR package in the `CHELSA` data set.
 
 ## Example - Download and plot data for a single country
 Download data for Philippines for year 2010 and generate a spatial, year summary
@@ -248,13 +248,14 @@ installed) it is not allowed on CRAN. It provides optional data for use with the
 GSODR package. See https://github.com/adamhsparks/GSODRdata for more.
 
 ```r
-install.packages("devtools")
-devtools::install_github("adamhsparks/GSOD.data")
+# install.packages("devtools")
+devtools::install_github("adamhsparks/GSODdata")
 ````
 
-### Working with climate data from GSODRdata
+### Working with climate data from `GSODRdata`
 
-Now that the extra data have been installed, take a look at the CHELSA data.
+Now that the extra data have been installed, take a look at the CHELSA data that
+are one of the data sets included in the `GSODRdata` package.
 CHELSA (Climatologies at high resolution for the earth’s land surface areas) are
 climate data at (30 arc sec) for the earth land surface areas. 
 
@@ -266,12 +267,13 @@ CHELSA is based on a quasi-mechanistical statistical downscaling of the ERA inte
 
 See http://chelsa-climate.org for more information on these data.
 
-```
+``` r
 library(GSODRdata)
+
 head(CHELSA)
 ```
 
-```
+``` r
 ##          STNID    LON    LAT CHELSA_bio10_1979-2013_V1_1 CHELSA_bio11_1979-2013_V1_1 CHELSA_bio1_1979-2013_V1_1
 ## 1 008268-99999 65.567 32.950                        30.0                         5.8                       18.3
 ## 2 010014-99999  5.341 59.792                        14.3                         1.0                        7.1
@@ -358,9 +360,8 @@ library(dplyr)
 library(ggplot2)
 library(reshape2)
 
-data(GSOD_clim)
 cnames <- paste0("CHELSA_temp_", 1:12, "_1979-2013")
-clim_temp <- GSOD_clim[GSOD_clim$STNID %in% pnts$STNID,
+clim_temp <- CHELSA[CHELSA$STNID %in% pnts$STNID,
                        paste(c("STNID", cnames))]
 clim_temp_df <- data.frame(STNID = rep(clim_temp$STNID, 12),
                            MONTHC = as.vector(sapply(1:12, rep,
@@ -405,16 +406,80 @@ decimal degrees.
 n <- nearest_stations(LAT = -27.5598, LON = 151.9507, distance = 50)
 
 n
+[1] "945510-99999" "945520-99999" "945620-99999" "949999-00170" "949999-00183" "955510-99999"
 
 toowoomba <- get_GSOD(years = 2015, station = n)
 
+#> This station, 945510-99999, only provides data for years 1956 to 2012.
+
+#> This station, 949999-00170, only provides data for years 1971 to 1984.
+
+#> This station, 949999-00183, only provides data for years 1983 to 1984.
+
+#> Checking requested station file for availability on server.
+
[…]
