### 2017-01-13 827a662
Many updates and corrections to documentation
