+++ vignettes/Working_with_spatial_and_climate_data.Rmd
@@ -11,39 +11,27 @@ vignette: >
 
 # Introduction
 
-The GSODR package provides the ability to interact with GSOD data using
-spatial methods. The `get_GSOD()` function allows for the data to be saved as
-a GeoPackage file which can be read by most GIS software packages or in R using
-R's GIS capabilities with contributed packages as well.
+The GSODR package provides the ability to interact with GSOD data using spatial methods. The `get_GSOD()` function allows for the data to be saved as a GeoPackage file which can be read by most GIS software packages or in R using R's GIS capabilities with contributed packages as well.
 
-Following is an example of how you might download and save GSOD annual data for
-a given country, Philippines in this example, and convert it into a KML file for
-viewing in GoogleEarth. The second portion uses the same GeoPackage file to
-import the data back into R and combine the GSOD data with CHELSA data from the
-[GSODRdata package](https://github.com/adamhsparks/GSODRdata) available from
-GitHub and plot the station temperatures for daily GSOD, average monthly GSOD
-and CHELSA temperatures (1979-2013). 
+Following is an example of how you might download and save GSOD annual data for a given country, Philippines in this example, and convert it into a KML file for viewing in GoogleEarth. The second portion uses the same GeoPackage file to
+import the data back into R and combine the GSOD data with CHELSA data from the [GSODRdata package](https://github.com/adamhsparks/GSODRdata) available from GitHub and plot the station temperatures for daily GSOD, average monthly GSOD and CHELSA temperatures (1979-2013). 
 
-## Example - Download and plot data for a single country
-Download data for Philippines for year 2010 and generate a spatial, year summary
-file, PHL-2010.gpkg, in the user's home directory.
+#### Example - Working with climate data from GSODRdata
 
-### Install GSODRdata package
+This example will demonstrate how to download data for Philippines for year 2010 and generate a spatial, year summary file, PHL-2010.gpkg, in the user's home directory and link it with climate data from the `GSODRdata` package.
 
-This package is only available from GitHub; due to its large size (5.5Mb
-installed) it is not allowed on CRAN. It provides optional data for use with the
-GSODR package.  See https://github.com/adamhsparks/GSODRdata for more.
+##### Install `GSODRdata` package
 
 ```r
-#install.packages("devtools")
+# install.packages("devtools")
 devtools::install_github("adamhsparks/GSODdata")
-```
+````
 
-### Working with climate data from GSODRdata
+##### Working with climate data from `GSODRdata`
 
-Now that the extra data have been installed, take a look at the CHELSA data.
-CHELSA (Climatologies at high resolution for the earth’s land surface areas) are
-climate data at (30 arc sec) for the earth land surface areas. 
+Now that the extra data have been installed, take a look at the CHELSA data that are one of the data sets included in the `GSODRdata` package.
+
+CHELSA (Climatologies at high resolution for the earth’s land surface areas) are climate data at (30 arc sec) for the earth land surface areas. 
 
 **Description of CHELSA data from CHELSA website**
 
[…]

[…]
