### 2017-01-13 2f3b810
More corrections and enhancements to documentation
