+++ vignettes/Working_with_spatial_and_climate_data.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Working with spatial and climate data from GSODR"
-author: "Adam H Sparks and Tomislav Hengl"
+author: "Tomislav Hengl and Adam H Sparks"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
