+++ vignettes/Specified_stations_for_a_range_of_years.md
@@ -1,6 +1,6 @@
 ---
 title: "GSODR use case: Specified years/stations vignette"
-author: "Tomislav Hengl and Adam H Sparks"
+author: "Adam H Sparks"
 date: "2017-01-13"
 output: rmarkdown::html_vignette
 vignette: >
@@ -11,30 +11,20 @@ vignette: >
 
 # Gather Data for IRRI Central Luzon Survey Loop
 
-The IRRI (International Rice Research Institute) survey loop in Central Luzon
-is a study that aims to monitor the changes in rice farming in the major rice
-producing area of the Philippines - the Central Luzon region, which is called as
-the "rice bowl of the Philippines". Data have been collected in this project
-since the 1960s. See, <http://ricestat.irri.org/fhsd/php/panel.php?page=1&sortBy=title&sortOrder=ascending#>
-for the panel data.
+The IRRI (International Rice Research Institute) survey loop in Central Luzon is a study that aims to monitor the changes in rice farming in the major rice producing area of the Philippines - the Central Luzon region, which is called as the "rice bowl of the Philippines". Data have been collected in this project since the 1960s. See, <http://ricestat.irri.org/fhsd/php/panel.php?page=1&sortBy=title&sortOrder=ascending#> for the panel data.
 
-This vignette details how to find and retrieve weather data for the area that
-this survey covers for the time period of 1960-2016. Methods that are detailed
-include: 
+This vignette details how to find and retrieve weather data for the area that this survey covers for the time period of 1960-2016. Methods that are detailed include: 
   * retrieving a spatial object of provincial level data;
   * subsetting this data for the provinces of interest;
   * merging the polygons into one object;
   * finding the centroid of this resulting polygon;
-  * using the centroid of the polygon to find stations within 100km of this
-  point;
-  * determining which stations provide data for the specified time-period, 1960-
-  2016; and 
+  * using the centroid of the polygon to find stations within 100km of this point;
+  * determining which stations provide data for the specified time-period, 1960-2016; and 
   * downloading the station files and creating a single CSV file of the data.
 
 ## Retrieve PHL Provincial Data and Select Loop Provinces
-First we retrieve data from GADM.org that will provide the provincial spatial
-data for the survey area. We will then use this to find the centroid, which will
-be used to find the nearest stations.
+
+First we retrieve data from GADM.org that will provide the provincial spatial data for the survey area. We will then use this to find the centroid, which will be used to find the nearest stations.
 
 ```r
 library(raster)
@@ -56,9 +46,8 @@ Central_Luzon <- RP[RP@data$NAME_1 == "Pampanga" |
 ```
 
 ## Dissolve Polygons and Find Centroid of Loop Survey Area
-Now that we have the provincial data imported, we will dissolve the polygons and
-find the centroid.
 
+Now that we have the provincial data imported, we will dissolve the polygons and find the centroid.
 
 ```r
 library(rgeos)
@@ -66,6 +55,7 @@ Central_Luzon <- gUnaryUnion(Central_Luzon)
 centroid <- gCentroid(Central_Luzon)
 ```
 
+Next, make a list of stations that are within this area.
 
 ```r
 library(GSODR)
@@ -91,14 +81,10 @@ loop_stations <- loop_stations[loop_stations$BEGIN <= 19591231 &
 ```
 
 ## Using get_GSOD to Fetch the Requested Station Files
-Using the `get_GSOD()` function may not work with this many station and year
-combinations. Often the FTP server becomes overwhelmed and stops responding
-to requests.
 
-This example shows how you could construct a query using the `get_GSOD()`
-function. Be aware that it may result in incomplete data and error from the
-server. If it does this, see the following option for using the `reformat_GSOD()`
-function.
[…]

[…]
