### 2017-01-13 #79 @adamhsparks
I'm following Hadley's [preference](http://r-pkgs.had.co.nz/description.html#author):

> You can also use separate Maintainer and Author fields. I prefer not to use these fields because Authors@R offers richer metadata.
