### 2017-01-13 e83386c
Update documentation
Replace “\*” in ROxygen with “*”
