+++ R/get_GSOD.R
@@ -122,7 +122,7 @@
 #' so this will sometimes not be the max for the calendar day. Missing =
 #' NA}
 #' \item{MAX_FLAG}{Blank indicates max temp was taken from the explicit max
-#' temp report and not from the 'hourly' data. An "\*" indicates max temp was
+#' temp report and not from the 'hourly' data. An "*" indicates max temp was
 #' derived from the hourly data (i.e., highest hourly or synoptic-reported
 #' temperature)}
 #' \item{MIN}{Minimum temperature reported during the day converted to
@@ -130,7 +130,7 @@
 #' so this will sometimes not be the max for the calendar day. Missing =
 #' NA}
 #' \item{MIN_FLAG}{Blank indicates max temp was taken from the explicit max
-#' temp report and not from the 'hourly' data. An "\*" indicates min temp was
+#' temp report and not from the 'hourly' data. An "*" indicates min temp was
 #' derived from the hourly data (i.e., highest hourly or synoptic-reported
 #' temperature)}
 #' \item{PRCP}{Total precipitation (rain and/or melted snow) reported during
