### 2017-01-16 5214d89
Update README.Rmd/md and index.html
Move final fields and units earlier in the file so it's more easily
found. Then demonstrate how to use it.
