+++ README.Rmd
@@ -69,6 +69,118 @@ If you find bugs, please file a [report as an issue](https://github.com/adamhspa
 devtools::install_github("adamhsparks/GSODR", build_vignettes = TRUE)
 ```
 
+------
+
+Final data format and contents
+------------------------------
+
+The `GSODR` package returns data with the following fields/units.
+
+-   **STNID** - Station number (WMO/DATSAV3 number) for the location;
+
+-   **WBAN** - number where applicable--this is the historical "Weather Bureau Air Force Navy" number - with WBAN being the acronym;
+
+-   **STN\_NAME** - Unique text identifier;
+
+-   **CTRY** - Country in which the station is located;
+
+-   **LAT** - Latitude. *Station dropped in cases where values are < -90 or > 90 degrees or Lat = 0 and Lon = 0*;
+
+-   **LON** - Longitude. *Station dropped in cases where values are < -180 or > 180 degrees or Lat = 0 and Lon = 0*;
+
+-   **ELEV\_M** - Elevation in metres;
+
+-   **ELEV\_M\_SRTM\_90m** - Elevation in metres corrected for possible errors, derived from the CGIAR-CSI SRTM 90m database (Jarvis et al. 2008);
+
+-   **YEARMODA** - Date in YYYY-mm-dd format;
+
+-   **YEAR** - The year (YYYY);
+
+-   **MONTH** - The month (mm);
+
+-   **DAY** - The day (dd);
+
+-   **YDAY** - Sequential day of year (not in original GSOD);
+
+-   **TEMP** - Mean daily temperature converted to degrees C to tenths. Missing = NA;
+
+-   **TEMP\_CNT** - Number of observations used in calculating mean daily temperature;
+
+-   **DEWP** - Mean daily dew point converted to degrees C to tenths. Missing = NA;
+
+-   **DEWP\_CNT** - Number of observations used in calculating mean daily dew point;
+
+-   **SLP** - Mean sea level pressure in millibars to tenths. Missing = NA;
+
+-   **SLP\_CNT** - Number of observations used in calculating mean sea level pressure;
+
+-   **STP** - Mean station pressure for the day in millibars to tenths. Missing = NA;
+
+-   **STP\_CNT** - Number of observations used in calculating mean station pressure;
+
+-   **VISIB** - Mean visibility for the day converted to kilometres to tenths Missing = NA;
+
+-   **VISIB\_CNT** - Number of observations used in calculating mean daily visibility;
+
+-   **WDSP** - Mean daily wind speed value converted to metres/second to tenths Missing = NA;
+
+-   **WDSP\_CNT** - Number of observations used in calculating mean daily wind speed;
+
+-   **MXSPD** - Maximum sustained wind speed reported for the day converted to metres/second to tenths. Missing = NA;
+
+-   **GUST** - Maximum wind gust reported for the day converted to metres/second to tenths. Missing = NA;
+
+-   **MAX** - Maximum temperature reported during the day converted to Celsius to tenths--time of max temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA;
+
+-   **MAX\_FLAG** - Blank indicates max temp was taken from the explicit max temp report and not from the 'hourly' data. An "\*" indicates max temp was derived from the hourly data (i.e., highest hourly or synoptic-reported temperature);
+
+-   **MIN** - Minimum temperature reported during the day converted to Celsius to tenths--time of min temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA;
+
+-   **MIN\_FLAG** - Blank indicates max temp was taken from the explicit min temp report and not from the 'hourly' data. An "\*" indicates min temp was derived from the hourly data (i.e., highest hourly or synoptic-reported temperature);
+
[…]
