### 2017-01-16 4f6c96c
Correct bad links; fix Authors@R field
