+++ README.Rmd
@@ -584,8 +584,7 @@ WMO Resolution 40. NOAA Policy
 > non-commercial international activities without restriction. The
 > non-U.S. data cannot be redistributed for commercial purposes.
 > Re-distribution of these data by others must provide this same
-> notification." [WMO Resolution 40. NOAA
-> Policy](http://www.wmo.int/pages/about/Resolution40.html)
+> notification." [WMO Resolution 40. NOAA Policy](https://public.wmo.int/en/our-mandate/what-we-do/data-exchange-and-technology-transfer)
 
 Code of Conduct
 ---------------
