+++ vignettes/Working_with_spatial_and_climate_data.Rmd
@@ -200,13 +200,7 @@ for more detail on the correction methods.
 
 *Users of these data should take into account the following (from the [NCDC website](http://www7.ncdc.noaa.gov/CDO/cdoselect.cmd?datasetabbv=GSOD&countryabbv=&georegionabbv=)):*
 
-> "The following data and products may have conditions placed on their 
-international commercial use. They can be used within the U.S. or for
-non-commercial international activities without restriction. The
-non-U.S. data cannot be redistributed for commercial purposes.
-Re-distribution of these data by others must provide this same
-notification." [WMO Resolution 40. NOAA
-Policy](http://www.wmo.int/pages/about/Resolution40.html)
+> "The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non-commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification." [WMO Resolution 40. NOAA Policy](https://public.wmo.int/en/our-mandate/what-we-do/data-exchange-and-technology-transfer)
 
 # References
 Stachelek, J. 2016. Using the Geopackage Format with R. 
