### 2017-01-16 9deab68
Change licence to MIT
Change for intraropensci consistency
