### 2017-01-16 d0955ad
Update static page docs
