### 2017-01-16 e45be2f
Update NEWS.md file
