### 2017-01-18 #79 @jeffreyhanson
@adamhsparks Ok, sounds good.
