### 2017-01-18 #79 @jeffreyhanson
@adamhsparks @sckott

**Overall, I am very impressed with this package. I think it will make a great addition to rOpenSci's
arsenal. I think it's almost perfect, I just found a few issues when running tests and checks.**

1. Some tests fail. Let me know if there's any other information I can provide to help with
   debugging. I've included details on the testing environment below.

```
Failed -------------------------------------------------------------------------

1. Failure: reformat_GSOD file_list parameter reformats data properly (@test-reformat_GSOD.R#12)
nrow(x) not equal to 1454.
1/1 mismatches
[1] 0 - 1454 == -1454

2. Failure: reformat_GSOD file_list parameter reformats data properly (@test-reformat_GSOD.R#13)
ncol(x) not equal to 48.
1/1 mismatches
[1] 0 - 48 == -48

3. Failure: reformat_GSOD dsn parameter reformats data properly (@test-reformat_GSOD.R#23)
nrow(x) not equal to 1454.
1/1 mismatches
[1] 0 - 1454 == -1454

4. Failure: reformat_GSOD dsn parameter reformats data properly (@test-reformat_GSOD.R#24)
ncol(x) not equal to 48.
1/1 mismatches
[1] 0 - 48 == -48
```

2. Some examples fail (run with `devtools::run_examples(run=FALSE)`).
  * `get_gsod` function.
    + Line 197: this needs to be commented out.
    + Lines 199-200: This example yields an error. This might have something
      to do with my installation of GDAL. Or the package might be a missing dependency in the
      DESCRIPTION?

```
get_GSOD(years = 2010, country = "Philippines", dsn = "~/",
            filename = "Philippines_GSOD", GPKG = TRUE, max_missing = 5)
trying URL 'ftp://ftp.ncdc.noaa.gov/pub/data/gsod/2010/gsod_2010.tar'
ftp data connection made, file length 98222080 bytes
==================================================
downloaded 93.7 MB

Checking stations against max_missing value
Starting data file processing
  |======================================================================| 100%

Writing GeoPackage File to Disk.

Error in rgdal::writeOGR(GSOD_XY, dsn = path.expand(outfile), layer = "GSOD",  :
  No such driver: GPKG
```

  * `get_station_list` function.
    + Line 11: `fetch_station_list()` should be `get_station_list()`

3. The documentation could be enhanced.
  * Several functions are missing an `@return` field in the documentation. In most cases they just
    need to describe the class of the returned object for completeness sake. For example,
    `get_station_list` just needs something like a `\code{\link[data.table]{data.table} object`.

### Test environment

```
R version 3.3.2 (2016-10-31)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 14.04.5 LTS

locale:
[1] C

attached base packages:

other attached packages:
[1] GSODR_1.0.0     testthat_1.0.2  devtools_1.12.0

 [1] Rcpp_0.12.8       magrittr_1.5      roxygen2_5.0.1    maps_3.1.1
 [5] lattice_0.20-34   R6_2.2.0          stringr_1.1.0     plyr_1.8.4
 [9] dplyr_0.5.0       fields_8.10       tools_3.3.2       rgdal_1.2-5
[13] grid_3.3.2        data.table_1.10.0 spam_1.4-0        R.oo_1.21.0
[17] DBI_0.5-1         withr_1.0.2       digest_0.6.10     assertthat_0.1
[21] tibble_1.2        crayon_1.3.2      readr_1.0.0       purrr_0.2.2
[25] bitops_1.0-6      R.utils_2.5.0     RCurl_1.95-4.8    curl_2.3
[29] evaluate_0.10     memoise_1.0.0     sp_1.2-4          stringi_1.1.2
[33] R.methodsS3_1.7.1
```

PS. Thank you for introducing me to the .gpkg format. It looks much easier to work with than ESRI Shapefiles for vector data.
