### 2017-01-18 63c643c
Comment out uncommented line in documentation
Line 197 of `get_GSOD()` was uncommented and should be commented
