### 2017-01-18 0471d9c
Specify rgdal version in DESCRIPTION
gpkg driver was introduced with 1.1-9, necessary for `GSODR` to work
properly
