### 2017-01-18 9937554
Correct line 11 of get_station_list.R
`fetch_station_list()` should be `get_station_list()` in example
