### 2017-01-18 2b15192
Update documentation to go with previous commit
