### 2017-01-18 7627c8a
Corrections in README
`get_station_list()` was incorrectly omitted from list of main
functions that `GSODR` provides.
