+++ README.md
@@ -27,15 +27,17 @@ to make it easy to find, transfer and format the data you need for use
 in analysis. The GSODR package provides four main functions for
 facilitating this:
 
--   `get_GSOD()` - the main function that will query and transfer files
-    from the FTP server, reformat them and return a data.frame in R or
-    save a file to disk  
--   `reformat_GSOD()` - the workhorse, this function takes individual
-    station files on the local disk and reformats them returning a
-    data.frame in R  
--   `nearest_stations()` - this function returns a dataframe containing
-    a list of stations and their metadata that fall within the given
-    radius of a point specified by the user
+-   `get_GSOD()` - queries and transfers files from the FTP server,
+    reformats them and returns a `data.frame()` object in R or saves a
+    file to disk with options for a GeoPackage or comma seperated values
+    (CSV) file,  
+-   `reformat_GSOD()` - the workhorse, takes individual station files on
+    the local disk and reformats them returning a data.frame in R,  
+-   `nearest_stations()` - returns a dataframe containing a list of
+    stations and their metadata that fall within the given radius of a
+    point specified by the user,  
+-   `get_station_list()` - downloads the latest station list from the
+    NCDC FTP server.
 
 When reformatting data either with `get_GSOD()` or `reformat_GSOD()`,
 all units are converted to International System of Units (SI), e.g.,
