+++ README.Rmd
@@ -21,9 +21,10 @@ Introduction to GSODR
 The GSOD or [Global Surface Summary of the Day (GSOD)](https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod) data provided by the US National Climatic Data Center (NCDC) are a valuable source of weather data with global coverage. However, the data files are cumbersome and difficult to work with. The GSODR package aims to make it easy to find, transfer and format the data you need for use in
 analysis. The GSODR package provides four main functions for facilitating this:
 
--   `get_GSOD()` - the main function that will query and transfer files from the FTP server, reformat them and return a data.frame in R or save a file to disk  
--   `reformat_GSOD()` - the workhorse, this function takes individual station files on the local disk and reformats them returning a data.frame in R  
--   `nearest_stations()` - this function returns a dataframe containing a list of stations and their metadata that fall within the given radius of a point specified by the user
+-   `get_GSOD()` - queries and transfers files from the FTP server, reformats them and returns a `data.frame()` object in R or saves a file to disk with options for a GeoPackage or comma seperated values (CSV) file,  
+-   `reformat_GSOD()` - the workhorse, takes individual station files on the local disk and reformats them returning a data.frame in R,  
+-   `nearest_stations()` - returns a dataframe containing a list of stations and their metadata that fall within the given radius of a point specified by the user,  
+-   `get_station_list()` - downloads the latest station list from the NCDC FTP server.
 
 When reformatting data either with `get_GSOD()` or `reformat_GSOD()`, all units are converted to International System of Units (SI), e.g., inches to millimetres and Fahrenheit to Celsius. File output can be saved as a Comma Separated Value (CSV) file or in a spatial GeoPackage (GPKG) file, implemented by most major GIS software, summarising each year by
 station, which also includes vapour pressure and relative humidity variables calculated from existing data in GSOD.
