### 2017-01-18 3608d45
Minor edits for clarity in README
