### 2017-01-18 b19699c
Improve documentation
Add `@return` field for the four major functions
