+++ man/get_GSOD.Rd
@@ -51,6 +51,10 @@ not specified by the user, defaults to the current working directory.}
 save it locally in a user specified location, if \code{dsn} is not specified
 by the user, defaults to the current working directory.}
 }
+\value{
+A \code{data.frame} object of weather data or a comma-separated value
+(CSV) or GeoPackage (GPKG) file saved to local disk.
+}
 \description{
 This function automates downloading, cleaning, reformatting of data from
 the Global Surface Summary of the Day (GSOD) data provided by the US National
