### 2017-01-18 #79 @adamhsparks
Hi @jeffreyhanson,
Thanks for this complete review. I've addressed most of your comments already. At least the easy ones related to documentation.

Right now I'm checking on your GPKG error.

I've set up a fresh R install on an ElementaryOS VM (based on Xenial) and installed GSODR on that. With a fresh installation, I'm able to create a GPKG file and I have the same version of rgdal that you do, 1.2-5, though I did add a minimum version to the DESCRIPTION, >= 1.1-9, so I don't think that's the issue and I don't think I'm missing dependencies since it works for me on a fresh install.

What does `rgdal::ogrDrivers` return? I'm wondering if your lib-gdal is outdated so it's an OS issue, not an R/rgdal issue.
