### 2017-01-18 #79 @jeffreyhanson
Here's the output. It looks like it's an issue with my GDAL installation, so feel free to disregard the issue.

```
rgdal::ogrDrivers()
             name write
1          ARCGEN FALSE
2          AVCBin FALSE
3          AVCE00 FALSE
4      AeronavFAA FALSE
5             BNA  TRUE
6             CSV  TRUE
7         CouchDB  TRUE
8             DGN  TRUE
9            DODS FALSE
10            DXF  TRUE
11         EDIGEO FALSE
12 ESRI Shapefile  TRUE
13  ElasticSearch  TRUE
14            GFT  TRUE
15            GML  TRUE
16            GMT  TRUE
17       GPSBabel  TRUE
18  GPSTrackMaker  TRUE
19            GPX  TRUE
20        GeoJSON  TRUE
21         GeoRSS  TRUE
22     Geoconcept  TRUE
23       Geomedia FALSE
24            HTF FALSE
25         Idrisi FALSE
26     Interlis 1  TRUE
27     Interlis 2  TRUE
28            KML  TRUE
29         LIBKML  TRUE
30   MSSQLSpatial  TRUE
31   MapInfo File  TRUE
32         Memory  TRUE
33          MySQL  TRUE
34            NAS FALSE
35           ODBC  TRUE
36            ODS  TRUE
37           OGDI FALSE
38            OSM FALSE
39        OpenAir FALSE
40         PCIDSK  TRUE
41            PDF  TRUE
42            PDS FALSE
43         PGDump  TRUE
44           PGeo FALSE
45     PostgreSQL  TRUE
46            REC FALSE
47            S57  TRUE
48           SDTS FALSE
49       SEGUKOOA FALSE
50           SEGY FALSE
51         SQLite  TRUE
52            SUA FALSE
53            SVG FALSE
54          TIGER  TRUE
55        UK .NTF FALSE
56            VFK FALSE
57            VRT FALSE
58            WFS FALSE
59            XLS FALSE
60           XLSX  TRUE
61         XPlane FALSE
 ```
