### 2017-01-18 #79 @adamhsparks
@jeffreyhanson,
RE 1. : I have files saved in the `data raw` folder to check the functionality of `reformat_GSOD()`, which works with local files and I used a hardcoded path. I need to fix that. Good catch!
