### 2017-01-18 5ed4bd4
Fix hard-coded file paths in tests
Testing the `reformat_GSOD.R` function requires local files. I've
removed the files from the data-raw directory and now download them to
the R `tempdir` directory for testing.

Four files are downloaded, the `GSODR` functions are checked and the
files are removed at the end of the test.
