### 2017-01-18 120328a
Fix tests for reformat_GSOD()
Cleanup and redownload the files to be sure for both tests.
