+++ tests/testthat/test-reformat_GSOD.R
@@ -19,10 +19,20 @@ test_that("reformat_GSOD file_list parameter reformats data properly", {
   expect_equal(nrow(x), 1454)
   expect_equal(ncol(x), 48)
   expect_type(x, "list")
+  unlink(destinations)
 })
 
 test_that("reformat_GSOD dsn parameter reformats data properly", {
   skip_on_cran()
+  
+  ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/1960/"
+  test_files <- c("066000-99999-1960.op.gz", "066200-99999-1960.op.gz",
+                  "066700-99999-1960.op.gz", "066800-99999-1960.op.gz")
+  dest <- tempdir()
+  destinations <- paste0(dest, "/", test_files)
+  
+  Map(function(u, d) download.file(u, d, mode = "wb"),
+      paste0(ftp_base, test_files), destinations)
 
   x <- reformat_GSOD(dsn = dest)
   expect_equal(nrow(x), 1454)
