### 2017-01-18 1f523ef
Remove any extra .op.gz files in `tempdir()` before testing `reformat_GSOD`
