### 2017-01-18 9fc62ff
Revert "Remove any extra .op.gz files in `tempdir()` before testing `reformat_GSOD`"
This reverts commit 1f523ef0c35e1ebeaa089262b5ee6fc5b66328b6.
