+++ tests/testthat/test-reformat_GSOD.R
@@ -1,7 +1,7 @@
 
 # Check that reformat_GSOD functions properly ----------------------------------
 context("reformat_GSOD")
-file.remove(file.path(tempdir(), list.files(tempdir(), pattern = ".op.gz$"))) 
+
 test_that("reformat_GSOD file_list parameter reformats data properly", {
   skip_on_cran()
 
@@ -24,7 +24,7 @@ test_that("reformat_GSOD file_list parameter reformats data properly", {
 
 test_that("reformat_GSOD dsn parameter reformats data properly", {
   skip_on_cran()
-  file.remove(file.path(tempdir(), list.files(tempdir(), pattern = ".op.gz$"))) 
+  
   ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/1960/"
   test_files <- c("066000-99999-1960.op.gz", "066200-99999-1960.op.gz",
                   "066700-99999-1960.op.gz", "066800-99999-1960.op.gz")
