### 2017-01-18 cd72e57
Fix tests for `reformat_GSOD`
Remove any .op.gz files in the tempdir before testing
