### 2017-01-18 6f49fc9
Fix merge conflict
