+++ tests/testthat/test-reformat_GSOD.R
@@ -24,11 +24,7 @@ test_that("reformat_GSOD file_list parameter reformats data properly", {
 
 test_that("reformat_GSOD dsn parameter reformats data properly", {
   skip_on_cran()
-<<<<<<< HEAD
-  
-=======
   file.remove(file.path(tempdir(), list.files(tempdir(), pattern = ".op.gz$"))) 
->>>>>>> devel
   ftp_base <- "ftp://ftp.ncdc.noaa.gov/pub/data/gsod/1960/"
   test_files <- c("066000-99999-1960.op.gz", "066200-99999-1960.op.gz",
                   "066700-99999-1960.op.gz", "066800-99999-1960.op.gz")
@@ -37,11 +33,7 @@ test_that("reformat_GSOD dsn parameter reformats data properly", {
   
   Map(function(u, d) download.file(u, d, mode = "wb"),
       paste0(ftp_base, test_files), destinations)
-<<<<<<< HEAD
 
-=======
-  
->>>>>>> devel
   x <- reformat_GSOD(dsn = dest)
   expect_equal(nrow(x), 1454)
   expect_equal(ncol(x), 48)
