### 2017-01-18 #79 @sckott
thanks @jeffreyhanson  for taking another look

@adamhsparks looks like we're good to go then since looks like you dealt with that `reformat_GSOD` thing
