### 2017-01-18 7e96544
Fix bug related to dsn and add test to check for this
