### 2017-01-18 #79 @adamhsparks
@sckott I think we're good. I was waiting to check the results of Appveyor and Travis this morning when I woke up and see if everything was good as far as the `reformat_GSOD` testing.

I added another test related to the file saving parameters this morning based on a bug I found last night.

Everything looks good on winbuilder now as well. I'm ready to go if you are.
