### 2017-01-18 7bd73bf
Update cran comments
