### 2017-01-19 #79 @sckott
Approved 👍  Thx for your submission.

- I've added you to a GSODR team for on the rOpenSci Github org. Accept and transfer the repo (under "Settings") to rOpenSci
- Be sure to update any links for CI, coverage, and documentation.
- Add ropensci footer image to bottom of readme `[![ropensci_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)`
- Let me know if you are interested in doing a blog post about your package on the rOpenSci blog, either long-form (https://ropensci.org/blog/) or short (https://ropensci.org/tech-notes/)

I'll get back to you soon on the JOSS stuff in a bit
