### 2017-01-19 6431998
Move repository to ropensci
Update links in badges and documentation

Fix one outdated link that pointed to old repository name for GSODRdata
on GitHub as well
