+++ README.Rmd
@@ -8,9 +8,9 @@ output:
 GSODR: Global Surface Summary Daily Weather Data in R
 =====================================================
 
-[![Travis-CI Build Status](https://travis-ci.org/adamhsparks/GSODR.svg?branch=master)](https://travis-ci.org/adamhsparks/GSODR)
-[![Build status](https://ci.appveyor.com/api/projects/status/s09kh2nj59o35ob1/branch/master?svg=true)](https://ci.appveyor.com/project/adamhsparks/gsodr/branch/master)
-[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/GSODR/master.svg)](https://codecov.io/github/adamhsparks/GSODR?branch=master)
+[![Travis-CI Build Status](https://travis-ci.org/ropensci/GSODR.svg?branch=master)](https://travis-ci.org/ropensci/GSODR)
+[![Build status](https://ci.appveyor.com/api/projects/status/s09kh2nj59o35ob1/branch/master?svg=true)](https://ci.appveyor.com/project/ropensci/gsodr/branch/master)
+[![Coverage Status](https://img.shields.io/codecov/c/github/ropensci/GSODR/master.svg)](https://codecov.io/github/ropensci/GSODR?branch=master)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/GSODR?color=blue)](https://github.com/metacran/cranlogs.app)
 [![cran version](http://www.r-pkg.org/badges/version/GSODR)](https://cran.r-project.org/package=GSODR)
 [![DOI](https://zenodo.org/badge/32764550.svg)](https://zenodo.org/badge/latestdoi/32764550)
@@ -63,11 +63,11 @@ install.packages("GSODR")
 
 A development version is available from from GitHub. If you wish to install the development version that may have new features (but also may not work properly), install the [devtools package](https://CRAN.R-project.org/package=devtools), available from CRAN. We strive to keep the master branch on GitHub functional and working properly, although this may not always happen.
 
-If you find bugs, please file a [report as an issue](https://github.com/adamhsparks/GSODR/issues).
+If you find bugs, please file a [report as an issue](https://github.com/ropensci/GSODR/issues).
 
 ``` r
 #install.packages("devtools")
-devtools::install_github("adamhsparks/GSODR", build_vignettes = TRUE)
+devtools::install_github("ropensci/GSODR", build_vignettes = TRUE)
 ```
 
 ------
@@ -573,7 +573,7 @@ CHELSA (climatic surfaces at 1 km resolution) is based on a quasi-mechanistical 
 #### Elevation Values
 
 90m hole-filled SRTM digital elevation (Jarvis *et al.* 2008) was used to identify and correct/remove elevation errors in data for station locations between -60˚ and 60˚ latitude. This applies to cases here where elevation was missing in the reported values as well. In case the station reported an elevation and the DEM does not, the station reported is taken. For stations beyond -60˚ and 60˚ latitude, the values are station reported values in every instance. See
-<https://github.com/adamhsparks/GSODR/blob/devel/data-raw/fetch_isd-history.md> for more detail on the correction methods.
+<https://github.com/ropensci/GSODR/blob/devel/data-raw/fetch_isd-history.md> for more detail on the correction methods.
 
 WMO Resolution 40. NOAA Policy
 ------------------------------
