### 2017-01-19 c402852
Add ropensci footer
