### 2017-01-19 883167c
Update remainder of links to previous repository
Finalise the change from adamhsparks/GSODR to ropensci/GSODR
