### 2017-01-19 637ea0e
Cleanup README
Minor edits removing redundant wording and moving a sentence around for
better clarity.
