### 2017-01-19 45efe25
Update .travis.yml
Update for move to ropensci repository
