### 2017-01-19 1f35309
Attempt to get `covr` to report tests skipped on CRAN
