+++ .travis.yml
@@ -44,6 +44,7 @@ notifications:
 env:
   matrix:
     - RVERSION=release
+    - NOT_CRAN=true
 
 after_success:
   - Rscript -e 'covr::codecov()'
