### 2017-01-19 60abe39
Fix .travis.yml matrix
