+++ .travis.yml
@@ -43,8 +43,8 @@ notifications:
 
 env:
   matrix:
-    - RVERSION=release
-    - NOT_CRAN=true
+    - r: release
+      not_cran: true
 
 after_success:
   - Rscript -e 'covr::codecov()'
