### 2017-01-19 74e108b
Update coverage badge
