### 2017-01-19 43bef44
Install packages from apt
Rather than building everything from source, install library
dependencies using apt and change language to R
