+++ .travis.yml
@@ -1,37 +1,14 @@
-language: c
+language: r
 sudo: required
+cache: packages
+
+env:
+  matrix:
+    - NOT_CRAN=true
 
 before_install:
-  - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
-  - chmod 755 pkg-build.sh
-  - ./pkg-build.sh bootstrap
-  - cd
-  - wget http://download.osgeo.org/geos/geos-3.5.0.tar.bz2
-  - tar xjf geos-3.5.0.tar.bz2
-  - cd geos-3.5.0
-  - ./configure --prefix=/usr
-  - make
-  - sudo make install
-  - wget http://download.osgeo.org/gdal/2.0.1/gdal-2.0.1.tar.gz
-  - tar zxf gdal-2.0.1.tar.gz
-  - cd gdal-2.0.1
-  - ./configure --prefix=/usr
-  - make
-  - sudo make install
-  - cd
-  - wget http://download.osgeo.org/proj/proj-4.9.2.tar.gz
-  - tar zxvf proj-4.9.2.tar.gz
-  - cd proj-4.9.2
-  - ./configure --prefix=/usr
-  - make
-  - sudo make install
-  - cd /home/travis/build/ropensci/GSODR
-
-install:
-  - ./pkg-build.sh install_deps
-
-script:
-  - ./pkg-build.sh run_tests
+- sudo apt-get -qq update
+  - sudo apt-get install -y libcurl4-openssl-dev, libgdal-dev, libproj-dev
 
 after_failure:
   - ./pkg-build.sh dump_logs
@@ -40,11 +17,6 @@ notifications:
   email:
     on_success: change
     on_failure: change
-
-env:
-  matrix:
-    - r: release
-      not_cran: true
 
 after_success:
   - Rscript -e 'covr::codecov()'
