### 2017-01-19 7e9b735
Modify apt-get update command
