+++ .travis.yml
@@ -7,8 +7,8 @@ env:
     - NOT_CRAN=true
 
 before_install:
-- sudo apt-get -qq update
-  - sudo apt-get install -y libcurl4-openssl-dev, libgdal-dev, libproj-dev
+- sudo apt-get update
+- sudo apt-get install -y libcurl4-openssl-dev, libgdal-dev, libproj-dev
 
 after_failure:
   - ./pkg-build.sh dump_logs
