### 2017-01-19 3342abb
fix apt-get install command
