### 2017-01-19 c60be2e
Switch to Trusty for geospatial packages
