### 2017-01-19 959c965
Update version number
Development version 1.0.1

Update documentation
