### 2017-01-20 b89ea7f
Move test for get_station_list to its own file
