### 2017-01-20 #79 @adamhsparks
@sckott now that the dust is settling and I've moved everything over, I'm happy to do a short-form blog post.
