### 2017-01-20 #79 @sckott
@adamhsparks great, on the blog post: We'll send an email to get that started

on JOSS:

I saw that you did a new release, what is the Zenodo DOI for that?

You can submit to https://github.com/openjournals/joss-reviews/issues now
