### 2017-01-20 #79 @sckott
@adamhsparks blarg, sorry, meant to give you https links for the footer image, plz change to `[![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)`
